### 2021-01-04 #2938 @whedon
[PRE REVIEW]: libCEED: Fast algebra for high-order element-based discretizations
**Submitting author:** @jedbrown (<a href="http://orcid.org/0000-0002-9945-0639">Jed Brown</a>)
**Repository:** <a href="https://github.com/CEED/libCEED" target ="_blank">https://github.com/CEED/libCEED</a>
**Version:** v0.8.0
