### 2021-01-04 d791217
JOSS paper: fix DOI in bib
