+++ doc/papers/joss/paper.bib
@@ -1,5 +1,4 @@
 @article{Brown:2010,
-  Adsurl = {https://doi.org/10.1007/s10915-010-9396-8},
   Author = {{Brown}, J.},
   Journal = {Journal of Scientific Computing},
   Volume = 45,
@@ -180,8 +179,7 @@
   year         = 2020,
   publisher    = {Zenodo},
   version      = {0.7},
-  doi          = {10.5281/zenodo.4302737},
-  url          = {https://doi.org/10.5281/zenodo.4302737}
+  doi          = {10.5281/zenodo.4302737}
 }
 
 @InProceedings{libceed-paper-proc-scipy-2020,
@@ -191,8 +189,7 @@
   pages     = {85 - 90},
   year      = {2020},
   editor    = {{M}eghann {A}garwal and {C}hris {C}alloway and {D}illon {N}iederhut and {D}avid {S}hupe},
-  doi       = {10.25080/Majora-342d178e-00c},
-  url       = {https://doi.org/10.25080/Majora-342d178e-00c}
+  doi       = {10.25080/Majora-342d178e-00c}
 }
 
 @misc{LIBXSMM,
@@ -237,7 +234,6 @@ year = {2015},
 isbn = {9781450340052},
 publisher = {Association for Computing Machinery},
 address = {New York, NY, USA},
-url = {https://doi.org/10.1145/2833157.2833162},
 doi = {10.1145/2833157.2833162},
 abstract = {Dynamic, interpreted languages, like Python, are attractive for domain-experts and scientists experimenting with new ideas. However, the performance of the interpreter is often a barrier when scaling to larger data sets. This paper presents a just-in-time compiler for Python that focuses in scientific and array-oriented computing. Starting with the simple syntax of Python, Numba compiles a subset of the language into efficient machine code that is comparable in performance to a traditional compiled language. In addition, we share our experience in building a JIT compiler using LLVM[1].},
 booktitle = {Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC},
@@ -267,8 +263,7 @@ series = {LLVM '15}
  number        = {7825},
  pages         = {357--362},
  doi           = {10.1038/s41586-020-2649-2},
- publisher     = {Springer Science and Business Media {LLC}},
- url           = {https://doi.org/10.1038/s41586-020-2649-2}
+ publisher     = {Springer Science and Business Media {LLC}}
 }
 
 @misc{OCCAwebsite,
@@ -283,7 +278,7 @@ series = {LLVM '15}
   volume = "37",
   pages = "70 - 92",
   year = "1980",
-  doi = "https://doi.org/10.1016/0021-9991(80)90005-4",
+  doi = "10.1016/0021-9991(80)90005-4",
   author = "Steven A Orszag"
 }
 
