### 2021-01-04 a920638
Operator - add operator/qfunction field compatibility checks
