### 2021-01-04 8109189
QFunction - do not allow adding fields to QFunction in use with an operator
