### 2021-01-04 8930f3d
QFunction - do not allow adding fields to QFunction in use with an operator
