+++ interface/ceed-qfunction.c
@@ -590,6 +590,11 @@ int CeedQFunctionCreateIdentity(Ceed ceed, CeedInt size, CeedEvalMode inmode,
 **/
 int CeedQFunctionAddInput(CeedQFunction qf, const char *fieldname, CeedInt size,
                           CeedEvalMode emode) {
+  if (qf->operatorsset)
+    // LCOV_EXCL_START
+    return CeedError(qf->ceed, 1,
+                     "QFunction cannot be changed when in use by an operator");
+  // LCOV_EXCL_STOP
   if ((emode == CEED_EVAL_WEIGHT) && (size != 1))
     // LCOV_EXCL_START
     return CeedError(qf->ceed, 1, "CEED_EVAL_WEIGHT should have size 1");
@@ -618,6 +623,11 @@ int CeedQFunctionAddInput(CeedQFunction qf, const char *fieldname, CeedInt size,
 **/
 int CeedQFunctionAddOutput(CeedQFunction qf, const char *fieldname,
                            CeedInt size, CeedEvalMode emode) {
+  if (qf->operatorsset)
+    // LCOV_EXCL_START
+    return CeedError(qf->ceed, 1,
+                     "QFunction cannot be changed when in use by an operator");
+  // LCOV_EXCL_STOP
   if (emode == CEED_EVAL_WEIGHT)
     // LCOV_EXCL_START
     return CeedError(qf->ceed, 1, "Cannot create QFunction output with "
