+++ .github/workflows/rust-test-with-style.yml
@@ -22,10 +22,9 @@ jobs:
         override: true
     - name: Rust test with coverage
       uses: actions-rs/tarpaulin@v0.1
-      with:
-        env:
-          CC: ${{ matrix.compiler }}
-          FC: gfortran-9
+      env:
+        CC: ${{ matrix.compiler }}
+        FC: gfortran-9
     - name: Codecov upload
       uses: codecov/codecov-action@v1.0.2
       #with:
