### 2021-01-04 f1fc39c
Actions - add Rust code coverage
