### 2021-01-04 #2938 @jedbrown
:wave: Happy new year! I've fixed the DOI issue. For reviewers, I would consider

* Direct domain expertise
  * https://github.com/pbauman - libMesh and multiphysics solvers, now at AMD
  * https://github.com/FreddieWitherden - PyFR (similar to high-order DG), Python-based optimization and GPU offloading
  * https://github.com/lukeolson - algebraic multigrid, finite elements, and GPUs, Python
* Volunteered to review and have relevant expertise, but have not yet reviewed for JOSS (or not recently)
  * https://github.com/munkm - meshing, visualization, nuclear engineering apps, Python and Rust
  * https://github.com/fhorrobin - astrophysics, CFD, relevant languages
  * https://github.com/thelfer - solid mechanics
