### 2021-01-07 f204958
[Fix] Removes shadowing of variable
