### 2021-01-07 #2945 @jedbrown
FYI, we intend to release v0.8.0 no later than March. It's feature-complete from our perspective, but I'd love to include reviewer feedback (which could involve API changes). In any case, we'll tag v0.8.0 by the time this review completes.
