+++ rust/src/vector.rs
@@ -116,7 +116,7 @@ impl fmt::Display for Vector {
 impl Vector {
     // Constructors
     pub fn create(ceed: &crate::Ceed, n: usize) -> Self {
-        let n = n as i32;
+        let n = i32::try_from(n).unwrap();
         let mut ptr = std::ptr::null_mut();
         unsafe { bind_ceed::CeedVectorCreate(ceed.ptr, n, &mut ptr) };
         Self { ptr: ptr }
@@ -180,7 +180,7 @@ impl Vector {
     /// assert_eq!(n, 10, "Incorrect length");
     /// ```
     pub fn length(&self) -> usize {
-        let mut n = 0;
+        let mut n: i32 = 0;
         unsafe { bind_ceed::CeedVectorGetLength(self.ptr, &mut n) };
         usize::try_from(n).unwrap()
     }
