### 2021-01-07 5aa72d1
Actions - include Rust doctests
