### 2021-01-07 c38e850
Actions - Rust doctest coverange only avaliable in nightly
