### 2021-01-07 9e4c553
Rust - drop explict typing of values passed to C get* functions
