+++ rust/src/elem_restriction.rs
@@ -306,7 +306,7 @@ impl ElemRestriction {
     /// assert_eq!(c, compstride, "Incorrect component stride");
     /// ```
     pub fn comp_stride(&self) -> usize {
-        let mut compstride: i32 = 0;
+        let mut compstride = 0;
         unsafe { bind_ceed::CeedElemRestrictionGetCompStride(self.ptr, &mut compstride) };
         usize::try_from(compstride).unwrap()
     }
@@ -328,7 +328,7 @@ impl ElemRestriction {
     /// assert_eq!(n, nelem, "Incorrect number of elements");
     /// ```
     pub fn num_elements(&self) -> usize {
-        let mut numelem: i32 = 0;
+        let mut numelem = 0;
         unsafe { bind_ceed::CeedElemRestrictionGetNumElements(self.ptr, &mut numelem) };
         usize::try_from(numelem).unwrap()
     }
@@ -351,7 +351,7 @@ impl ElemRestriction {
     /// assert_eq!(e, elem_size, "Incorrect element size");
     /// ```
     pub fn elem_size(&self) -> usize {
-        let mut elemsize: i32 = 0;
+        let mut elemsize = 0;
         unsafe { bind_ceed::CeedElemRestrictionGetElementSize(self.ptr, &mut elemsize) };
         usize::try_from(elemsize).unwrap()
     }
@@ -373,7 +373,7 @@ impl ElemRestriction {
     /// assert_eq!(lsize, nelem + 1);
     /// ```
     pub fn lvector_size(&self) -> usize {
-        let mut lsize: i32 = 0;
+        let mut lsize = 0;
         unsafe { bind_ceed::CeedElemRestrictionGetLVectorSize(self.ptr, &mut lsize) };
         usize::try_from(lsize).unwrap()
     }
@@ -396,7 +396,7 @@ impl ElemRestriction {
     /// assert_eq!(n, ncomp, "Incorrect number of components");
     /// ```
     pub fn num_components(&self) -> usize {
-        let mut ncomp: i32 = 0;
+        let mut ncomp = 0;
         unsafe { bind_ceed::CeedElemRestrictionGetNumComponents(self.ptr, &mut ncomp) };
         usize::try_from(ncomp).unwrap()
     }
