### 2021-01-07 f361963
Rust - modify version numbers tests, cargo-tarpaulin and cargo have different calling paths in tests
