+++ rust/tests/version-numbers.rs
@@ -1,9 +1,23 @@
 #[test]
 fn test_readme_deps() {
-    version_sync::assert_markdown_deps_updated!("../README.rst");
+    let path = std::env::current_dir().unwrap();
+    let path = path.to_str().unwrap();
+    let length = path.len();
+    if &path[length-4..length] == "rust" {
+        version_sync::assert_markdown_deps_updated!("../README.rst");
+    } else {
+        version_sync::assert_markdown_deps_updated!("README.rst");
+    }
 }
 
 #[test]
 fn test_html_root_url() {
-    version_sync::assert_html_root_url_updated!("src/lib.rs");
+    let path = std::env::current_dir().unwrap();
+    let path = path.to_str().unwrap();
+    let length = path.len();
+    if &path[length-4..length] == "rust" {
+        version_sync::assert_html_root_url_updated!("src/lib.rs");
+    } else {
+        version_sync::assert_html_root_url_updated!("rust/src/lib.rs");
+    }
 }
