### 2021-01-07 84dc6ce
Examples - add separate Rust example quiet mode
