+++ examples/rust/ex1-volume/src/main.rs
@@ -51,6 +51,7 @@ fn example_1(options: opt::Opt) -> Result<(), String> {
         num_qpts,
         problem_size_requested,
         test,
+        quiet,
         gallery,
     } = options;
     assert!(dim >= 1 && dim <= 3);
@@ -66,7 +67,7 @@ fn example_1(options: opt::Opt) -> Result<(), String> {
     }
 
     // Summary output
-    if !test {
+    if !quiet {
         println!("Selected options: [command line option] : <current value>");
         println!("    Ceed specification [-c] : {}", ceed_spec);
         println!("    Mesh dimension     [-d] : {}", dim);
@@ -91,7 +92,7 @@ fn example_1(options: opt::Opt) -> Result<(), String> {
 
     // Determine mesh size from approximate problem size
     let num_xyz = mesh::cartesian_mesh_size(dim, solution_degree, problem_size);
-    if !test {
+    if !quiet {
         print!("\nMesh size                   : nx = {}", num_xyz[0]);
         if dim > 1 {
             print!(", ny = {}", num_xyz[1]);
@@ -110,7 +111,7 @@ fn example_1(options: opt::Opt) -> Result<(), String> {
         mesh::build_cartesian_restriction(&ceed, dim, num_xyz, solution_degree, 1, num_qpts);
     let mesh_size = restr_mesh.lvector_size();
     let solution_size = restr_solution.lvector_size();
-    if !test {
+    if !quiet {
         println!("Number of mesh nodes        : {}", mesh_size / dim);
         println!("Number of solution nodes    : {}", solution_size);
     }
@@ -246,29 +247,28 @@ fn example_1(options: opt::Opt) -> Result<(), String> {
     let volume: f64 = v.view().iter().sum();
 
     // Output results
-    if !test {
+    if !quiet {
         println!("Exact mesh volume           : {:.12}", exact_volume);
         println!("Computed mesh volume        : {:.12}", volume);
         println!(
             "Volume error                : {:.12e}",
             volume - exact_volume
         );
+    }
+    let tolerance = if dim == 1 {
+        1E-14
+    } else if dim == 2 {
+        1E-7
     } else {
-        let tolerance = if dim == 1 {
-            1E-14
-        } else if dim == 2 {
-            1E-7
-        } else {
-            1E-5
-        };
-        let error = (volume - exact_volume).abs();
-        if error > tolerance {
-            println!("Volume error: {:.12e}", error);
-            return Err(format!(
-                "Volume error too large - expected: {:.12e}, actual: {:.12e}",
-                tolerance, error
-            ));
-        }
+        1E-5
+    };
+    let error = (volume - exact_volume).abs();
+    if error > tolerance {
+        println!("Volume error too large: {:.12e}", error);
+        return Err(format!(
+            "Volume error too large - expected: {:.12e}, actual: {:.12e}",
+            tolerance, error
+        ));
     }
     Ok(())
 }
@@ -289,6 +289,7 @@ mod tests {
         let num_qpts = 6;
         let problem_size_requested = -1;
         let test = true;
+        let quiet = false;
         let gallery = false;
         let options = opt::Opt {
             ceed_spec,
@@ -298,6 +299,7 @@ mod tests {
             num_qpts,
             problem_size_requested,
             test,
+            quiet,
             gallery,
         };
         assert!(example_1(options).is_ok());
@@ -312,6 +314,7 @@ mod tests {
         let num_qpts = 6;
         let problem_size_requested = -1;
         let test = true;
+        let quiet = true;
         let gallery = false;
         let options = opt::Opt {
             ceed_spec,
@@ -321,6 +324,7 @@ mod tests {
             num_qpts,
             problem_size_requested,
             test,
+            quiet,
             gallery,
         };
[…]
