### 2021-01-07 e43e6ad
Rust - simplify version tests
