+++ rust/tests/version-numbers.rs
@@ -2,22 +2,20 @@
 fn test_readme_deps() {
     let path = std::env::current_dir().unwrap();
     let path = path.to_str().unwrap();
-    let length = path.len();
-    if &path[length - 4..length] == "rust" {
-        version_sync::assert_markdown_deps_updated!("../README.rst");
+    version_sync::assert_markdown_deps_updated!(if &path[path.len() - 4..path.len()] == "rust" {
+        "../README.rst"
     } else {
-        version_sync::assert_markdown_deps_updated!("README.rst");
-    }
+        "README.rst"
+    });
 }
 
 #[test]
 fn test_html_root_url() {
     let path = std::env::current_dir().unwrap();
     let path = path.to_str().unwrap();
-    let length = path.len();
-    if &path[length - 4..length] == "rust" {
-        version_sync::assert_html_root_url_updated!("src/lib.rs");
+    version_sync::assert_html_root_url_updated!(if &path[path.len() - 4..path.len()] == "rust" {
+        "src/lib.rs"
     } else {
-        version_sync::assert_html_root_url_updated!("rust/src/lib.rs");
-    }
+        "rust/src/lib.rs"
+    });
 }
