### 2021-01-07 e436547
CodeCov - try adding gallery to 'fixes'
