+++ examples/ceed/ex1-volume.c
@@ -39,9 +39,11 @@
 //
 // Next line is grep'd from tap.sh to set its arguments
 // Test in 1D-3D
+//TESTARGS -ceed {ceed_resource} -d 1 -t
 //TESTARGS -ceed {ceed_resource} -d 2 -t
 //TESTARGS -ceed {ceed_resource} -d 3 -t
 //TESTARGS -ceed {ceed_resource} -d 1 -t -g
+//TESTARGS -ceed {ceed_resource} -d 2 -t -g
 //TESTARGS -ceed {ceed_resource} -d 3 -t -g
 
 /// @file
