### 2021-01-07 7295fff
Rust - test 3d examples without quiet flag
