### 2021-01-07 569a46b
Actions - use gcc-10 for compatibility in gcov reports
