+++ .github/workflows/rust-test-with-style.yml
@@ -8,7 +8,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-20.04]
-        compiler: [gcc-9]
+        compiler: [gcc-10]
 
     runs-on: ${{ matrix.os }}
 
