+++ examples/rust/ex1-volume/src/main.rs
@@ -356,9 +356,59 @@ mod tests {
     }
 
     #[test]
+    fn example_1_1d_gallery() {
+        let ceed_spec = "/cpu/self/ref/serial".to_string();
+        let dim = 1;
+        let mesh_degree = 4;
+        let solution_degree = 4;
+        let num_qpts = 6;
+        let problem_size_requested = -1;
+        let test = true;
+        let quiet = true;
+        let gallery = true;
+        let options = opt::Opt {
+            ceed_spec,
+            dim,
+            mesh_degree,
+            solution_degree,
+            num_qpts,
+            problem_size_requested,
+            test,
+            quiet,
+            gallery,
+        };
+        assert!(example_1(options).is_ok());
+    }
+
+    #[test]
     fn example_1_2d_gallery() {
         let ceed_spec = "/cpu/self/ref/serial".to_string();
         let dim = 2;
+        let mesh_degree = 4;
+        let solution_degree = 4;
+        let num_qpts = 6;
+        let problem_size_requested = -1;
+        let test = true;
+        let quiet = true;
+        let gallery = true;
+        let options = opt::Opt {
+            ceed_spec,
+            dim,
+            mesh_degree,
+            solution_degree,
+            num_qpts,
+            problem_size_requested,
+            test,
+            quiet,
+            gallery,
+        };
+        assert!(example_1(options).is_ok());
+    }
+
+    #[test]
+    fn example_1_3d_gallery() {
+        let ceed_spec = "/cpu/self/ref/serial".to_string();
+        let dim = 3;
         let mesh_degree = 4;
         let solution_degree = 4;
         let num_qpts = 6;
