### 2021-01-07 5b8f7ec
Gitlab - make it easier to find complier info in logs
