+++ .github/workflows/rust-test-with-style.yml
@@ -31,6 +31,7 @@ jobs:
     - name: Codecov upload
       uses: codecov/codecov-action@v1.0.2
       with:
+        directory: target/tarpaulin
         token: ${{secrets.CODECOV_TOKEN}}
 
   style:
