### 2021-01-08 9879cfa
Actions - exclude non-rust files from rust coverage
