+++ .github/workflows/rust-test-with-style.yml
@@ -27,7 +27,7 @@ jobs:
         CC: ${{ matrix.compiler }}
         FC: gfortran-9
       with:
-        args: '--run-types Doctests Tests'
+        args: '--run-types Doctests Tests --exclude-files backends/* gallery/* include/* interface/*'
     - name: Codecov upload
       uses: codecov/codecov-action@v1.0.2
       with:
