### 2021-01-12 7fdaa51
Cov - exclude weak symbols code, not useful cov info
