+++ gallery/ceed-gallerytemplate.h
@@ -26,7 +26,7 @@
 
 #ifndef gallerytemplate_h
 #define gallerytemplate_h
-
+// LCOV_EXCL_START
 CEED_QFUNCTION(GalleryTemplate)(void *ctx, const CeedInt Q,
                                 const CeedScalar *const *in,
                                 CeedScalar *const *out) {
@@ -43,5 +43,5 @@ CEED_QFUNCTION(GalleryTemplate)(void *ctx, const CeedInt Q,
 
   return 0;
 }
-
+// LCOV_EXCL_STOP
 #endif // gallerytemplate_h
