### 2021-01-12 153a420
Examples - clean up options creation for tests in Rust examples
