### 2021-01-12 9eacf7e
Examples - simplify tolerance for testing Rust examples
