### 2021-01-12 2ab01f6
Examples - common Rust example code in shared local crate
