+++ examples/rust/ex1-volume/src/mesh.rs
@@ -1,207 +1,0 @@
-// Copyright (c) 2017-2021, Lawrence Livermore National Security, LLC.
-// Produced at the Lawrence Livermore National Laboratory. LLNL-CODE-734707.
-// All Rights reserved. See files LICENSE and NOTICE for details.
-//
-// This file is part of CEED, a collection of benchmarks, miniapps, software
-// libraries and APIs for efficient high-order finite element and spectral
-// element discretizations for exascale applications. For more information and
-// source code availability see http://github.com/ceed.
-//
-// The CEED research is supported by the Exascale Computing Project 17-SC-20-SC,
-// a collaborative effort of two U.S. Department of Energy organizations (Office
-// of Science and the National Nuclear Security Administration) responsible for
-// the planning and preparation of a capable exascale ecosystem, including
-// software, applications, hardware, advanced system engineering and early
-// testbed platforms, in support of the nation's exascale computing imperative.
-
-use libceed::{prelude::*, Ceed};
-
-// ----------------------------------------------------------------------------
-// Determine problem size in each dimension from size and dimenison
-// ----------------------------------------------------------------------------
-pub(crate) fn cartesian_mesh_size(
-    dim: usize,
-    solution_degree: usize,
-    problem_size: i64,
-) -> [usize; 3] {
-    // Use the approximate formula:
-    //    prob_size ~ num_elem * degree^dim
-    let mut num_elem = problem_size / solution_degree.pow(dim as u32) as i64;
-    let mut s = 0; // find s: num_elem / 2 < 2^s <= num_elem
-    while num_elem > 1 {
-        num_elem /= 2;
-        s += 1;
-    }
-
-    // Size per dimension
-    let mut r = s % dim;
-    let mut num_xyz = [0; 3];
-    for d in 0..dim {
-        let mut sd = s / dim;
-        if r > 0 {
-            sd += 1;
-            r -= 1;
-        }
-        num_xyz[d] = 1 << sd;
-    }
-    num_xyz
-}
-
-// ----------------------------------------------------------------------------
-// Build element restriction objects for the mesh
-// ----------------------------------------------------------------------------
-pub(crate) fn build_cartesian_restriction(
-    ceed: &Ceed,
-    dim: usize,
-    num_xyz: [usize; 3],
-    degree: usize,
-    num_comp: usize,
-    num_qpts: usize,
-) -> (ElemRestriction, ElemRestriction) {
-    let p = degree + 1;
-    let num_nodes = p.pow(dim as u32); // number of nodes per element
-    let elem_qpts = num_qpts.pow(dim as u32); // number of quadrature pts per element
-
-    // Problem dimensions
-    let mut num_d = [0; 3];
-    let mut num_elem = 1;
-    let mut scalar_size = 1;
-    for d in 0..dim {
-        num_elem *= num_xyz[d];
-        num_d[d] = num_xyz[d] * (p - 1) + 1;
-        scalar_size *= num_d[d];
-    }
-
-    // elem:         0             1                 n-1
-    //        |---*-...-*---|---*-...-*---|- ... -|--...--|
-    // nodes: 0   1    p-1  p  p+1       2*p             n*p
-    let mut elem_nodes = vec![0; num_elem * num_nodes];
-    for e in 0..num_elem {
-        let mut e_xyz = [1; 3];
-        let mut re = e;
-        for d in 0..dim {
-            e_xyz[d] = re % num_xyz[d];
-            re /= num_xyz[d];
-        }
-        let loc_offset = e * num_nodes;
[…]
