+++ examples/rust/mesh/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "ex2-surface"
+name = "mesh"
 version = "0.1.0"
 authors = [
     "Jeremy L Thompson <thompson.jeremy.luke@gmail.com>",
@@ -9,4 +9,3 @@ edition = "2018"
 [dependencies]
 structopt = { version = "0.3", default-features = false }
 libceed = { path = "../../../rust" }
-mesh = { path = "../mesh" }

[…]
