### 2021-01-12 ded9b81
Rust - add examples (#683)
* Rust - initial version of ex1

Example - move closure

Co-authored-by: Jed Brown <jed@jedbrown.org>

typo fix

Co-authored-by: Jed Brown <jed@jedbrown.org>

minor tidying

order -> degree

drop operator applicaiton messages

* rust: make top-level workspace and use libceed from path

This allows a single "cargo test" from the top level.

* rust: add version-sync (to help us remember when we publish a crate)

* examples - consistency update

* Rust - add Cargo.lock to top level gitignore

* rust - refactor example to add tests

* Rust - test and style from workspace
[…]
