+++ rust/src/qfunction.rs
@@ -71,8 +71,8 @@ pub(crate) struct QFunctionCore {
 struct QFunctionTrampolineData {
     number_inputs: usize,
     number_outputs: usize,
-    input_sizes: [i32; MAX_QFUNCTION_FIELDS],
-    output_sizes: [i32; MAX_QFUNCTION_FIELDS],
+    input_sizes: [usize; MAX_QFUNCTION_FIELDS],
+    output_sizes: [usize; MAX_QFUNCTION_FIELDS],
     user_f: Box<QFunctionUserClosure>,
 }
 
@@ -175,7 +175,7 @@ impl fmt::Display for QFunctionByName {
 // -----------------------------------------------------------------------------
 impl QFunctionCore {
     // Common implementation
-    pub fn apply(&self, Q: i32, u: &[Vector], v: &[Vector]) {
+    pub fn apply(&self, Q: usize, u: &[Vector], v: &[Vector]) {
         let mut u_c = [std::ptr::null_mut(); MAX_QFUNCTION_FIELDS];
         for i in 0..std::cmp::min(MAX_QFUNCTION_FIELDS, u.len()) {
             u_c[i] = u[i].ptr;
@@ -184,6 +184,7 @@ impl QFunctionCore {
         for i in 0..std::cmp::min(MAX_QFUNCTION_FIELDS, v.len()) {
             v_c[i] = v[i].ptr;
         }
+        let Q = i32::try_from(Q).unwrap();
         unsafe {
             bind_ceed::CeedQFunctionApply(self.ptr, Q, u_c.as_mut_ptr(), v_c.as_mut_ptr());
         }
@@ -219,7 +220,7 @@ unsafe extern "C" fn trampoline(
         .iter()
         .enumerate()
         .map(|(i, &x)| {
-            std::slice::from_raw_parts(x, (trampoline_data.input_sizes[i] * q) as usize) as &[f64]
+            std::slice::from_raw_parts(x, trampoline_data.input_sizes[i] * q as usize) as &[f64]
         })
         .zip(inputs_array.iter_mut())
         .for_each(|(x, a)| *a = x);
@@ -232,7 +233,7 @@ unsafe extern "C" fn trampoline(
         .iter()
         .enumerate()
         .map(|(i, &x)| {
-            std::slice::from_raw_parts_mut(x, (trampoline_data.output_sizes[i] * q) as usize)
+            std::slice::from_raw_parts_mut(x, trampoline_data.output_sizes[i] * q as usize)
                 as &mut [f64]
         })
         .zip(outputs_array.iter_mut())
@@ -247,7 +248,7 @@ unsafe extern "C" fn trampoline(
 // -----------------------------------------------------------------------------
 impl QFunction {
     // Constructor
-    pub fn create(ceed: &crate::Ceed, vlength: i32, user_f: Box<QFunctionUserClosure>) -> Self {
+    pub fn create(ceed: &crate::Ceed, vlength: usize, user_f: Box<QFunctionUserClosure>) -> Self {
         let source_c = CString::new("").expect("CString::new failed");
         let mut ptr = std::ptr::null_mut();
 
@@ -267,6 +268,7 @@ impl QFunction {
         };
 
         // Create QFunction
+        let vlength = i32::try_from(vlength).unwrap();
         unsafe {
             bind_ceed::CeedQFunctionCreateInterior(
                 ceed.ptr,
@@ -331,7 +333,7 @@ impl QFunction {
     /// let mut u = [0.; Q];
     /// let mut v = [0.; Q];
     ///
-    /// for i in 0..Q as usize {
+    /// for i in 0..Q {
     ///   let x = 2. * (i as f64)/((Q as f64) - 1.) - 1.;
     ///   u[i] = 2. + 3. * x + 5. * x * x;
[…]
