+++ examples/rust/ex2-surface/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "ex2-surface"
+version = "0.1.0"
+authors = [
+    "Jeremy L Thompson <thompson.jeremy.luke@gmail.com>",
+]
+edition = "2018"
+
+[dependencies]
+structopt = { version = "0.3", default-features = false }
+libceed = { path = "../../../rust" }
+mesh = { path = "../mesh" }

[…]
