### 2021-01-12 7aa233f
Examples - drop unused dep in common Rust example code crate
