+++ examples/rust/mesh/Cargo.toml
@@ -7,5 +7,4 @@ authors = [
 edition = "2018"
 
 [dependencies]
-structopt = { version = "0.3", default-features = false }
 libceed = { path = "../../../rust" }
