+++ examples/rust/ex1-volume/src/main.rs
@@ -321,7 +321,7 @@ mod tests {
             num_qpts: 6,
             problem_size_requested: -1,
             test: true,
-            quiet: true,
+            quiet: false,
             gallery: false,
         };
         assert!(example_1(options).is_ok());
