### 2021-01-12 308cbf1
Examples - also exclude main driver from tests in Rust ex
