### 2021-01-12 954a603
Actions - Rust workspace moved target dir for Rust docs (#694)
* Actions - Rust workspace moved target dir for Rust docs

* Examples - drop unused dep in common Rust example code crate

* Examples - restore accidentally missed lines

* Examples - ignore structopt testing in Rust examples

* Examples - also exclude main driver from tests in Rust ex
