+++ examples/rust/ex2-surface/src/main.rs
@@ -36,8 +36,9 @@ mod opt;
 mod transform;
 
 // ----------------------------------------------------------------------------
-// Example 1
+// Example 2
 // ----------------------------------------------------------------------------
+#[cfg(not(tarpaulin_include))]
 fn main() -> Result<(), String> {
     let options = opt::Opt::from_args();
     example_2(options)
@@ -402,7 +403,7 @@ mod tests {
             num_qpts: 6,
             problem_size_requested: -1,
             test: true,
-            quiet: true,
+            quiet: false,
             gallery: false,
         };
         assert!(example_2(options).is_ok());
