### 2021-01-13 925dff3
removed the forcing term of case 4
