+++ examples/fluids/euler-vortex.h
@@ -102,7 +102,7 @@ static inline int Exact_Euler(CeedInt dim, CeedScalar time,
   const CeedScalar S = (gamma - 1.) * vortex_strength * vortex_strength /
                        (8.*gamma*M_PI*M_PI);
 
-  // Exact Solutions
+  // Exact Solutions (traveling vortex)
   //const CeedScalar rho = 1.;
   //const CeedScalar P = 1.;
   //const CeedScalar E = 2.;
@@ -114,7 +114,6 @@ static inline int Exact_Euler(CeedInt dim, CeedScalar time,
 
   // Initial Conditions
   if (0) { // Case 1: constant zero velocity, density constant, total energy constant
-
     const CeedScalar rho = 1.;
     const CeedScalar P = 1.;
     const CeedScalar E = 2.;
@@ -127,7 +126,6 @@ static inline int Exact_Euler(CeedInt dim, CeedScalar time,
     q[4] = E;
   }
   if (0) { // Case 2: constant nonzero velocity, density constant, total energy constant
-
     const CeedScalar rho = 1.;
     const CeedScalar P = 1.;
     const CeedScalar E = 2.;
@@ -143,26 +141,22 @@ static inline int Exact_Euler(CeedInt dim, CeedScalar time,
     // (so density and internal energy will be non-constant),
     // but the velocity should stay zero and the bubble won't diffuse
     // (for Euler, where there is no thermal conductivity)
-
     const CeedScalar P = 1.;
-    const CeedScalar T = 1. - S * exp(1. - r*r); // I am not sure!!!
+    const CeedScalar T = 1. - S * exp(1. - r*r);
     const CeedScalar rho = P / (R*T);
-    //const CeedScalar E = 2.;
     const CeedScalar u[3] = {0., 0., 0.};
 
     q[0] = rho;
     q[1] = rho * u[0];
     q[2] = rho * u[1];
     q[3] = rho * u[2];
-    q[4] = rho * ( cv * T + (u[0]*u[0] +
-                             u[1]*u[1])/2. ); // It doesn't really change!
+    q[4] = rho * ( cv * T + (u[0]*u[0] + u[1]*u[1])/2. );
   }
   if (1) { // Case 4: constant nonzero velocity, pressure constant
     // (so density and internal energy will be non-constant),
     // it should be transported across the domain, but velocity stays constant
-
     const CeedScalar P = 1.;
-    const CeedScalar T = 1. - S * exp(1. - r*r); // P / rho - S * exp(1. - r*r);
+    const CeedScalar T = 1. - S * exp(1. - r*r);
     const CeedScalar rho = P / (R*T);
     const CeedScalar u[3] = {1.1, 1.2, 0.};
 
@@ -268,12 +262,12 @@ static inline int MMSforce_Euler(CeedInt dim, CeedScalar time,
   const CeedScalar C = vortex_strength * exp((1. - r*r)/2.) / (2. * M_PI);
   const CeedScalar S = (gamma - 1.) * vortex_strength * vortex_strength /
                        (8.*gamma*M_PI*M_PI);
-  CeedScalar u[3] = {etv_mean_velocity[0] - C*y0,
-                     etv_mean_velocity[1] + C*x0,
-                     0.
-                    };
+  const CeedScalar u[3] = {etv_mean_velocity[0] - C*y0,
+                           etv_mean_velocity[1] + C*x0,
+                           0.
+                          };
 
-  // Forcing term for Manufactured solution
+  // Forcing term for Manufactured solution (traveling vortex)
   if (0) {
     force[0] = 0.;
[…]
