### 2021-01-14 2561194
removed periodicity
