### 2021-01-14 2177edd
removed periodicity
