### 2021-01-14 8741176
Added required variables to surface context
