+++ examples/fluids/navierstokes.c
@@ -941,6 +941,8 @@ int main(int argc, char **argv) {
   CeedScalar cv              = 717.;     // J/(kg K)
   CeedScalar cp              = 1004.;    // J/(kg K)
   CeedScalar vortex_strength = 5.;       // -
+  CeedScalar T_inlet         = 300.;     // K
+  CeedScalar P_inlet         = 1.e5;     // Pa
   CeedScalar g               = 9.81;     // m/s^2
   CeedScalar lambda          = -2./3.;   // -
   CeedScalar mu              = 75.;      // Pa s, dynamic viscosity
@@ -1091,6 +1093,24 @@ int main(int argc, char **argv) {
                        "Warning! Use -vortex_strength only with -problem euler_vortex\n");
     CHKERRQ(ierr);
   }
+  PetscBool userTinlet;
+  ierr = PetscOptionsScalar("-T_inlet", "Incoming Temperature",
+                            NULL, T_inlet, &T_inlet, &userTinlet);
+  CHKERRQ(ierr);
+  if (problemChoice != NS_EULER_VORTEX && userTinlet) {
+    ierr = PetscPrintf(comm,
+                       "Warning! Use -T_inlet only with -problem euler_vortex\n");
+    CHKERRQ(ierr);
+  }
+  PetscBool userPinlet;
+  ierr = PetscOptionsScalar("-P_inlet", "Incoming Pressure",
+                            NULL, P_inlet, &P_inlet, &userPinlet);
+  CHKERRQ(ierr);
+  if (problemChoice != NS_EULER_VORTEX && userPinlet) {
+    ierr = PetscPrintf(comm,
+                       "Warning! Use -P_inlet only with -problem euler_vortex\n");
+    CHKERRQ(ierr);
+  }
   ierr = PetscOptionsScalar("-g", "Gravitational acceleration",
                             NULL, g, &g, NULL); CHKERRQ(ierr);
   ierr = PetscOptionsScalar("-lambda",
@@ -1199,6 +1219,8 @@ int main(int argc, char **argv) {
   cv *= JperkgK;
   cp *= JperkgK;
   Rd = cp - cv;
+  T_inlet *= Kelvin;
+  P_inlet *= Pascal;
   g *= mpersquareds;
   mu *= Pascal * second;
   k *= WpermK;
@@ -1563,6 +1585,11 @@ int main(int argc, char **argv) {
   struct SurfaceContext_ ctxSurfaceData = {
     .E_wind = E_wind,
     .strong_form = strong_form,
+    .T_inlet = T_inlet,
+    .P_inlet = P_inlet,
+    .etv_mean_velocity[0] = etv_mean_velocity[0],
+    .etv_mean_velocity[1] = etv_mean_velocity[1],
+    .etv_mean_velocity[2] = etv_mean_velocity[2],
     .implicit = implicit,
   };
   CeedQFunctionContextCreate(ceed, &ctxSurface);
@@ -1598,7 +1625,7 @@ int main(int argc, char **argv) {
   case NS_EULER_VORTEX:
     if (qf_ics) CeedQFunctionSetContext(qf_ics, ctxEuler);
     if (qf_rhsVol) CeedQFunctionSetContext(qf_rhsVol, ctxEuler);
-    if (qf_applySur) CeedQFunctionSetContext(qf_applySur, ctxEuler);
+    if (qf_applySur) CeedQFunctionSetContext(qf_applySur, ctxSurface);
   }
 
   // Set up PETSc context
