+++ examples/fluids/euler-vortex.h
@@ -45,6 +45,19 @@ struct EulerContext_ {
 };
 #endif
 
+#ifndef surface_context_struct
+#define surface_context_struct
+typedef struct SurfaceContext_ *SurfaceContext;
+struct SurfaceContext_ {
+  CeedScalar E_wind;               // advection(2d)
+  CeedScalar strong_form;          // advection(2d)
+  CeedScalar T_inlet;              // euler_vortex
+  CeedScalar P_inlet;              // euler_vortex
+  CeedScalar etv_mean_velocity[3]; // euler_vortex
+  bool implicit;
+};
+#endif
+
 // *****************************************************************************
 // This function sets the initial conditions and the boundary conditions
 //
