+++ examples/fluids/navierstokes.c
@@ -879,10 +879,10 @@ static PetscErrorCode SetUpDM(DM dm, problemData *problem, PetscInt degree,
                              3, comps, (void(*)(void))problem->bc, NULL,
                              bc->nwall, bc->walls, ctxSetupData); CHKERRQ(ierr);
       } else if (problem->bc == Exact_Euler) {
-        PetscInt bcMMS[4] = {3, 4, 5, 6};
-        ierr = DMAddBoundary(dm, DM_BC_ESSENTIAL, "mms", "Face Sets", 0,
-                             0, NULL, (void(*)(void))problem->bc, NULL,
-                             4, bcMMS, ctxMMS); CHKERRQ(ierr);
+        PetscInt comps[3] = {1, 2, 3};
+        ierr = DMAddBoundary(dm, DM_BC_ESSENTIAL, "wall", "Face Sets", 0,
+                             3, comps, (void(*)(void))problem->bc, NULL,
+                             bc->nwall, bc->walls, ctxMMS); CHKERRQ(ierr);
       } else
         SETERRQ(PETSC_COMM_SELF, PETSC_ERR_ARG_NULL,
                 "Undefined boundary conditions for this problem");
