### 2021-01-17 38fc7e6
Default P_inlet 1. -> 1.e3 so that it becomes 1 after the unit conversion
