### 2021-01-17 cfc558c
Default P_inlet 1. -> 1.e3 so that it becomes 1 after the unit conversion
