### 2021-01-19 9eacc1a
ETV: added more specific conditional statements for the in/outflow formulations
