### 2021-01-19 a3ddc43
ETV: use enum and command line options for the test cases
