### 2021-01-19 2d7ca1a
ETV: use enum and command line options for the test cases
