### 2021-01-19 3a27e6b
use a better command line for turning off time stepping
