### 2021-01-20 b2f2822
ETV: fixed the zero velocity test cases on the boundaries (had to apply EulerContext for the boundaries as well)
