### 2021-01-20 a45a933
ETV: fixed the zero velocity test cases on the boundaries (had to apply EulerContext for the boundaries as well)
