### 2021-01-20 95d07a4
examples/fluids: fix stale debugging pressure
