### 2021-01-21 f3aea92
examples/fluids: fix Euler vortex initial condition
