### 2021-01-21 3d6e121
ETV: turn off wall and slip BCs ( in the x and y directions) in SetUpDM() instead of CreateOperatorForDomain()
