### 2021-01-21 b3bd059
ETV: turn off wall and slip BCs ( in the x and y directions) in SetUpDM() instead of CreateOperatorForDomain()
