### 2021-01-21 7bd142c
examples/fluids: fix entropy for vortex vs bubble
