### 2021-01-21 b36211b
ETV: Changed the conditional statement for identifying in/outflow faces to fix the artifact in linear translation cases
