### 2021-01-21 9dd8864
remove constant memory usage for hip-shared
