### 2021-01-21 54dd8e9
remove constant array pointers from hip-shared basis data struct
