### 2021-01-22 17515be
ETV: added lz=1 and fixed max time to the arguments of the convergence tests
