### 2021-01-23 CEED/libCEED#696 @FreddieWitherden
Error/Exception Handling
I am currently reviewing libCEED as part of https://github.com/openjournals/joss-reviews/issues/2945.

One issue I've identified is around how errors are handled.  Specifically, standard operating procedure at the moment is along the lines of:
```
ierr = MightFail(); CeedChk(ierr);
```
where `CeedChk` expands to a return statement should `ierr` be non-zero.  This paradigm is problematic as it typically leaves the entity which the function is operating on in an undefined state.  This makes it virtually impossible for a user of libCEED do any meaningful error handling.  Moreover, it appears as if there several situations where this can result in memory being leaked; if a method allocates some memory (which succeeds) but then calls another internal method (which fails) the current code will immediately return without first freeing the temporary object it allocated.

There are also a few oddities.  For example:
https://github.com/CEED/libCEED/blob/06988bf74cc6ac18eacafe7930f080803395ba29/interface/ceed-operator.c#L67-L68
where the second `ierr` check appears to be redundant.
