### 2021-01-23 CEED/libCEED#697 @FreddieWitherden
Obtaining libCEED version at compile time/run time
I am currently reviewing libCEED as part of https://github.com/openjournals/joss-reviews/issues/2945.

Looking through the documentation/header files I am wondering how one can obtain the libCEED version at both compile time and run time (the latter being useful for those who are interfacing at the ABI level).
