### 2021-01-24 CEED/libCEED#696 @knepley
Don't you just use:
`ierr = MightFail();
if (ierr) {
  <do you cleanup>
  CeedChk(ierr);
}`
This is exactly what an exception implementation would do, without the bizarre exception discovery problems.
