### 2021-01-24 CEED/libCEED#696 @jedbrown
Thanks @FreddieWitherden. The case you mention is definitely a (harmless but for reader confusion) mistake. The broader complaint is super valid. libCEED has prioritized error behavior that is useful for callers (and library developers) to fix their code, which is usually what you want for compiled code, but not for interactive work where one wants to keep a valid state in a console/notebook environment. A common issue is that lower-level libraries (such as MPI) don't ensure leak-free valid state on errors, but since we want libCEED to be usable from interactive environments, I agree that it would make sense to at least communicate when an error is returning with valid state.

This becomes a really tall order (hard to implement correctly and hard to test) in the general case without type-level support, but I think we should at least cover the sort of cases that a Python or Julia user is likely to encounter in interactive environments (shape mismatch, incomplete specification of fields).

Cc: @jeremylt
