### 2021-01-24 1070991
Add version macros and CeedGetVersion() [fix #697]
Requested-by: Freddie Witherden <freddie@witherden.org>
