### 2021-01-24 CEED/libCEED#696 @FreddieWitherden
> Thanks @FreddieWitherden. The case you mention is definitely a (harmless but for reader confusion) mistake. The broader complaint is super valid. libCEED has prioritized error behavior that is useful for callers (and library developers) to fix their code, which is usually what you want for compiled code, but not for interactive work where one wants to keep a valid state in a console/notebook environment. A common issue is that lower-level libraries (such as MPI) don't ensure leak-free valid state on errors, but since we want libCEED to be usable from interactive environments, I agree that it would make sense to at least communicate when an error is returning with valid state.

In instances for errors are typically the result of user-error I've found the following two strategies helpful.

The first is simply to backtrace and exit upon encountering such an error.  This is nice as there is no possibility for a user to miss an error (say because they forgot to check a return code) and they always get a backtrace.  MPI implementations will sometimes do this if you really irk them, or at the least immediately write something out to stderr.

Another is to save the error state in a field variable on the associated object and check it before each API call.  For example
```
int myfun(Ceed *ceed)
{
    if (ceed->err) return CEED_INVALID_STATE;

    // ...
    int ierr = MightFail(); if (ierr) { ceed->err = ierr; ceed->err_file = __FILE__; ceed->err_line =  __LINE__; return ierr; }
}
```
then have a function such as `int CeedGetAndResetErr(ceed)` which fetches the error and resets the internal state.  (With similar ideas holding for related objects.)  This allows users of the C library to chain a set of mundane functions they do not expect will fail and handle the errors towards the end of a block but without a risk of missing where the error occurred.
