### 2021-01-25 02450a2
examples/fluids: Max Error -> Relative Error
