+++ examples/fluids/stdoutParsing.py
@@ -4,7 +4,7 @@ from pathlib import Path
 
 # Regex to parse STDOUT of the navierstokes run
 logreg = re.compile(
-    r".*(?:^Degree of FEM Space: (\d+)).*(?:^Global FEM nodes: (\d{2,})).*(?:^dm_plex_box_faces: (\S+)).*(?:^Time taken for solution: (\d*\.?\d+)).*(?:^Max Error: (\d*\.?\d+))",
+    r".*(?:^Degree of FEM Space: (\d+)).*(?:^Global FEM nodes: (\d{2,})).*(?:^dm_plex_box_faces: (\S+)).*(?:^Time taken for solution: (\d*\.?\d+)).*(?:^Relative Error: (\d*\.?\d+))",
     re.DOTALL | re.MULTILINE,
 )
 
