### 2021-01-25 ce58a7c
examples/fluids: Max Error -> Relative Error
