### 2021-01-25 bdedb62
examples/fluids/convergence_plot: Some clean-up
