### 2021-01-25 c915279
fluids/examples/verification-test.sh: Some clean-up
