### 2021-01-25 a81047f
examples/fluids/convergence_plot.py: added copyright
