### 2021-01-26 35efc48
examples/fluids: Pipe convergence test results to a CSV file and use it as an input for the plotting
