### 2021-01-26 50c0860
examples/fluids: Renamed files related to the convergence study so that they all start with conv_
