### 2021-01-26 4b32fe0
examples/fluids: Take the CSV file from the user for the plotting
