### 2021-01-26 bb7933b
error - add error enum
