### 2021-01-26 387335a
error - add error enum
