### 2021-01-26 7f386f6
error - update error enum names and numbering
