### 2021-01-26 a5956e7
error - update error enum names and numbering
