+++ backends/blocked/ceed-blocked-operator.c
@@ -318,7 +318,8 @@ static inline int CeedOperatorInputBasis_Blocked(CeedInt e, CeedInt Q,
       CeedChk(ierr);
       Ceed ceed;
       ierr = CeedBasisGetCeed(basis, &ceed); CeedChk(ierr);
-      return CeedError(ceed, 1, "Ceed evaluation mode not implemented");
+      return CeedError(ceed, CEED_ERROR_BACKEND,
+                       "Ceed evaluation mode not implemented");
       // LCOV_EXCL_STOP
     }
     }
@@ -379,14 +380,16 @@ static inline int CeedOperatorOutputBasis_Blocked(CeedInt e, CeedInt Q,
     case CEED_EVAL_WEIGHT: {
       Ceed ceed;
       ierr = CeedOperatorGetCeed(op, &ceed); CeedChk(ierr);
-      return CeedError(ceed, 1, "CEED_EVAL_WEIGHT cannot be an output "
+      return CeedError(ceed, CEED_ERROR_BACKEND,
+                       "CEED_EVAL_WEIGHT cannot be an output "
                        "evaluation mode");
     }
     case CEED_EVAL_DIV:
     case CEED_EVAL_CURL: {
       Ceed ceed;
       ierr = CeedOperatorGetCeed(op, &ceed); CeedChk(ierr);
-      return CeedError(ceed, 1, "Ceed evaluation mode not implemented");
+      return CeedError(ceed, CEED_ERROR_BACKEND,
+                       "Ceed evaluation mode not implemented");
       // LCOV_EXCL_STOP
     }
     }
@@ -558,7 +561,8 @@ static int CeedOperatorLinearAssembleQFunction_Blocked(CeedOperator op,
   // Check for identity
   if (impl->identityqf)
     // LCOV_EXCL_START
-    return CeedError(ceed, 1, "Assembling identity qfunctions not supported");
+    return CeedError(ceed, CEED_ERROR_BACKEND,
+                     "Assembling identity qfunctions not supported");
   // LCOV_EXCL_STOP
 
   // Input Evecs and Restriction
@@ -603,7 +607,8 @@ static int CeedOperatorLinearAssembleQFunction_Blocked(CeedOperator op,
   // Check sizes
   if (!numactivein || !numactiveout)
     // LCOV_EXCL_START
-    return CeedError(ceed, 1, "Cannot assemble QFunction without active inputs "
+    return CeedError(ceed, CEED_ERROR_BACKEND,
+                     "Cannot assemble QFunction without active inputs "
                      "and outputs");
   // LCOV_EXCL_STOP
 
