+++ backends/opt/ceed-opt-blocked.c
@@ -40,7 +40,8 @@ static int CeedInit_Opt_Blocked(const char *resource, Ceed ceed) {
   if (strcmp(resource, "/cpu/self") && strcmp(resource, "/cpu/self/opt")
       && strcmp(resource, "/cpu/self/opt/blocked"))
     // LCOV_EXCL_START
-    return CeedError(ceed, 1, "Opt backend cannot use resource: %s", resource);
+    return CeedError(ceed, CEED_ERROR_BACKEND,
+                     "Opt backend cannot use resource: %s", resource);
   // LCOV_EXCL_STOP
   ierr = CeedSetDeterministic(ceed, true); CeedChk(ierr);
 
