+++ backends/memcheck/ceed-memcheck-serial.c
@@ -27,7 +27,8 @@ static int CeedInit_Memcheck(const char *resource, Ceed ceed) {
   if (strcmp(resource, "/cpu/self/memcheck")
       && strcmp(resource, "/cpu/self/memcheck/serial"))
     // LCOV_EXCL_START
-    return CeedError(ceed, 1, "Valgrind Memcheck backend cannot use resource: %s",
+    return CeedError(ceed, CEED_ERROR_BACKEND,
+                     "Valgrind Memcheck backend cannot use resource: %s",
                      resource);
   // LCOV_EXCL_STOP
 
