+++ backends/cuda-shared/ceed-cuda-shared.c
@@ -29,7 +29,8 @@ static int CeedInit_Cuda_shared(const char *resource, Ceed ceed) {
   const int nrc = 9; // number of characters in resource
   if (strncmp(resource, "/gpu/cuda/shared", nrc))
     // LCOV_EXCL_START
-    return CeedError(ceed, 1, "Cuda backend cannot use resource: %s", resource);
+    return CeedError(ceed, CEED_ERROR_BACKEND,
+                     "Cuda backend cannot use resource: %s", resource);
   // LCOV_EXCL_STOP
   ierr = CeedSetDeterministic(ceed, true); CeedChk(ierr);
 

[…]
