+++ backends/cuda/ceed-cuda-restriction.c
@@ -203,7 +203,8 @@ int CeedElemRestrictionApplyBlock_Cuda(CeedElemRestriction r, CeedInt block,
   int ierr;
   Ceed ceed;
   ierr = CeedElemRestrictionGetCeed(r, &ceed); CeedChk(ierr);
-  return CeedError(ceed, 1, "Backend does not implement blocked restrictions");
+  return CeedError(ceed, CEED_ERROR_BACKEND,
+                   "Backend does not implement blocked restrictions");
   // LCOV_EXCL_STOP
 }
 
@@ -429,7 +430,8 @@ int CeedElemRestrictionCreate_Cuda(CeedMemType mtype, CeedCopyMode cmode,
     }
   } else {
     // LCOV_EXCL_START
-    return CeedError(ceed, 1, "Only MemType = HOST or DEVICE supported");
+    return CeedError(ceed, CEED_ERROR_BACKEND,
+                     "Only MemType = HOST or DEVICE supported");
     // LCOV_EXCL_STOP
   }
 
@@ -477,6 +479,7 @@ int CeedElemRestrictionCreateBlocked_Cuda(const CeedMemType mtype,
   int ierr;
   Ceed ceed;
   ierr = CeedElemRestrictionGetCeed(r, &ceed); CeedChk(ierr);
-  return CeedError(ceed, 1, "Backend does not implement blocked restrictions");
+  return CeedError(ceed, CEED_ERROR_BACKEND,
+                   "Backend does not implement blocked restrictions");
 }
 //------------------------------------------------------------------------------
