+++ backends/ref/ceed-ref.c
@@ -27,7 +27,8 @@ static int CeedInit_Ref(const char *resource, Ceed ceed) {
   if (strcmp(resource, "/cpu/self") && strcmp(resource, "/cpu/self/ref")
       && strcmp(resource, "/cpu/self/ref/serial"))
     // LCOV_EXCL_START
-    return CeedError(ceed, 1, "Ref backend cannot use resource: %s", resource);
+    return CeedError(ceed, CEED_ERROR_BACKEND,
+                     "Ref backend cannot use resource: %s", resource);
   // LCOV_EXCL_STOP
   ierr = CeedSetDeterministic(ceed, true); CeedChk(ierr);
 

[…]
