### 2021-01-26 bfdd5b8
tests/tap.sh: Print out test names for fluids examples
inspired-by: @jeremylt

tests/tap.sh: drop quotation marks from the printed test names
