+++ tests/tap.sh
@@ -20,7 +20,7 @@ fi
 
 # for examples/ceed petsc*, mfem*, or ex* grep the code to fetch arguments from a TESTARGS line
 declare -a allargs
-declare -a suffices
+declare -a names
 if [ ${1::6} == "petsc-" ]; then
     allargs=$(grep -F //TESTARGS examples/petsc/${1:6}.c* | cut -d\  -f2- )
 elif [ ${1::5} == "mfem-" ]; then
@@ -32,9 +32,11 @@ elif [ ${1::4} == "nek-" ]; then
       allargs+=("$(awk -v i="$i" '/C_TESTARGS/,/\n/{j++}j==i+1{print; exit}' examples/nek/bps/${1:4}.usr* | cut -d\  -f2- )")
     done
 elif [ ${1::7} == "fluids-" ]; then
-    # get all test configurations
     numconfig=$(grep -F //TESTARGS examples/fluids/${1:7}.c* | wc -l)
     for ((i=0;i<${numconfig};++i)); do
+      # get test name
+      names+=("$(awk -v i="$i" '/\/\/TESTARGS/,/\n/{j++}j==i+1{print substr($1,18,length($1)-19)}' examples/fluids/${1:7}.c)")
+      # get all test configurations
       allargs+=("$(awk -v i="$i" '/\/\/TESTARGS/,/\n/{j++}j==i+1{print; exit}' examples/fluids/${1:7}.c | cut -d\  -f2- )")
     done
 elif [ ${1::7} == "solids-" ]; then
@@ -57,6 +59,7 @@ trap 'rm -f ${tmpfiles}' EXIT
 # test configurations loop
 for ((j=0;j<${#allargs[@]};++j)); do
 args=${allargs[$j]}
+printf "# Test Name: ${names[$j]}\n"
 printf "# TESTARGS: $args\n"
 
 # backends loop
