+++ interface/ceed-qfunction.c
@@ -76,7 +76,7 @@ int CeedQFunctionRegister(const char *name, const char *source,
                           int (*init)(Ceed, const char *, CeedQFunction)) {
   if (num_qfunctions >= sizeof(qfunctions) / sizeof(qfunctions[0]))
     // LCOV_EXCL_START
-    return CeedError(NULL, 1, "Too many gallery QFunctions");
+    return CeedError(NULL, CEED_ERROR_UNSUPPORTED, "Too many gallery QFunctions");
   // LCOV_EXCL_STOP
 
   strncpy(qfunctions[num_qfunctions].name, name, CEED_MAX_RESOURCE_LEN);
@@ -87,7 +87,7 @@ int CeedQFunctionRegister(const char *name, const char *source,
   qfunctions[num_qfunctions].f = f;
   qfunctions[num_qfunctions].init = init;
   num_qfunctions++;
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 /**
@@ -118,7 +118,7 @@ static int CeedQFunctionFieldSet(CeedQFunctionField *f,const char *fieldname,
   (*f)->fieldname = tmp;
   (*f)->size = size;
   (*f)->emode = emode;
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 /**
@@ -142,8 +142,7 @@ static int CeedQFunctionFieldView(CeedQFunctionField field, CeedInt fieldnumber,
           "      EvalMode: \"%s\"\n",
           inout, fieldnumber, field->fieldname, field->size,
           CeedEvalModes[field->emode]);
-
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 
@@ -159,7 +158,7 @@ static int CeedQFunctionFieldView(CeedQFunctionField field, CeedInt fieldnumber,
 **/
 int CeedQFunctionSetFortranStatus(CeedQFunction qf, bool status) {
   qf->fortranstatus = status;
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 /// @}
@@ -182,7 +181,7 @@ int CeedQFunctionSetFortranStatus(CeedQFunction qf, bool status) {
 **/
 int CeedQFunctionGetCeed(CeedQFunction qf, Ceed *ceed) {
   *ceed = qf->ceed;
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 /**
@@ -197,7 +196,7 @@ int CeedQFunctionGetCeed(CeedQFunction qf, Ceed *ceed) {
 **/
 int CeedQFunctionGetVectorLength(CeedQFunction qf, CeedInt *vlength) {
   *vlength = qf->vlength;
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 /**
@@ -215,7 +214,7 @@ int CeedQFunctionGetNumArgs(CeedQFunction qf, CeedInt *numinput,
                             CeedInt *numoutput) {
   if (numinput) *numinput = qf->numinputfields;
   if (numoutput) *numoutput = qf->numoutputfields;
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 /**
@@ -230,7 +229,7 @@ int CeedQFunctionGetNumArgs(CeedQFunction qf, CeedInt *numinput,
 **/
 int CeedQFunctionGetSourcePath(CeedQFunction qf, char **source) {
   *source = (char *) qf->sourcepath;
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 /**
@@ -245,7 +244,7 @@ int CeedQFunctionGetSourcePath(CeedQFunction qf, char **source) {
 **/
[…]
