+++ interface/ceed-elemrestriction.c
@@ -56,7 +56,7 @@ int CeedPermutePadOffsets(const CeedInt *offsets, CeedInt *blkoffsets,
       for (int k = 0; k < elemsize; k++)
         blkoffsets[e*elemsize + k*blksize + j]
           = offsets[CeedIntMin(e+j,nelem-1)*elemsize + k];
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 /// @}
@@ -79,7 +79,7 @@ int CeedPermutePadOffsets(const CeedInt *offsets, CeedInt *blkoffsets,
 **/
 int CeedElemRestrictionGetCeed(CeedElemRestriction rstr, Ceed *ceed) {
   *ceed = rstr->ceed;
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 /**
@@ -97,12 +97,13 @@ int CeedElemRestrictionGetStrides(CeedElemRestriction rstr,
                                   CeedInt (*strides)[3]) {
   if (!rstr->strides)
     // LCOV_EXCL_START
-    return CeedError(rstr->ceed, 1, "ElemRestriction has no stride data");
+    return CeedError(rstr->ceed, CEED_ERROR_UNSUPPORTED,
+                     "ElemRestriction has no stride data");
   // LCOV_EXCL_STOP
 
   for (int i = 0; i<3; i++)
     (*strides)[i] = rstr->strides[i];
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 /**
@@ -124,12 +125,13 @@ int CeedElemRestrictionGetOffsets(CeedElemRestriction rstr, CeedMemType mtype,
 
   if (!rstr->GetOffsets)
     // LCOV_EXCL_START
-    return CeedError(rstr->ceed, 1, "Backend does not support GetOffsets");
+    return CeedError(rstr->ceed, CEED_ERROR_UNSUPPORTED,
+                     "Backend does not support GetOffsets");
   // LCOV_EXCL_STOP
 
   ierr = rstr->GetOffsets(rstr, mtype, offsets); CeedChk(ierr);
   rstr->numreaders++;
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 /**
@@ -146,7 +148,7 @@ int CeedElemRestrictionRestoreOffsets(CeedElemRestriction rstr,
                                       const CeedInt **offsets) {
   *offsets = NULL;
   rstr->numreaders--;
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 /**
@@ -161,7 +163,7 @@ int CeedElemRestrictionRestoreOffsets(CeedElemRestriction rstr,
 **/
 int CeedElemRestrictionIsStrided(CeedElemRestriction rstr, bool *isstrided) {
   *isstrided = rstr->strides ? 1 : 0;
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 /**
@@ -178,13 +180,14 @@ int CeedElemRestrictionHasBackendStrides(CeedElemRestriction rstr,
     bool *hasbackendstrides) {
   if (!rstr->strides)
     // LCOV_EXCL_START
-    return CeedError(rstr->ceed, 1, "ElemRestriction has no stride data");
+    return CeedError(rstr->ceed, CEED_ERROR_UNSUPPORTED,
+                     "ElemRestriction has no stride data");
   // LCOV_EXCL_STOP
 
   *hasbackendstrides = ((rstr->strides[0] == CEED_STRIDES_BACKEND[0]) &&
                         (rstr->strides[1] == CEED_STRIDES_BACKEND[1]) &&
                         (rstr->strides[2] == CEED_STRIDES_BACKEND[2]));
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 /**
[…]
