+++ interface/ceed-qfunction-register.c
@@ -27,5 +27,5 @@ int CeedQFunctionRegisterAll() {
 #define MACRO(name) CeedChk(name());
 #include "../gallery/ceed-gallery-list.h"
 #undef MACRO
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
