+++ interface/ceed.c
@@ -108,8 +108,9 @@ CeedRequest *const CEED_REQUEST_ORDERED = &ceed_request_ordered;
 **/
 int CeedRequestWait(CeedRequest *req) {
   if (!*req)
-    return 0;
-  return CeedError(NULL, 2, "CeedRequestWait not implemented");
+    return CEED_ERROR_SUCCESS;
+  return CeedError(NULL, CEED_ERROR_UNSUPPORTED,
+                   "CeedRequestWait not implemented");
 }
 
 /// @}
@@ -195,11 +196,11 @@ int CeedMallocArray(size_t n, size_t unit, void *p) {
   int ierr = posix_memalign((void **)p, CEED_ALIGN, n*unit);
   if (ierr)
     // LCOV_EXCL_START
-    return CeedError(NULL, ierr, "posix_memalign failed to allocate %zd "
+    return CeedError(NULL, CEED_ERROR_TERMINAL,
+                     "posix_memalign failed to allocate %zd "
                      "members of size %zd\n", n, unit);
   // LCOV_EXCL_STOP
-
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 /**
@@ -221,11 +222,11 @@ int CeedCallocArray(size_t n, size_t unit, void *p) {
   *(void **)p = calloc(n, unit);
   if (n && unit && !*(void **)p)
     // LCOV_EXCL_START
-    return CeedError(NULL, 1, "calloc failed to allocate %zd members of size "
+    return CeedError(NULL, CEED_ERROR_TERMINAL,
+                     "calloc failed to allocate %zd members of size "
                      "%zd\n", n, unit);
   // LCOV_EXCL_STOP
-
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 /**
@@ -247,11 +248,11 @@ int CeedReallocArray(size_t n, size_t unit, void *p) {
   *(void **)p = realloc(*(void **)p, n*unit);
   if (n && unit && !*(void **)p)
     // LCOV_EXCL_START
-    return CeedError(NULL, 1, "realloc failed to allocate %zd members of size "
+    return CeedError(NULL, CEED_ERROR_TERMINAL,
+                     "realloc failed to allocate %zd members of size "
                      "%zd\n", n, unit);
   // LCOV_EXCL_STOP
-
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 /** Free memory allocated using CeedMalloc() or CeedCalloc()
@@ -263,7 +264,7 @@ int CeedReallocArray(size_t n, size_t unit, void *p) {
 int CeedFree(void *p) {
   free(*(void **)p);
   *(void **)p = NULL;
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 /**
@@ -285,7 +286,7 @@ int CeedRegister(const char *prefix, int (*init)(const char *, Ceed),
                  unsigned int priority) {
   if (num_backends >= sizeof(backends) / sizeof(backends[0]))
     // LCOV_EXCL_START
-    return CeedError(NULL, 1, "Too many backends");
+    return CeedError(NULL, CEED_ERROR_TERMINAL, "Too many backends");
   // LCOV_EXCL_STOP
 
   strncpy(backends[num_backends].prefix, prefix, CEED_MAX_RESOURCE_LEN);
[…]
