+++ interface/ceed-qfunctioncontext.c
@@ -41,7 +41,7 @@
 **/
 int CeedQFunctionContextGetCeed(CeedQFunctionContext ctx, Ceed *ceed) {
   *ceed = ctx->ceed;
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 /**
@@ -56,7 +56,7 @@ int CeedQFunctionContextGetCeed(CeedQFunctionContext ctx, Ceed *ceed) {
 **/
 int CeedQFunctionContextGetState(CeedQFunctionContext ctx, uint64_t *state) {
   *state = ctx->state;
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 /**
@@ -72,7 +72,7 @@ int CeedQFunctionContextGetState(CeedQFunctionContext ctx, uint64_t *state) {
 int CeedQFunctionContextGetContextSize(CeedQFunctionContext ctx,
                                        size_t *ctxsize) {
   *ctxsize = ctx->ctxsize;
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 /**
@@ -87,7 +87,7 @@ int CeedQFunctionContextGetContextSize(CeedQFunctionContext ctx,
 **/
 int CeedQFunctionContextGetBackendData(CeedQFunctionContext ctx, void *data) {
   *(void **)data = ctx->data;
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 /**
@@ -102,7 +102,7 @@ int CeedQFunctionContextGetBackendData(CeedQFunctionContext ctx, void *data) {
 **/
 int CeedQFunctionContextSetBackendData(CeedQFunctionContext ctx, void *data) {
   ctx->data = data;
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 /// @}
@@ -133,11 +133,12 @@ int CeedQFunctionContextCreate(Ceed ceed, CeedQFunctionContext *ctx) {
 
     if (!delegate)
       // LCOV_EXCL_START
-      return CeedError(ceed, 1, "Backend does not support ContextCreate");
+      return CeedError(ceed, CEED_ERROR_UNSUPPORTED,
+                       "Backend does not support ContextCreate");
     // LCOV_EXCL_STOP
 
     ierr = CeedQFunctionContextCreate(delegate, ctx); CeedChk(ierr);
-    return 0;
+    return CEED_ERROR_SUCCESS;
   }
 
   ierr = CeedCalloc(1, ctx); CeedChk(ierr);
@@ -145,7 +146,7 @@ int CeedQFunctionContextCreate(Ceed ceed, CeedQFunctionContext *ctx) {
   ceed->refcount++;
   (*ctx)->refcount = 1;
   ierr = ceed->QFunctionContextCreate(*ctx); CeedChk(ierr);
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 /**
@@ -170,7 +171,8 @@ int CeedQFunctionContextSetData(CeedQFunctionContext ctx, CeedMemType mtype,
 
   if (!ctx->SetData)
     // LCOV_EXCL_START
-    return CeedError(ctx->ceed, 1, "Backend does not support ContextSetData");
+    return CeedError(ctx->ceed, CEED_ERROR_UNSUPPORTED,
+                     "Backend does not support ContextSetData");
   // LCOV_EXCL_STOP
 
   if (ctx->state % 2 == 1)
@@ -184,7 +186,7 @@ int CeedQFunctionContextSetData(CeedQFunctionContext ctx, CeedMemType mtype,
[…]
