### 2021-01-26 2226dbd
error - update backends to use CEED_ERROR_SUCCESS and CeedChkBackend
