+++ backends/cuda-gen/ceed-cuda-gen-qfunction.c
@@ -29,7 +29,7 @@ static int CeedQFunctionApply_Cuda_gen(CeedQFunction qf, CeedInt Q,
                                        CeedVector *U, CeedVector *V) {
   int ierr;
   Ceed ceed;
-  ierr = CeedQFunctionGetCeed(qf, &ceed); CeedChk(ierr);
+  ierr = CeedQFunctionGetCeed(qf, &ceed); CeedChkBackend(ierr);
   return CeedError(ceed, CEED_ERROR_BACKEND,
                    "Backend does not implement QFunctionApply");
 }
@@ -40,13 +40,13 @@ static int CeedQFunctionApply_Cuda_gen(CeedQFunction qf, CeedInt Q,
 static int CeedQFunctionDestroy_Cuda_gen(CeedQFunction qf) {
   int ierr;
   CeedQFunction_Cuda_gen *data;
-  ierr = CeedQFunctionGetData(qf, &data); CeedChk(ierr);
+  ierr = CeedQFunctionGetData(qf, &data); CeedChkBackend(ierr);
   Ceed ceed;
-  ierr = CeedQFunctionGetCeed(qf, &ceed); CeedChk(ierr);
+  ierr = CeedQFunctionGetCeed(qf, &ceed); CeedChkBackend(ierr);
   ierr = cudaFree(data->d_c); CeedChk_Cu(ceed, ierr);
-  ierr = CeedFree(&data->qFunctionSource); CeedChk(ierr);
-  ierr = CeedFree(&data); CeedChk(ierr);
-  return 0;
+  ierr = CeedFree(&data->qFunctionSource); CeedChkBackend(ierr);
+  ierr = CeedFree(&data); CeedChkBackend(ierr);
+  return CEED_ERROR_SUCCESS;
 }
 
 //------------------------------------------------------------------------------
@@ -57,11 +57,11 @@ static int loadCudaFunction(CeedQFunction qf, char *c_src_file) {
   Ceed ceed;
   CeedQFunctionGetCeed(qf, &ceed);
   CeedQFunction_Cuda_gen *data;
-  ierr = CeedQFunctionGetData(qf, &data); CeedChk(ierr);
+  ierr = CeedQFunctionGetData(qf, &data); CeedChkBackend(ierr);
 
   // Find source file
   char *cuda_file;
-  ierr = CeedCalloc(CUDA_MAX_PATH, &cuda_file); CeedChk(ierr);
+  ierr = CeedCalloc(CUDA_MAX_PATH, &cuda_file); CeedChkBackend(ierr);
   memcpy(cuda_file, c_src_file, strlen(c_src_file));
   const char *last_dot = strrchr(cuda_file, '.');
   if (!last_dot)
@@ -86,13 +86,13 @@ static int loadCudaFunction(CeedQFunction qf, char *c_src_file) {
   rewind(fp);
 
   // Allocate memory for entire content
-  ierr = CeedCalloc(lSize+1, &buffer); CeedChk(ierr);
+  ierr = CeedCalloc(lSize+1, &buffer); CeedChkBackend(ierr);
 
   // Copy the file into the buffer
   if (1 != fread(buffer, lSize, 1, fp)) {
     // LCOV_EXCL_START
     fclose(fp);
-    ierr = CeedFree(&buffer); CeedChk(ierr);
+    ierr = CeedFree(&buffer); CeedChkBackend(ierr);
     CeedError(ceed, CEED_ERROR_BACKEND,
               "Couldn't read the Cuda file for the QFunction.");
     // LCOV_EXCL_STOP
@@ -103,16 +103,16 @@ static int loadCudaFunction(CeedQFunction qf, char *c_src_file) {
   char *fields_string =
[…]
