+++ backends/cuda/ceed-cuda-basis.c
@@ -345,11 +345,11 @@ int CeedBasisApply_Cuda(CeedBasis basis, const CeedInt nelem,
                         CeedEvalMode emode, CeedVector u, CeedVector v) {
   int ierr;
   Ceed ceed;
-  ierr = CeedBasisGetCeed(basis, &ceed); CeedChk(ierr);
+  ierr = CeedBasisGetCeed(basis, &ceed); CeedChkBackend(ierr);
   Ceed_Cuda *ceed_Cuda;
-  ierr = CeedGetData(ceed, &ceed_Cuda); CeedChk(ierr);
+  ierr = CeedGetData(ceed, &ceed_Cuda); CeedChkBackend(ierr);
   CeedBasis_Cuda *data;
-  ierr = CeedBasisGetData(basis, &data); CeedChk(ierr);
+  ierr = CeedBasisGetData(basis, &data); CeedChkBackend(ierr);
   const CeedInt transpose = tmode == CEED_TRANSPOSE;
   const int maxblocksize = 32;
 
@@ -357,14 +357,14 @@ int CeedBasisApply_Cuda(CeedBasis basis, const CeedInt nelem,
   const CeedScalar *d_u;
   CeedScalar *d_v;
   if (emode != CEED_EVAL_WEIGHT) {
-    ierr = CeedVectorGetArrayRead(u, CEED_MEM_DEVICE, &d_u); CeedChk(ierr);
+    ierr = CeedVectorGetArrayRead(u, CEED_MEM_DEVICE, &d_u); CeedChkBackend(ierr);
   }
-  ierr = CeedVectorGetArray(v, CEED_MEM_DEVICE, &d_v); CeedChk(ierr);
+  ierr = CeedVectorGetArray(v, CEED_MEM_DEVICE, &d_v); CeedChkBackend(ierr);
 
   // Clear v for transpose operation
   if (tmode == CEED_TRANSPOSE) {
     CeedInt length;
-    ierr = CeedVectorGetLength(v, &length); CeedChk(ierr);
+    ierr = CeedVectorGetLength(v, &length); CeedChkBackend(ierr);
     ierr = cudaMemset(d_v, 0, length * sizeof(CeedScalar));
     CeedChk_Cu(ceed,ierr);
   }
@@ -376,13 +376,13 @@ int CeedBasisApply_Cuda(CeedBasis basis, const CeedInt nelem,
                           &data->d_interp1d, &d_u, &d_v
                          };
     CeedInt Q1d, dim;
-    ierr = CeedBasisGetNumQuadraturePoints1D(basis, &Q1d); CeedChk(ierr);
-    ierr = CeedBasisGetDimension(basis, &dim); CeedChk(ierr);
+    ierr = CeedBasisGetNumQuadraturePoints1D(basis, &Q1d); CeedChkBackend(ierr);
+    ierr = CeedBasisGetDimension(basis, &dim); CeedChkBackend(ierr);
     CeedInt blocksize = CeedIntPow(Q1d, dim);
     blocksize = blocksize > maxblocksize ? maxblocksize : blocksize;
 
     ierr = CeedRunKernelCuda(ceed, data->interp, nelem, blocksize, interpargs);
-    CeedChk(ierr);
+    CeedChkBackend(ierr);
   } break;
   case CEED_EVAL_GRAD: {
     void *gradargs[] = {(void *) &nelem, (void *) &transpose, &data->d_interp1d,
@@ -391,7 +391,7 @@ int CeedBasisApply_Cuda(CeedBasis basis, const CeedInt nelem,
     CeedInt blocksize = maxblocksize;
 
[…]
