### 2021-01-26 c7dcf5e
error - update backends to use CEED_ERROR_SUCCESS and CeedChkBackend
