+++ backends/memcheck/ceed-memcheck-serial.c
@@ -36,12 +36,12 @@ static int CeedInit_Memcheck(const char *resource, Ceed ceed) {
   //   through unless overridden
   Ceed ceedref;
   CeedInit("/cpu/self/ref/serial", &ceedref);
-  ierr = CeedSetDelegate(ceed, ceedref); CeedChk(ierr);
+  ierr = CeedSetDelegate(ceed, ceedref); CeedChkBackend(ierr);
 
   ierr = CeedSetBackendFunction(ceed, "Ceed", ceed, "QFunctionCreate",
-                                CeedQFunctionCreate_Memcheck); CeedChk(ierr);
+                                CeedQFunctionCreate_Memcheck); CeedChkBackend(ierr);
 
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 //------------------------------------------------------------------------------
