+++ backends/ref/ceed-ref-qfunction.c
@@ -26,44 +26,44 @@ static int CeedQFunctionApply_Ref(CeedQFunction qf, CeedInt Q,
                                   CeedVector *U, CeedVector *V) {
   int ierr;
   CeedQFunction_Ref *impl;
-  ierr = CeedQFunctionGetData(qf, &impl); CeedChk(ierr);
+  ierr = CeedQFunctionGetData(qf, &impl); CeedChkBackend(ierr);
 
   CeedQFunctionContext ctx;
-  ierr = CeedQFunctionGetContext(qf, &ctx); CeedChk(ierr);
+  ierr = CeedQFunctionGetContext(qf, &ctx); CeedChkBackend(ierr);
   void *ctxData = NULL;
   if (ctx) {
     ierr = CeedQFunctionContextGetData(ctx, CEED_MEM_HOST, &ctxData);
-    CeedChk(ierr);
+    CeedChkBackend(ierr);
   }
 
   CeedQFunctionUser f = NULL;
-  ierr = CeedQFunctionGetUserFunction(qf, &f); CeedChk(ierr);
+  ierr = CeedQFunctionGetUserFunction(qf, &f); CeedChkBackend(ierr);
 
   CeedInt nIn, nOut;
-  ierr = CeedQFunctionGetNumArgs(qf, &nIn, &nOut); CeedChk(ierr);
+  ierr = CeedQFunctionGetNumArgs(qf, &nIn, &nOut); CeedChkBackend(ierr);
 
   for (int i = 0; i<nIn; i++) {
     ierr = CeedVectorGetArrayRead(U[i], CEED_MEM_HOST, &impl->inputs[i]);
-    CeedChk(ierr);
+    CeedChkBackend(ierr);
   }
   for (int i = 0; i<nOut; i++) {
     ierr = CeedVectorGetArray(V[i], CEED_MEM_HOST, &impl->outputs[i]);
-    CeedChk(ierr);
+    CeedChkBackend(ierr);
   }
 
-  ierr = f(ctxData, Q, impl->inputs, impl->outputs); CeedChk(ierr);
+  ierr = f(ctxData, Q, impl->inputs, impl->outputs); CeedChkBackend(ierr);
 
   for (int i = 0; i<nIn; i++) {
-    ierr = CeedVectorRestoreArrayRead(U[i], &impl->inputs[i]); CeedChk(ierr);
+    ierr = CeedVectorRestoreArrayRead(U[i], &impl->inputs[i]); CeedChkBackend(ierr);
   }
   for (int i = 0; i<nOut; i++) {
-    ierr = CeedVectorRestoreArray(V[i], &impl->outputs[i]); CeedChk(ierr);
+    ierr = CeedVectorRestoreArray(V[i], &impl->outputs[i]); CeedChkBackend(ierr);
   }
   if (ctx) {
-    ierr = CeedQFunctionContextRestoreData(ctx, &ctxData); CeedChk(ierr);
+    ierr = CeedQFunctionContextRestoreData(ctx, &ctxData); CeedChkBackend(ierr);
   }
 
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 //------------------------------------------------------------------------------
@@ -72,13 +72,13 @@ static int CeedQFunctionApply_Ref(CeedQFunction qf, CeedInt Q,
 static int CeedQFunctionDestroy_Ref(CeedQFunction qf) {
   int ierr;
   CeedQFunction_Ref *impl;
-  ierr = CeedQFunctionGetData(qf, &impl); CeedChk(ierr);
+  ierr = CeedQFunctionGetData(qf, &impl); CeedChkBackend(ierr);
 
-  ierr = CeedFree(&impl->inputs); CeedChk(ierr);
-  ierr = CeedFree(&impl->outputs); CeedChk(ierr);
[…]
