### 2021-01-26 e661aee
error - use new errors in gallery
