+++ gallery/poisson2d/ceed-poisson2dbuild.c
@@ -31,7 +31,8 @@ static int CeedQFunctionInit_Poisson2DBuild(Ceed ceed, const char *requested,
   const char *name = "Poisson2DBuild";
   if (strcmp(name, requested))
     // LCOV_EXCL_START
-    return CeedError(ceed, 1, "QFunction '%s' does not match requested name: %s",
+    return CeedError(ceed, CEED_ERROR_UNSUPPORTED,
+                     "QFunction '%s' does not match requested name: %s",
                      name, requested);
   // LCOV_EXCL_STOP
 
@@ -44,7 +45,7 @@ static int CeedQFunctionInit_Poisson2DBuild(Ceed ceed, const char *requested,
   ierr = CeedQFunctionAddOutput(qf, "qdata", dim*(dim+1)/2, CEED_EVAL_NONE);
   CeedChk(ierr);
 
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 /**
