+++ gallery/mass1d/ceed-mass1dbuild.c
@@ -31,7 +31,8 @@ static int CeedQFunctionInit_Mass1DBuild(Ceed ceed, const char *requested,
   const char *name = "Mass1DBuild";
   if (strcmp(name, requested))
     // LCOV_EXCL_START
-    return CeedError(ceed, 1, "QFunction '%s' does not match requested name: %s",
+    return CeedError(ceed, CEED_ERROR_UNSUPPORTED,
+                     "QFunction '%s' does not match requested name: %s",
                      name, requested);
   // LCOV_EXCL_STOP
 
@@ -43,7 +44,7 @@ static int CeedQFunctionInit_Mass1DBuild(Ceed ceed, const char *requested,
   CeedChk(ierr);
   ierr = CeedQFunctionAddOutput(qf, "qdata", 1, CEED_EVAL_NONE); CeedChk(ierr);
 
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 /**
