+++ backends/blocked/ceed-blocked-operator.c
@@ -127,7 +127,7 @@ static int CeedOperatorSetupFields_Blocked(CeedQFunction qf,
       break; // Not implemented
     }
   }
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 //------------------------------------------------------------------------------
@@ -137,7 +137,7 @@ static int CeedOperatorSetup_Blocked(CeedOperator op) {
   int ierr;
   bool setupdone;
   ierr = CeedOperatorIsSetupDone(op, &setupdone); CeedChkBackend(ierr);
-  if (setupdone) return 0;
+  if (setupdone) return CEED_ERROR_SUCCESS;
   Ceed ceed;
   ierr = CeedOperatorGetCeed(op, &ceed); CeedChkBackend(ierr);
   CeedOperator_Blocked *impl;
@@ -206,7 +206,7 @@ static int CeedOperatorSetup_Blocked(CeedOperator op) {
 
   ierr = CeedOperatorSetSetupDone(op); CeedChkBackend(ierr);
 
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 //------------------------------------------------------------------------------
@@ -249,7 +249,7 @@ static inline int CeedOperatorSetupInputs_Blocked(CeedInt numinputfields,
       CeedChkBackend(ierr);
     }
   }
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 //------------------------------------------------------------------------------
@@ -327,7 +327,7 @@ static inline int CeedOperatorInputBasis_Blocked(CeedInt e, CeedInt Q,
     }
     }
   }
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 //------------------------------------------------------------------------------
@@ -398,7 +398,7 @@ static inline int CeedOperatorOutputBasis_Blocked(CeedInt e, CeedInt Q,
     }
     }
   }
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 //------------------------------------------------------------------------------
@@ -427,7 +427,7 @@ static inline int CeedOperatorRestoreInputs_Blocked(CeedInt numinputfields,
       CeedChkBackend(ierr);
     }
   }
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 //------------------------------------------------------------------------------
@@ -527,7 +527,7 @@ static int CeedOperatorApplyAdd_Blocked(CeedOperator op, CeedVector invec,
   ierr = CeedOperatorRestoreInputs_Blocked(numinputfields, qfinputfields,
          opinputfields, false, impl); CeedChkBackend(ierr);
 
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 //------------------------------------------------------------------------------
@@ -703,7 +703,7 @@ static int CeedOperatorLinearAssembleQFunction_Blocked(CeedOperator op,
   ierr = CeedVectorDestroy(&lvec); CeedChkBackend(ierr);
   ierr = CeedElemRestrictionDestroy(&blkrstr); CeedChkBackend(ierr);
 
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 //------------------------------------------------------------------------------
@@ -738,7 +738,7 @@ static int CeedOperatorDestroy_Blocked(CeedOperator op) {
   ierr = CeedFree(&impl->qvecsout); CeedChkBackend(ierr);
 
   ierr = CeedFree(&impl); CeedChkBackend(ierr);
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 //------------------------------------------------------------------------------
@@ -760,6 +760,6 @@ int CeedOperatorCreate_Blocked(CeedOperator op) {
[…]
