+++ backends/cuda/ceed-cuda-restriction.c
@@ -190,7 +190,7 @@ static int CeedElemRestrictionApply_Cuda(CeedElemRestriction r,
   // Restore arrays
   ierr = CeedVectorRestoreArrayRead(u, &d_u); CeedChkBackend(ierr);
   ierr = CeedVectorRestoreArray(v, &d_v); CeedChkBackend(ierr);
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 //------------------------------------------------------------------------------
@@ -225,7 +225,7 @@ static int CeedElemRestrictionGetOffsets_Cuda(CeedElemRestriction rstr,
     *offsets = impl->d_ind;
     break;
   }
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 //------------------------------------------------------------------------------
@@ -246,7 +246,7 @@ static int CeedElemRestrictionDestroy_Cuda(CeedElemRestriction r) {
   ierr = cudaFree(impl->d_lvec_indices); CeedChk_Cu(ceed, ierr);
 
   ierr = CeedFree(&impl); CeedChkBackend(ierr);
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 //------------------------------------------------------------------------------
@@ -337,7 +337,7 @@ static int CeedElemRestrictionOffset_Cuda(const CeedElemRestriction r,
   ierr = CeedFree(&lvec_indices); CeedChkBackend(ierr);
   ierr = CeedFree(&toffsets); CeedChkBackend(ierr);
   ierr = CeedFree(&tindices); CeedChkBackend(ierr);
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 //------------------------------------------------------------------------------
@@ -468,7 +468,7 @@ int CeedElemRestrictionCreate_Cuda(CeedMemType mtype, CeedCopyMode cmode,
   ierr = CeedSetBackendFunction(ceed, "ElemRestriction", r, "Destroy",
                                 CeedElemRestrictionDestroy_Cuda);
   CeedChkBackend(ierr);
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 //------------------------------------------------------------------------------
