+++ backends/hip-gen/ceed-hip-gen-operator-build.cpp
@@ -774,7 +774,8 @@ CEED_INTERN int CeedHipGenOperatorBuild(CeedOperator op) {
   oper = "CeedKernel_Hip_gen_" + qFunctionName;
 
   code << "\n#define CEED_QFUNCTION(name) inline __device__ int name\n";
-  code << "\n#define CeedPragmaSIMD\n";
+  code << "#define CeedPragmaSIMD\n";
+  code << "#define CEED_ERROR_SUCCESS 0\n\n";
 
   // Find dim and Q1d
   bool useCollograd = true;
