+++ backends/cuda-gen/ceed-cuda-gen-operator-build.cpp
@@ -802,7 +802,8 @@ extern "C" int CeedCudaGenOperatorBuild(CeedOperator op) {
   oper = "CeedKernel_Cuda_gen_" + qFunctionName;
 
   code << "\n#define CEED_QFUNCTION(name) inline __device__ int name\n";
-  code << "\n#define CeedPragmaSIMD\n";
+  code << "#define CeedPragmaSIMD\n";
+  code << "#define CEED_ERROR_SUCCESS 0\n\n";
 
   // Find dim and Q1d
   bool useCollograd = true;
