+++ backends/cuda/ceed-cuda-qfunction-load.cpp
@@ -77,8 +77,9 @@ extern "C" int CeedCudaBuildQFunction(CeedQFunction qf) {
 
   // Defintions
   code << "\n#define CEED_QFUNCTION(name) inline __device__ int name\n";
-  code << "\n#define CeedPragmaSIMD\n";
-  code << "\n#define CEED_Q_VLA 1\n\n";
+  code << "#define CeedPragmaSIMD\n";
+  code << "#define CEED_ERROR_SUCCESS 0\n";
+  code << "#define CEED_Q_VLA 1\n\n";
   code << "typedef struct { const CeedScalar* inputs[16]; CeedScalar* outputs[16]; } Fields_Cuda;\n";
   code << qReadWriteS;
   code << qFunction;
