+++ interface/ceed-operator.c
@@ -256,7 +256,6 @@ static int CeedOperatorFieldView(CeedOperatorField field,
     fprintf(stream, "%s      Active vector\n", pre);
   else if (field->vec == CEED_VECTOR_NONE)
     fprintf(stream, "%s      No vector\n", pre);
-
   return CEED_ERROR_SUCCESS;
 }
 
@@ -320,7 +319,7 @@ static int CeedOperatorGetActiveBasis(CeedOperator op,
     int ierr;
     Ceed ceed;
     ierr = CeedOperatorGetCeed(op, &ceed); CeedChk(ierr);
-    return CeedError(ceed, CEED_ERROR_UNSUPPORTED,
+    return CeedError(ceed, CEED_ERROR_NONTERMINAL,
                      "No active basis found for automatic multigrid setup");
     // LCOV_EXCL_STOP
   }
@@ -521,7 +520,7 @@ int CeedOperatorGetCeed(CeedOperator op, Ceed *ceed) {
 int CeedOperatorGetNumElements(CeedOperator op, CeedInt *numelem) {
   if (op->composite)
     // LCOV_EXCL_START
-    return CeedError(op->ceed, CEED_ERROR_UNSUPPORTED,
+    return CeedError(op->ceed, CEED_ERROR_NONTERMINAL,
                      "Not defined for composite operator");
   // LCOV_EXCL_STOP
 
@@ -543,7 +542,7 @@ int CeedOperatorGetNumElements(CeedOperator op, CeedInt *numelem) {
 int CeedOperatorGetNumQuadraturePoints(CeedOperator op, CeedInt *numqpts) {
   if (op->composite)
     // LCOV_EXCL_START
-    return CeedError(op->ceed, CEED_ERROR_UNSUPPORTED,
+    return CeedError(op->ceed, CEED_ERROR_NONTERMINAL,
                      "Not defined for composite operator");
   // LCOV_EXCL_STOP
 
@@ -565,7 +564,7 @@ int CeedOperatorGetNumQuadraturePoints(CeedOperator op, CeedInt *numqpts) {
 int CeedOperatorGetNumArgs(CeedOperator op, CeedInt *numargs) {
   if (op->composite)
     // LCOV_EXCL_START
-    return CeedError(op->ceed, CEED_ERROR_UNSUPPORTED,
+    return CeedError(op->ceed, CEED_ERROR_NONTERMINAL,
                      "Not defined for composite operators");
   // LCOV_EXCL_STOP
 
@@ -603,7 +602,7 @@ int CeedOperatorIsSetupDone(CeedOperator op, bool *issetupdone) {
 int CeedOperatorGetQFunction(CeedOperator op, CeedQFunction *qf) {
   if (op->composite)
     // LCOV_EXCL_START
-    return CeedError(op->ceed, CEED_ERROR_UNSUPPORTED,
+    return CeedError(op->ceed, CEED_ERROR_NONTERMINAL,
                      "Not defined for composite operator");
   // LCOV_EXCL_STOP
 
@@ -641,7 +640,7 @@ int CeedOperatorIsComposite(CeedOperator op, bool *iscomposite) {
 int CeedOperatorGetNumSub(CeedOperator op, CeedInt *numsub) {
   if (!op->composite)
     // LCOV_EXCL_START
-    return CeedError(op->ceed, CEED_ERROR_UNSUPPORTED, "Not a composite operator");
+    return CeedError(op->ceed, CEED_ERROR_NONTERMINAL, "Not a composite operator");
   // LCOV_EXCL_STOP
 
   *numsub = op->numsub;
[…]
