+++ interface/ceed-basis.c
@@ -98,13 +98,15 @@ int CeedHouseholderApplyQ(CeedScalar *A, const CeedScalar *Q,
                           const CeedScalar *tau, CeedTransposeMode tmode,
                           CeedInt m, CeedInt n, CeedInt k,
                           CeedInt row, CeedInt col) {
+  int ierr;
   CeedScalar v[m];
   for (CeedInt ii=0; ii<k; ii++) {
     CeedInt i = tmode == CEED_TRANSPOSE ? ii : k-1-ii;
     for (CeedInt j=i+1; j<m; j++)
       v[j] = Q[j*k+i];
     // Apply Householder reflector (I - tau v v^T) collograd1d^T
-    CeedHouseholderReflect(&A[i*row], &v[i], tau[i], m-i, n, row, col);
+    ierr = CeedHouseholderReflect(&A[i*row], &v[i], tau[i], m-i, n, row, col);
+    CeedChk(ierr);
   }
   return CEED_ERROR_SUCCESS;
 }
@@ -208,6 +210,8 @@ int CeedBasisGetCollocatedGrad(CeedBasis basis, CeedScalar *collograd1d) {
   // QR Factorization, interp1d = Q R
   ierr = CeedBasisGetCeed(basis, &ceed); CeedChk(ierr);
   ierr = CeedQRFactorization(ceed, interp1d, tau, Q1d, P1d); CeedChk(ierr);
+  // Note: This function is for backend use, so all errors are terminal
+  //   and we do not need to clean up memory on failure.
 
   // Apply Rinv, collograd1d = grad1d Rinv
   for (i=0; i<Q1d; i++) { // Row i
@@ -223,8 +227,8 @@ int CeedBasisGetCollocatedGrad(CeedBasis basis, CeedScalar *collograd1d) {
   }
 
   // Apply Qtranspose, collograd = collograd Qtranspose
-  CeedHouseholderApplyQ(collograd1d, interp1d, tau, CEED_NOTRANSPOSE,
-                        Q1d, Q1d, P1d, 1, Q1d);
+  ierr = CeedHouseholderApplyQ(collograd1d, interp1d, tau, CEED_NOTRANSPOSE,
+                               Q1d, Q1d, P1d, 1, Q1d); CeedChk(ierr);
 
   ierr = CeedFree(&interp1d); CeedChk(ierr);
   ierr = CeedFree(&grad1d); CeedChk(ierr);
@@ -403,15 +407,6 @@ int CeedBasisCreateTensorH1(Ceed ceed, CeedInt dim, CeedInt ncomp, CeedInt P1d,
                             const CeedScalar *qweight1d, CeedBasis *basis) {
   int ierr;
 
-  if (dim<1)
-    // LCOV_EXCL_START
-    return CeedError(ceed, CEED_ERROR_DIMENSION,
-                     "Basis dimension must be a positive value");
-  // LCOV_EXCL_STOP
-  CeedElemTopology topo = dim == 1 ? CEED_LINE :
-                          dim == 2 ? CEED_QUAD :
-                          CEED_HEX;
-
   if (!ceed->BasisCreateTensorH1) {
     Ceed delegate;
     ierr = CeedGetObjectDelegate(ceed, &delegate, "Basis"); CeedChk(ierr);
@@ -427,7 +422,17 @@ int CeedBasisCreateTensorH1(Ceed ceed, CeedInt dim, CeedInt ncomp, CeedInt P1d,
                                    qweight1d, basis); CeedChk(ierr);
[…]
