+++ interface/ceed-qfunctioncontext.c
@@ -185,7 +185,6 @@ int CeedQFunctionContextSetData(CeedQFunctionContext ctx, CeedMemType mtype,
   ctx->ctxsize = size;
   ierr = ctx->SetData(ctx, mtype, cmode, data); CeedChk(ierr);
   ctx->state += 2;
-
   return CEED_ERROR_SUCCESS;
 }
 
@@ -225,7 +224,6 @@ int CeedQFunctionContextGetData(CeedQFunctionContext ctx, CeedMemType mtype,
 
   ierr = ctx->GetData(ctx, mtype, data); CeedChk(ierr);
   ctx->state += 1;
-
   return CEED_ERROR_SUCCESS;
 }
 
@@ -302,7 +300,6 @@ int CeedQFunctionContextDestroy(CeedQFunctionContext *ctx) {
   if ((*ctx)->Destroy) {
     ierr = (*ctx)->Destroy(*ctx); CeedChk(ierr);
   }
-
   ierr = CeedDestroy(&(*ctx)->ceed); CeedChk(ierr);
   ierr = CeedFree(ctx); CeedChk(ierr);
   return CEED_ERROR_SUCCESS;
