### 2021-01-26 0ba8c2e
operator - fix setting field qpts
