### 2021-01-26 29564a1
basis - add input/output dimension error checking
