### 2021-01-27 c24ac85
examples/fluids: Changed default P_inlet 1e3 -> 1
