### 2021-01-27 c62b51d
examples/fluids: Changed default P_inlet 1e3 -> 1
