### 2021-01-27 97c5e2b
doc: Fluids - improved the sample run to make the vortex and its movement more visible
