### 2021-01-27 ea50442
python - move basis utility methods to ceed object, no basis required or used
