+++ python/ceed.py
@@ -18,6 +18,7 @@ from _ceed_cffi import ffi, lib
 import sys
 import os
 import io
+import numpy as np
 import tempfile
 from abc import ABC
 from .ceed_vector import Vector
@@ -109,6 +110,179 @@ class Ceed():
         self._check_error(err_code)
 
         return memtype[0]
+
+    # --- Basis utility functions ---
+
+    # Gauss quadrature
+    def gauss_quadrature(self, q):
+        """Construct a Gauss-Legendre quadrature.
+
+           Args:
+             Q: number of quadrature points (integrates polynomials of
+                  degree 2*Q-1 exactly)
+
+           Returns:
+             (qref1d, qweight1d): array of length Q to hold the abscissa on [-1, 1]
+                                    and array of length Q to hold the weights"""
+
+        # Setup arguments
+        qref1d = np.empty(q, dtype="float64")
+        qweight1d = np.empty(q, dtype="float64")
+
+        qref1d_pointer = ffi.new("CeedScalar *")
+        qref1d_pointer = ffi.cast(
+            "CeedScalar *",
+            qref1d.__array_interface__['data'][0])
+
+        qweight1d_pointer = ffi.new("CeedScalar *")
+        qweight1d_pointer = ffi.cast(
+            "CeedScalar *",
+            qweight1d.__array_interface__['data'][0])
+
+        # libCEED call
+        err_code = lib.CeedGaussQuadrature(q, qref1d_pointer, qweight1d_pointer)
+        self._check_error(err_code)
+
+        return qref1d, qweight1d
+
+    # Lobatto quadrature
+    def lobatto_quadrature(self, q):
+        """Construct a Gauss-Legendre-Lobatto quadrature.
+
+           Args:
+             q: number of quadrature points (integrates polynomials of
+                  degree 2*Q-3 exactly)
+
+           Returns:
+             (qref1d, qweight1d): array of length Q to hold the abscissa on [-1, 1]
+                                    and array of length Q to hold the weights"""
+
+        # Setup arguments
+        qref1d = np.empty(q, dtype="float64")
+        qref1d_pointer = ffi.new("CeedScalar *")
+        qref1d_pointer = ffi.cast(
+            "CeedScalar *",
+            qref1d.__array_interface__['data'][0])
+
+        qweight1d = np.empty(q, dtype="float64")
+        qweight1d_pointer = ffi.new("CeedScalar *")
+        qweight1d_pointer = ffi.cast(
+            "CeedScalar *",
+            qweight1d.__array_interface__['data'][0])
+
+        # libCEED call
+        err_code = lib.CeedLobattoQuadrature(
+            q, qref1d_pointer, qweight1d_pointer)
+        self._check_error(err_code)
+
+        return qref1d, qweight1d
+
+    # QR factorization
+    def qr_factorization(self, mat, tau, m, n):
+        """Return QR Factorization of a matrix.
+
+           Args:
+             ceed: Ceed context currently in use
+             *mat: Numpy array holding the row-major matrix to be factorized in place
+             *tau: Numpy array to hold the vector of lengt m of scaling factors
+             m: number of rows
+             n: numbef of columns"""
+
+        # Setup arguments
+        mat_pointer = ffi.new("CeedScalar *")
+        mat_pointer = ffi.cast(
+            "CeedScalar *",
+            mat.__array_interface__['data'][0])
+
+        tau_pointer = ffi.new("CeedScalar *")
+        tau_pointer = ffi.cast(
+            "CeedScalar *",
+            tau.__array_interface__['data'][0])
+
+        # libCEED call
+        err_code = lib.CeedQRFactorization(
+            self._pointer[0], mat_pointer, tau_pointer, m, n)
+        self._check_error(err_code)
+
+        return mat, tau
+
[…]
