+++ python/ceed_basis.py
@@ -114,177 +114,6 @@ class Basis(ABC):
 
         return q_pointer[0]
 
-    # Gauss quadrature
-    @staticmethod
-    def gauss_quadrature(q):
-        """Construct a Gauss-Legendre quadrature.
-
-           Args:
-             Q: number of quadrature points (integrates polynomials of
-                  degree 2*Q-1 exactly)
-
-           Returns:
-             (qref1d, qweight1d): array of length Q to hold the abscissa on [-1, 1]
-                                    and array of length Q to hold the weights"""
-
-        # Setup arguments
-        qref1d = np.empty(q, dtype="float64")
-        qweight1d = np.empty(q, dtype="float64")
-
-        qref1d_pointer = ffi.new("CeedScalar *")
-        qref1d_pointer = ffi.cast(
-            "CeedScalar *",
-            qref1d.__array_interface__['data'][0])
-
-        qweight1d_pointer = ffi.new("CeedScalar *")
-        qweight1d_pointer = ffi.cast(
-            "CeedScalar *",
-            qweight1d.__array_interface__['data'][0])
-
-        # libCEED call
-        err_code = lib.CeedGaussQuadrature(q, qref1d_pointer, qweight1d_pointer)
-        self._ceed._check_error(err_code)
-
-        return qref1d, qweight1d
-
-    # Lobatto quadrature
-    @staticmethod
-    def lobatto_quadrature(q):
-        """Construct a Gauss-Legendre-Lobatto quadrature.
-
-           Args:
-             q: number of quadrature points (integrates polynomials of
-                  degree 2*Q-3 exactly)
-
-           Returns:
-             (qref1d, qweight1d): array of length Q to hold the abscissa on [-1, 1]
-                                    and array of length Q to hold the weights"""
-
-        # Setup arguments
-        qref1d = np.empty(q, dtype="float64")
-        qref1d_pointer = ffi.new("CeedScalar *")
-        qref1d_pointer = ffi.cast(
-            "CeedScalar *",
-            qref1d.__array_interface__['data'][0])
-
-        qweight1d = np.empty(q, dtype="float64")
-        qweight1d_pointer = ffi.new("CeedScalar *")
-        qweight1d_pointer = ffi.cast(
-            "CeedScalar *",
-            qweight1d.__array_interface__['data'][0])
-
-        # libCEED call
-        err_code = lib.CeedLobattoQuadrature(
-            q, qref1d_pointer, qweight1d_pointer)
-        self._ceed._check_error(err_code)
-
-        return qref1d, qweight1d
-
-    # QR factorization
-    @staticmethod
-    def qr_factorization(ceed, mat, tau, m, n):
-        """Return QR Factorization of a matrix.
-
-           Args:
-             ceed: Ceed context currently in use
-             *mat: Numpy array holding the row-major matrix to be factorized in place
-             *tau: Numpy array to hold the vector of lengt m of scaling factors
-             m: number of rows
-             n: numbef of columns"""
-
-        # Setup arguments
-        mat_pointer = ffi.new("CeedScalar *")
-        mat_pointer = ffi.cast(
-            "CeedScalar *",
-            mat.__array_interface__['data'][0])
-
-        tau_pointer = ffi.new("CeedScalar *")
-        tau_pointer = ffi.cast(
-            "CeedScalar *",
-            tau.__array_interface__['data'][0])
-
-        # libCEED call
-        lib.CeedQRFactorization(
-            ceed._pointer[0], mat_pointer, tau_pointer, m, n)
-
-        return mat, tau
-
-    # Symmetric Schur decomposition
-    @staticmethod
-    def symmetric_schur_decomposition(ceed, mat, n):
-        """Return symmetric Schur decomposition of a symmetric matrix
[…]
