### 2021-01-27 9ccc7e8
python - force exit with negative error code
