### 2021-01-27 1a76722
rust - initial work to add error handling logic
