### 2021-01-28 CEED/libCEED#700 @FreddieWitherden
CeedRequestWait is documented but not implemented
I am currently reviewing libCEED as part of https://github.com/openjournals/joss-reviews/issues/2945.

The method CeedRequestWait is documented and seems extremely interesting.  However, looking through the source it does not appear to be actually implemented.  It might be prudent to relegate this method and associated paraphernalia to a feature branch until it is ready for mainline.  If nothing else it may be that when implementing that that a blocker raises its head or a simpler approach shows itself.
