### 2021-01-28 6ca4bac
Temporary change for convergence test and debugging
remove printing the arguments
