+++ examples/fluids/conv_test.sh
@@ -22,11 +22,11 @@ declare -A run_flags
     run_flags[dm_plex_box_faces]=20,20,1
     run_flags[P_inlet]=1
     run_flags[T_inlet]=1
-    run_flags[euler_test]=none
-    run_flags[lx]=1e3
-    run_flags[ly]=1e3
+    run_flags[euler_test]=t4
+    run_flags[lx]=1000
+    run_flags[ly]=1000
     run_flags[lz]=1
-    run_flags[ts_max_time]=.02
+    run_flags[ts_max_time]=.2
     run_flags[ts_rk_type]=5bs
     run_flags[ts_rtol]=1e-10
     run_flags[ts_atol]=1e-10
@@ -34,10 +34,10 @@ declare -A run_flags
 declare -A test_flags
     test_flags[degree_start]=1
     test_flags[degree_stride]=1
-    test_flags[degree_end]=3
+    test_flags[degree_end]=4
     test_flags[res_start]=6
-    test_flags[res_stride]=2
-    test_flags[res_end]=10
+    test_flags[res_stride]=4
+    test_flags[res_end]=20
 
 echo ",mesh_res,degree,rel_error" > conv_test_result.csv
 i=0
@@ -57,4 +57,5 @@ for ((d=${test_flags[degree_start]}; d<=${test_flags[degree_end]}; d+=${test_fla
 done
 
 # Compare the output CSV file with the reference file
-diff conv_test_result.csv tests-output/fluids_navierstokes_etv.csv
+# diff conv_test_result.csv tests-output/fluids_navierstokes_etv.csv
+python3 conv_plot.py --conv_result_file conv_test_result.csv
