### 2021-01-28 5889717
rust - add ceed.resource method
