### 2021-01-28 89cbae8
rust - add results for methods that may fail
