+++ rust/src/vector.rs
@@ -29,12 +29,12 @@ use crate::prelude::*;
 // -----------------------------------------------------------------------------
 #[derive(Debug, Clone, Copy)]
 pub enum VectorOpt<'a> {
-    Some(&'a Vector),
+    Some(&'a Vector<'a>),
     Active,
     None,
 }
 /// Construct a VectorOpt reference from a Vector reference
-impl<'a> From<&'a Vector> for VectorOpt<'a> {
+impl<'a> From<&'a Vector<'_>> for VectorOpt<'a> {
     fn from(vec: &'a Vector) -> Self {
         debug_assert!(vec.ptr != unsafe { bind_ceed::CEED_VECTOR_NONE });
         debug_assert!(vec.ptr != unsafe { bind_ceed::CEED_VECTOR_ACTIVE });
@@ -56,10 +56,11 @@ impl<'a> VectorOpt<'a> {
 // CeedVector context wrapper
 // -----------------------------------------------------------------------------
 #[derive(Debug)]
-pub struct Vector {
+pub struct Vector<'a> {
+    ceed: &'a crate::Ceed,
     pub(crate) ptr: bind_ceed::CeedVector,
 }
-impl From<&'_ Vector> for bind_ceed::CeedVector {
+impl From<&'_ Vector<'_>> for bind_ceed::CeedVector {
     fn from(vec: &Vector) -> Self {
         vec.ptr
     }
@@ -68,7 +69,7 @@ impl From<&'_ Vector> for bind_ceed::CeedVector {
 // -----------------------------------------------------------------------------
 // Destructor
 // -----------------------------------------------------------------------------
-impl Drop for Vector {
+impl<'a> Drop for Vector<'a> {
     fn drop(&mut self) {
         let not_none_and_active = self.ptr != unsafe { bind_ceed::CEED_VECTOR_NONE }
             && self.ptr != unsafe { bind_ceed::CEED_VECTOR_ACTIVE };
@@ -82,13 +83,13 @@ impl Drop for Vector {
 // -----------------------------------------------------------------------------
 // Display
 // -----------------------------------------------------------------------------
-impl fmt::Display for Vector {
+impl<'a> fmt::Display for Vector<'a> {
     /// View a Vector
     ///
     /// ```
     /// # use libceed::prelude::*;
     /// # let ceed = libceed::Ceed::default_init();
-    /// let vec = libceed::vector::Vector::from_slice(&ceed, &[1., 2., 3.,]);
+    /// let vec = libceed::vector::Vector::from_slice(&ceed, &[1., 2., 3.,]).unwrap();
     /// assert_eq!(vec.to_string(), "CeedVector length 3
     ///     1.00000000
     ///     2.00000000
@@ -113,17 +114,21 @@ impl fmt::Display for Vector {
 // -----------------------------------------------------------------------------
 // Implementations
 // -----------------------------------------------------------------------------
-impl Vector {
+impl<'a> Vector<'a> {
     // Constructors
-    pub fn create(ceed: &crate::Ceed, n: usize) -> Self {
+    pub fn create(ceed: &'a crate::Ceed, n: usize) -> crate::Result<Self> {
         let n = i32::try_from(n).unwrap();
         let mut ptr = std::ptr::null_mut();
-        unsafe { bind_ceed::CeedVectorCreate(ceed.ptr, n, &mut ptr) };
-        Self { ptr: ptr }
[…]
