+++ rust/src/operator.rs
@@ -23,22 +23,23 @@ use crate::prelude::*;
 // -----------------------------------------------------------------------------
 // CeedOperator context wrapper
 // -----------------------------------------------------------------------------
-pub(crate) struct OperatorCore {
+pub(crate) struct OperatorCore<'a> {
+    ceed: &'a crate::Ceed,
     ptr: bind_ceed::CeedOperator,
 }
 
-pub struct Operator {
-    op_core: OperatorCore,
+pub struct Operator<'a> {
+    op_core: OperatorCore<'a>,
 }
 
-pub struct CompositeOperator {
-    op_core: OperatorCore,
+pub struct CompositeOperator<'a> {
+    op_core: OperatorCore<'a>,
 }
 
 // -----------------------------------------------------------------------------
 // Destructor
 // -----------------------------------------------------------------------------
-impl Drop for OperatorCore {
+impl<'a> Drop for OperatorCore<'a> {
     fn drop(&mut self) {
         unsafe {
             bind_ceed::CeedOperatorDestroy(&mut self.ptr);
@@ -49,7 +50,7 @@ impl Drop for OperatorCore {
 // -----------------------------------------------------------------------------
 // Display
 // -----------------------------------------------------------------------------
-impl fmt::Display for OperatorCore {
+impl<'a> fmt::Display for OperatorCore<'a> {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         let mut ptr = std::ptr::null_mut();
         let mut sizeloc = crate::MAX_BUFFER_LENGTH;
@@ -68,7 +69,7 @@ impl fmt::Display for OperatorCore {
 /// ```
 /// # use libceed::prelude::*;
 /// # let ceed = libceed::Ceed::default_init();
-/// let qf = ceed.q_function_interior_by_name("Mass1DBuild");
+/// let qf = ceed.q_function_interior_by_name("Mass1DBuild").unwrap();
 ///
 /// // Operator field arguments
 /// let ne = 3;
@@ -78,21 +79,21 @@ impl fmt::Display for OperatorCore {
 ///   ind[2 * i + 0] = i as i32;
 ///   ind[2 * i + 1] = (i + 1) as i32;
 /// }
-/// let r = ceed.elem_restriction(ne, 2, 1, 1, ne + 1, MemType::Host, &ind);
+/// let r = ceed.elem_restriction(ne, 2, 1, 1, ne + 1, MemType::Host, &ind).unwrap();
 /// let strides : [i32; 3] = [1, q as i32, q as i32];
-/// let rq = ceed.strided_elem_restriction(ne, 2, 1, q*ne, strides);
+/// let rq = ceed.strided_elem_restriction(ne, 2, 1, q*ne, strides).unwrap();
 ///
-/// let b = ceed.basis_tensor_H1_Lagrange(1, 1, 2, q, QuadMode::Gauss);
+/// let b = ceed.basis_tensor_H1_Lagrange(1, 1, 2, q, QuadMode::Gauss).unwrap();
 ///
 /// // Operator fields
-/// let op = ceed.operator(&qf, QFunctionOpt::None, QFunctionOpt::None)
-///     .field("dx", &r, &b, VectorOpt::Active)
-///     .field("weights", ElemRestrictionOpt::None, &b, VectorOpt::None)
-///     .field("qdata", &rq, BasisOpt::Collocated, VectorOpt::Active);
+/// let op = ceed.operator(&qf, QFunctionOpt::None, QFunctionOpt::None).unwrap()
[…]

[…]
