### 2021-01-28 32d511c
rust - add results for methods that may fail
