+++ rust/src/lib.rs
@@ -41,7 +41,7 @@
 //!
 //! fn main() {
 //!     let ceed = libceed::Ceed::init("/cpu/self/ref");
-//!     let xc = ceed.vector_from_slice(&[0., 0.5, 1.0]);
+//!     let xc = ceed.vector_from_slice(&[0., 0.5, 1.0]).unwrap();
 //!     let xs = xc.view();
 //!     assert_eq!(xs[..], [0., 0.5, 1.0]);
 //! }
@@ -188,7 +188,7 @@ type Result<T> = std::result::Result<T, CeedError>;
 
 #[derive(Debug)]
 pub struct CeedError {
-    message: String,
+    pub message: String,
 }
 
 impl fmt::Display for CeedError {
@@ -377,9 +377,9 @@ impl Ceed {
     /// ```
     /// # use libceed::prelude::*;
     /// # let ceed = libceed::Ceed::default_init();
-    /// let vec = ceed.vector(10);
+    /// let vec = ceed.vector(10).unwrap();
     /// ```
-    pub fn vector(&self, n: usize) -> Vector {
+    pub fn vector(&self, n: usize) -> Result<Vector> {
         Vector::create(self, n)
     }
 
@@ -392,10 +392,10 @@ impl Ceed {
     /// ```
     /// # use libceed::prelude::*;
     /// # let ceed = libceed::Ceed::default_init();
-    /// let vec = ceed.vector_from_slice(&[1., 2., 3.]);
+    /// let vec = ceed.vector_from_slice(&[1., 2., 3.]).unwrap();
     /// assert_eq!(vec.length(), 3);
     /// ```
-    pub fn vector_from_slice(&self, slice: &[f64]) -> Vector {
+    pub fn vector_from_slice(&self, slice: &[f64]) -> Result<Vector> {
         Vector::from_slice(self, slice)
     }
 
@@ -430,7 +430,7 @@ impl Ceed {
     ///   ind[2 * i + 0] = i as i32;
     ///   ind[2 * i + 1] = (i + 1) as i32;
     /// }
-    /// let r = ceed.elem_restriction(nelem, 2, 1, 1, nelem + 1, MemType::Host, &ind);
+    /// let r = ceed.elem_restriction(nelem, 2, 1, 1, nelem + 1, MemType::Host, &ind).unwrap();
     /// ```
     pub fn elem_restriction(
         &self,
@@ -441,7 +441,7 @@ impl Ceed {
         lsize: usize,
         mtype: MemType,
         offsets: &[i32],
-    ) -> ElemRestriction {
+    ) -> Result<ElemRestriction> {
         ElemRestriction::create(
             self, nelem, elemsize, ncomp, compstride, lsize, mtype, offsets,
         )
@@ -473,7 +473,7 @@ impl Ceed {
     /// # let ceed = libceed::Ceed::default_init();
     /// let nelem = 3;
     /// let strides : [i32; 3] = [1, 2, 2];
-    /// let r = ceed.strided_elem_restriction(nelem, 2, 1, nelem * 2, strides);
+    /// let r = ceed.strided_elem_restriction(nelem, 2, 1, nelem * 2, strides).unwrap();
     /// ```
     pub fn strided_elem_restriction(
         &self,
@@ -482,7 +482,7 @@ impl Ceed {
         ncomp: usize,
         lsize: usize,
         strides: [i32; 3],
-    ) -> ElemRestriction {
+    ) -> Result<ElemRestriction> {
         ElemRestriction::create_strided(self, nelem, elemsize, ncomp, lsize, strides)
     }
 
@@ -519,7 +519,7 @@ impl Ceed {
     /// let qref1d    = [-0.86113631, -0.33998104,  0.33998104,  0.86113631];
[…]
