+++ examples/rust/ex2-surface/src/main.rs
@@ -39,12 +39,12 @@ mod transform;
 // Example 2
 // ----------------------------------------------------------------------------
 #[cfg(not(tarpaulin_include))]
-fn main() -> Result<(), String> {
+fn main() -> Result<(), libceed::CeedError> {
     let options = opt::Opt::from_args();
     example_2(options)
 }
 
-fn example_2(options: opt::Opt) -> Result<(), String> {
+fn example_2(options: opt::Opt) -> Result<(), libceed::CeedError> {
     // Process command line arguments
     let opt::Opt {
         ceed_spec,
@@ -92,10 +92,12 @@ fn example_2(options: opt::Opt) -> Result<(), String> {
     let ceed = Ceed::init(&ceed_spec);
 
     // Mesh and solution bases
-    let basis_mesh =
-        ceed.basis_tensor_H1_Lagrange(dim, ncomp_x, mesh_degree + 1, num_qpts, QuadMode::Gauss);
-    let basis_solution =
-        ceed.basis_tensor_H1_Lagrange(dim, 1, solution_degree + 1, num_qpts, QuadMode::Gauss);
+    let basis_mesh = ceed
+        .basis_tensor_H1_Lagrange(dim, ncomp_x, mesh_degree + 1, num_qpts, QuadMode::Gauss)
+        .unwrap();
+    let basis_solution = ceed
+        .basis_tensor_H1_Lagrange(dim, 1, solution_degree + 1, num_qpts, QuadMode::Gauss)
+        .unwrap();
 
     // Determine mesh size from approximate problem size
     let num_xyz = mesh::cartesian_mesh_size(dim, solution_degree, problem_size);
@@ -113,7 +115,7 @@ fn example_2(options: opt::Opt) -> Result<(), String> {
     // Build ElemRestriction objects describing the mesh and solution discrete
     // representations
     let (restr_mesh, _) =
-        mesh::build_cartesian_restriction(&ceed, dim, num_xyz, mesh_degree, ncomp_x, num_qpts);
+        mesh::build_cartesian_restriction(&ceed, dim, num_xyz, mesh_degree, ncomp_x, num_qpts)?;
     let (_, restr_qdata) = mesh::build_cartesian_restriction(
         &ceed,
         dim,
@@ -121,10 +123,10 @@ fn example_2(options: opt::Opt) -> Result<(), String> {
         solution_degree,
         dim * (dim + 1) / 2,
         num_qpts,
-    );
+    )?;
 
     let (restr_solution, _) =
-        mesh::build_cartesian_restriction(&ceed, dim, num_xyz, solution_degree, 1, num_qpts);
+        mesh::build_cartesian_restriction(&ceed, dim, num_xyz, solution_degree, 1, num_qpts)?;
     let mesh_size = restr_mesh.lvector_size();
     let solution_size = restr_solution.lvector_size();
     if !quiet {
@@ -133,7 +135,7 @@ fn example_2(options: opt::Opt) -> Result<(), String> {
     }
 
     // Create a Vector with the mesh coordinates
-    let mut mesh_coords = mesh::cartesian_mesh_coords(&ceed, dim, num_xyz, mesh_degree, mesh_size);
+    let mut mesh_coords = mesh::cartesian_mesh_coords(&ceed, dim, num_xyz, mesh_degree, mesh_size)?;
 
     // Apply a transformation to the mesh coordinates
     let exact_area = transform::transform_mesh_coordinates(dim, &mut mesh_coords);
@@ -210,14 +212,14 @@ fn example_2(options: opt::Opt) -> Result<(), String> {
         0
     };
     let qf_build_closure = ceed
-        .q_function_interior(1, Box::new(build_diff))
-        .input("dx", ncomp_x * dim, EvalMode::Grad)
-        .input("weights", 1, EvalMode::Weight)
[…]

[…]
