### 2021-01-28 61ca1c8
rust - also format doctests
