+++ rust/src/elem_restriction.rs
@@ -78,10 +78,12 @@ impl<'a> fmt::Display for ElemRestriction<'a> {
     /// let nelem = 3;
     /// let mut ind: Vec<i32> = vec![0; 2 * nelem];
     /// for i in 0..nelem {
-    ///   ind[2 * i + 0] = i as i32;
-    ///   ind[2 * i + 1] = (i + 1) as i32;
+    ///     ind[2 * i + 0] = i as i32;
+    ///     ind[2 * i + 1] = (i + 1) as i32;
     /// }
-    /// let r = ceed.elem_restriction(nelem, 2, 1, 1, nelem + 1, MemType::Host, &ind).unwrap();
+    /// let r = ceed
+    ///     .elem_restriction(nelem, 2, 1, 1, nelem + 1, MemType::Host, &ind)
+    ///     .unwrap();
     /// println!("{}", r);
     /// ```
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
@@ -177,10 +179,12 @@ impl<'a> ElemRestriction<'a> {
     /// let nelem = 3;
     /// let mut ind: Vec<i32> = vec![0; 2 * nelem];
     /// for i in 0..nelem {
-    ///   ind[2 * i + 0] = i as i32;
-    ///   ind[2 * i + 1] = (i + 1) as i32;
+    ///     ind[2 * i + 0] = i as i32;
+    ///     ind[2 * i + 1] = (i + 1) as i32;
     /// }
-    /// let r = ceed.elem_restriction(nelem, 2, 1, 1, nelem + 1, MemType::Host, &ind).unwrap();
+    /// let r = ceed
+    ///     .elem_restriction(nelem, 2, 1, 1, nelem + 1, MemType::Host, &ind)
+    ///     .unwrap();
     ///
     /// let lvector = r.create_lvector().unwrap();
     ///
@@ -203,10 +207,12 @@ impl<'a> ElemRestriction<'a> {
     /// let nelem = 3;
     /// let mut ind: Vec<i32> = vec![0; 2 * nelem];
     /// for i in 0..nelem {
-    ///   ind[2 * i + 0] = i as i32;
-    ///   ind[2 * i + 1] = (i + 1) as i32;
+    ///     ind[2 * i + 0] = i as i32;
+    ///     ind[2 * i + 1] = (i + 1) as i32;
     /// }
-    /// let r = ceed.elem_restriction(nelem, 2, 1, 1, nelem + 1, MemType::Host, &ind).unwrap();
+    /// let r = ceed
+    ///     .elem_restriction(nelem, 2, 1, 1, nelem + 1, MemType::Host, &ind)
+    ///     .unwrap();
     ///
     /// let evector = r.create_evector().unwrap();
     ///
@@ -229,10 +235,12 @@ impl<'a> ElemRestriction<'a> {
     /// let nelem = 3;
     /// let mut ind: Vec<i32> = vec![0; 2 * nelem];
     /// for i in 0..nelem {
-    ///   ind[2 * i + 0] = i as i32;
-    ///   ind[2 * i + 1] = (i + 1) as i32;
+    ///     ind[2 * i + 0] = i as i32;
+    ///     ind[2 * i + 1] = (i + 1) as i32;
     /// }
-    /// let r = ceed.elem_restriction(nelem, 2, 1, 1, nelem + 1, MemType::Host, &ind).unwrap();
+    /// let r = ceed
+    ///     .elem_restriction(nelem, 2, 1, 1, nelem + 1, MemType::Host, &ind)
+    ///     .unwrap();
     ///
     /// let (lvector, evector) = r.create_vectors().unwrap();
     ///
@@ -267,23 +275,26 @@ impl<'a> ElemRestriction<'a> {
     /// let nelem = 3;
[…]

[…]
