### 2021-01-28 9086585
rust - also format doctests
