+++ rust/src/qfunction.rs
@@ -129,25 +129,25 @@ impl<'a> fmt::Display for QFunctionCore<'a> {
 /// ```
 /// # use libceed::prelude::*;
 /// # let ceed = libceed::Ceed::default_init();
-/// let mut user_f = |
-///   [u, weights, ..]: QFunctionInputs,
-///   [v, ..]: QFunctionOutputs,
-/// |
-/// {
-///   // Iterate over quadrature points
-///   v
-///     .iter_mut()
-///     .zip(u.iter().zip(weights.iter()))
-///     .for_each(|(v, (u, w))| *v = u * w);
+/// let mut user_f = |[u, weights, ..]: QFunctionInputs, [v, ..]: QFunctionOutputs| {
+///     // Iterate over quadrature points
+///     v.iter_mut()
+///         .zip(u.iter().zip(weights.iter()))
+///         .for_each(|(v, (u, w))| *v = u * w);
 ///
-///   // Return clean error code
-///   0
+///     // Return clean error code
+///     0
 /// };
 ///
-/// let qf = ceed.q_function_interior(1, Box::new(user_f)).unwrap()
-///     .input("u", 1, EvalMode::Interp).unwrap()
-///     .input("weights", 1, EvalMode::Weight).unwrap()
-///     .output("v", 1, EvalMode::Interp).unwrap();
+/// let qf = ceed
+///     .q_function_interior(1, Box::new(user_f))
+///     .unwrap()
+///     .input("u", 1, EvalMode::Interp)
+///     .unwrap()
+///     .input("weights", 1, EvalMode::Weight)
+///     .unwrap()
+///     .output("v", 1, EvalMode::Interp)
+///     .unwrap();
 ///
 /// println!("{}", qf);
 /// ```
@@ -318,25 +318,25 @@ impl<'a> QFunction<'a> {
     /// ```
     /// # use libceed::prelude::*;
     /// # let ceed = libceed::Ceed::default_init();
-    /// let mut user_f = |
-    ///   [u, weights, ..]: QFunctionInputs,
-    ///   [v, ..]: QFunctionOutputs,
-    /// |
-    /// {
-    ///   // Iterate over quadrature points
-    ///   v
-    ///     .iter_mut()
-    ///     .zip(u.iter().zip(weights.iter()))
-    ///     .for_each(|(v, (u, w))| *v = u * w);
+    /// let mut user_f = |[u, weights, ..]: QFunctionInputs, [v, ..]: QFunctionOutputs| {
+    ///     // Iterate over quadrature points
+    ///     v.iter_mut()
+    ///         .zip(u.iter().zip(weights.iter()))
+    ///         .for_each(|(v, (u, w))| *v = u * w);
     ///
-    ///   // Return clean error code
-    ///   0
+    ///     // Return clean error code
+    ///     0
     /// };
     ///
-    /// let qf = ceed.q_function_interior(1, Box::new(user_f)).unwrap()
-    ///     .input("u", 1, EvalMode::Interp).unwrap()
-    ///     .input("weights", 1, EvalMode::Weight).unwrap()
-    ///     .output("v", 1, EvalMode::Interp).unwrap();
+    /// let qf = ceed
+    ///     .q_function_interior(1, Box::new(user_f))
+    ///     .unwrap()
+    ///     .input("u", 1, EvalMode::Interp)
+    ///     .unwrap()
+    ///     .input("weights", 1, EvalMode::Weight)
+    ///     .unwrap()
+    ///     .output("v", 1, EvalMode::Interp)
+    ///     .unwrap();
     ///
     /// const Q: usize = 8;
     /// let mut w = [0.; Q];
@@ -344,10 +344,10 @@ impl<'a> QFunction<'a> {
[…]
