### 2021-01-28 e913242
change c_* to s_* for shared memory arrays and remove old fcn declarations
