+++ backends/hip-shared/ceed-hip-shared-basis.c
@@ -125,7 +125,7 @@ inline __device__ void ContractTransposeX1d(CeedScalar *slice, const int tidx,
 // 1D interpolate to quadrature points
 //------------------------------------------------------------------------------
 inline __device__ void interp1d(const CeedInt nelem, const int transpose,
-                                const CeedScalar *c_B,
+                                const CeedScalar *s_B,
                                 const CeedScalar *__restrict__ d_U,
                                 CeedScalar *__restrict__ d_V,
                                 CeedScalar *slice) {
@@ -142,11 +142,11 @@ inline __device__ void interp1d(const CeedInt nelem, const int transpose,
     for (int comp = 0; comp < BASIS_NCOMP; comp++) {
       if (!transpose) {
         readDofs1d(elem, tidx, tidy, tidz, comp, nelem, d_U, slice);
-        ContractX1d(slice, tidx, tidy, tidz, r_t, c_B, r_V);
+        ContractX1d(slice, tidx, tidy, tidz, r_t, s_B, r_V);
         writeQuads1d(elem, tidx, tidy, comp, 0, nelem, r_V, d_V);
       } else {
         readQuads1d(elem, tidx, tidy, tidz, comp, 0, nelem, d_U, slice);
-        ContractTransposeX1d(slice, tidx, tidy, tidz, r_t, c_B, r_V);
+        ContractTransposeX1d(slice, tidx, tidy, tidz, r_t, s_B, r_V);
         writeDofs1d(elem, tidx, tidy, comp, nelem, r_V, d_V);
       }
     }
@@ -157,7 +157,7 @@ inline __device__ void interp1d(const CeedInt nelem, const int transpose,
 // 1D derivatives at quadrature points
 //------------------------------------------------------------------------------
 inline __device__ void grad1d(const CeedInt nelem, const int transpose,
-                              const CeedScalar *c_B, const CeedScalar *c_G,
+                              const CeedScalar *s_B, const CeedScalar *s_G,
                               const CeedScalar *__restrict__ d_U,
                               CeedScalar *__restrict__ d_V,
                               CeedScalar *slice) {
@@ -174,13 +174,13 @@ inline __device__ void grad1d(const CeedInt nelem, const int transpose,
     for(int comp = 0; comp < BASIS_NCOMP; comp++) {
       if (!transpose) {
         readDofs1d(elem, tidx, tidy, tidz, comp, nelem, d_U, slice);
-        ContractX1d(slice, tidx, tidy, tidz, r_U, c_G, r_V);
+        ContractX1d(slice, tidx, tidy, tidz, r_U, s_G, r_V);
         dim = 0;
         writeQuads1d(elem, tidx, tidy, comp, dim, nelem, r_V, d_V);
       } else {
         dim = 0;
         readQuads1d(elem, tidx, tidy, tidz, comp, dim, nelem, d_U, slice);
-        ContractTransposeX1d(slice, tidx, tidy, tidz, r_U, c_G, r_V);
+        ContractTransposeX1d(slice, tidx, tidy, tidz, r_U, s_G, r_V);
         writeDofs1d(elem, tidx, tidy, comp, nelem, r_V, d_V);
       }
     }
[…]
