### 2021-01-28 1f814bf
Revert "Temporary change for convergence test and debugging"
This reverts commit d326fa595e8c6b2326042f88f316f0d31095b876.
