+++ backends/xsmm/ceed-xsmm-serial.c
@@ -14,6 +14,8 @@
 // software, applications, hardware, advanced system engineering and early
 // testbed platforms, in support of the nation's exascale computing imperative.
 
+#include <ceed.h>
+#include <ceed-backend.h>
 #include "ceed-xsmm.h"
 
 //------------------------------------------------------------------------------
