+++ backends/ref/ceed-ref.h
@@ -14,9 +14,8 @@
 // software, applications, hardware, advanced system engineering and early
 // testbed platforms, in support of the nation's exascale computing imperative.
 
-#include <ceed-backend.h>
+#include <ceed.h>
 #include <string.h>
-#include <math.h>
 
 typedef struct {
   CeedScalar *collograd1d;
