+++ backends/cuda-shared/ceed-cuda-shared.c
@@ -14,8 +14,10 @@
 // software, applications, hardware, advanced system engineering and early
 // testbed platforms, in support of the nation's exascale computing imperative.
 
-#include <string.h>
+#include <ceed.h>
+#include <ceed-backend.h>
 #include <stdarg.h>
+#include <string.h>
 #include "ceed-cuda-shared.h"
 
 //------------------------------------------------------------------------------
