+++ interface/ceed-vector.c
@@ -14,8 +14,9 @@
 // software, applications, hardware, advanced system engineering and early
 // testbed platforms, in support of the nation's exascale computing imperative.
 
-#include <ceed-impl.h>
+#include <ceed.h>
 #include <ceed-backend.h>
+#include <ceed-impl.h>
 #include <math.h>
 
 /// @file
