+++ backends/avx/ceed-avx-tensor.c
@@ -14,6 +14,9 @@
 // software, applications, hardware, advanced system engineering and early
 // testbed platforms, in support of the nation's exascale computing imperative.
 
+#include <ceed.h>
+#include <ceed-backend.h>
+#include <immintrin.h>
 #include "ceed-avx.h"
 
 // c += a * b
