+++ backends/blocked/ceed-blocked.h
@@ -14,8 +14,8 @@
 // software, applications, hardware, advanced system engineering and early
 // testbed platforms, in support of the nation's exascale computing imperative.
 
+#include <ceed.h>
 #include <ceed-backend.h>
-#include <string.h>
 
 typedef struct {
   CeedScalar *colograd1d;
