+++ interface/ceed-operator.c
@@ -14,10 +14,11 @@
 // software, applications, hardware, advanced system engineering and early
 // testbed platforms, in support of the nation's exascale computing imperative.
 
-#include <ceed-impl.h>
+#include <ceed.h>
 #include <ceed-backend.h>
-#include <string.h>
+#include <ceed-impl.h>
 #include <math.h>
+#include <string.h>
 
 /// @file
 /// Implementation of CeedOperator interfaces
