### 2021-01-28 aabe6db
minor - drop unused CeedChk()
