+++ interface/ceed-operator.c
@@ -69,7 +69,7 @@ int CeedOperatorCreateFallback(CeedOperator op) {
   // Clone Op
   CeedOperator opref;
   ierr = CeedCalloc(1, &opref); CeedChk(ierr);
-  memcpy(opref, op, sizeof(*opref)); CeedChk(ierr);
+  memcpy(opref, op, sizeof(*opref));
   opref->data = NULL;
   opref->interfacesetup = false;
   opref->backendsetup = false;
@@ -80,7 +80,7 @@ int CeedOperatorCreateFallback(CeedOperator op) {
   // Clone QF
   CeedQFunction qfref;
   ierr = CeedCalloc(1, &qfref); CeedChk(ierr);
-  memcpy(qfref, (op->qf), sizeof(*qfref)); CeedChk(ierr);
+  memcpy(qfref, (op->qf), sizeof(*qfref));
   qfref->data = NULL;
   qfref->ceed = ceedref;
   ierr = ceedref->QFunctionCreate(qfref); CeedChk(ierr);
