### 2021-01-28 6697a05
error - rename terminal/nonterminal to major/minor
