+++ interface/ceed-types.c
@@ -18,11 +18,11 @@
 
 const char *const CeedErrorTypesShifted[] = {
   [CEED_ERROR_SUCCESS - CEED_ERROR_UNSUPPORTED] = "success",
-  [CEED_ERROR_NONTERMINAL - CEED_ERROR_UNSUPPORTED] = "nonterminal error",
+  [CEED_ERROR_MINOR - CEED_ERROR_UNSUPPORTED] = "generic minor error",
   [CEED_ERROR_DIMENSION - CEED_ERROR_UNSUPPORTED] = "dimension mismatch",
   [CEED_ERROR_INCOMPLETE - CEED_ERROR_UNSUPPORTED] = "object setup incomplete",
   [CEED_ERROR_ACCESS - CEED_ERROR_UNSUPPORTED] = "access lock in use",
-  [CEED_ERROR_TERMINAL - CEED_ERROR_UNSUPPORTED] = "terminal error",
+  [CEED_ERROR_MAJOR - CEED_ERROR_UNSUPPORTED] = "generic major error",
   [CEED_ERROR_BACKEND - CEED_ERROR_UNSUPPORTED] = "internal backend error",
   [CEED_ERROR_UNSUPPORTED - CEED_ERROR_UNSUPPORTED] = "operation unsupported by backend",
 };
