### 2021-01-28 1a364d4
error - rename terminal/nonterminal to major/minor
