+++ interface/ceed-operator.c
@@ -111,7 +111,7 @@ static int CeedOperatorCheckField(Ceed ceed, CeedQFunctionField qfield,
     ierr = CeedElemRestrictionGetNumComponents(r, &restr_ncomp);
   } else if (emode != CEED_EVAL_WEIGHT) {
     // LCOV_EXCL_START
-    return CeedError(ceed, CEED_ERROR_NONTERMINAL,
+    return CeedError(ceed, CEED_ERROR_MINOR,
                      "CEED_ELEMRESTRICTION_NONE can only be used "
                      "for a field with eval mode CEED_EVAL_WEIGHT");
     // LCOV_EXCL_STOP
@@ -319,7 +319,7 @@ static int CeedOperatorGetActiveBasis(CeedOperator op,
     int ierr;
     Ceed ceed;
     ierr = CeedOperatorGetCeed(op, &ceed); CeedChk(ierr);
-    return CeedError(ceed, CEED_ERROR_NONTERMINAL,
+    return CeedError(ceed, CEED_ERROR_MINOR,
                      "No active basis found for automatic multigrid setup");
     // LCOV_EXCL_STOP
   }
@@ -520,7 +520,7 @@ int CeedOperatorGetCeed(CeedOperator op, Ceed *ceed) {
 int CeedOperatorGetNumElements(CeedOperator op, CeedInt *numelem) {
   if (op->composite)
     // LCOV_EXCL_START
-    return CeedError(op->ceed, CEED_ERROR_NONTERMINAL,
+    return CeedError(op->ceed, CEED_ERROR_MINOR,
                      "Not defined for composite operator");
   // LCOV_EXCL_STOP
 
@@ -542,7 +542,7 @@ int CeedOperatorGetNumElements(CeedOperator op, CeedInt *numelem) {
 int CeedOperatorGetNumQuadraturePoints(CeedOperator op, CeedInt *numqpts) {
   if (op->composite)
     // LCOV_EXCL_START
-    return CeedError(op->ceed, CEED_ERROR_NONTERMINAL,
+    return CeedError(op->ceed, CEED_ERROR_MINOR,
                      "Not defined for composite operator");
   // LCOV_EXCL_STOP
 
@@ -564,7 +564,7 @@ int CeedOperatorGetNumQuadraturePoints(CeedOperator op, CeedInt *numqpts) {
 int CeedOperatorGetNumArgs(CeedOperator op, CeedInt *numargs) {
   if (op->composite)
     // LCOV_EXCL_START
-    return CeedError(op->ceed, CEED_ERROR_NONTERMINAL,
+    return CeedError(op->ceed, CEED_ERROR_MINOR,
                      "Not defined for composite operators");
   // LCOV_EXCL_STOP
 
@@ -602,7 +602,7 @@ int CeedOperatorIsSetupDone(CeedOperator op, bool *issetupdone) {
 int CeedOperatorGetQFunction(CeedOperator op, CeedQFunction *qf) {
   if (op->composite)
     // LCOV_EXCL_START
-    return CeedError(op->ceed, CEED_ERROR_NONTERMINAL,
+    return CeedError(op->ceed, CEED_ERROR_MINOR,
                      "Not defined for composite operator");
   // LCOV_EXCL_STOP
 
@@ -640,7 +640,7 @@ int CeedOperatorIsComposite(CeedOperator op, bool *iscomposite) {
 int CeedOperatorGetNumSub(CeedOperator op, CeedInt *numsub) {
   if (!op->composite)
     // LCOV_EXCL_START
-    return CeedError(op->ceed, CEED_ERROR_NONTERMINAL, "Not a composite operator");
+    return CeedError(op->ceed, CEED_ERROR_MINOR, "Not a composite operator");
[…]
