+++ interface/ceed-elemrestriction.c
@@ -97,7 +97,7 @@ int CeedElemRestrictionGetStrides(CeedElemRestriction rstr,
                                   CeedInt (*strides)[3]) {
   if (!rstr->strides)
     // LCOV_EXCL_START
-    return CeedError(rstr->ceed, CEED_ERROR_NONTERMINAL,
+    return CeedError(rstr->ceed, CEED_ERROR_MINOR,
                      "ElemRestriction has no stride data");
   // LCOV_EXCL_STOP
 
@@ -180,7 +180,7 @@ int CeedElemRestrictionHasBackendStrides(CeedElemRestriction rstr,
     bool *hasbackendstrides) {
   if (!rstr->strides)
     // LCOV_EXCL_START
-    return CeedError(rstr->ceed, CEED_ERROR_NONTERMINAL,
+    return CeedError(rstr->ceed, CEED_ERROR_MINOR,
                      "ElemRestriction has no stride data");
   // LCOV_EXCL_STOP
 
@@ -209,7 +209,7 @@ int CeedElemRestrictionGetELayout(CeedElemRestriction rstr,
                                   CeedInt (*layout)[3]) {
   if (!rstr->layout[0])
     // LCOV_EXCL_START
-    return CeedError(rstr->ceed, CEED_ERROR_NONTERMINAL,
+    return CeedError(rstr->ceed, CEED_ERROR_MINOR,
                      "ElemRestriction has no layout data");
   // LCOV_EXCL_STOP
 

[…]
