### 2021-01-28 7961cc9
rust - set ErrorStore as default errorhandler
