### 2021-01-28 5525050
add syncthreads after array read
