+++ backends/xsmm/ceed-xsmm-serial.c
@@ -16,6 +16,8 @@
 
 #include <ceed.h>
 #include <ceed-backend.h>
+#include <stdbool.h>
+#include <string.h>
 #include "ceed-xsmm.h"
 
 //------------------------------------------------------------------------------
