+++ interface/ceed-basis.c
@@ -18,8 +18,8 @@
 #include <ceed-backend.h>
 #include <ceed-impl.h>
 #include <math.h>
+#include <stdbool.h>
 #include <stdio.h>
-#include <stdlib.h>
 #include <string.h>
 
 /// @file
