### 2021-01-28 9c167f8
python - revert error handing change for python
