+++ python/ceed.py
@@ -78,12 +78,10 @@ class Ceed():
     # Error handler
     def _check_error(self, err_code):
         """Check return code and retrieve error message for non-zero code"""
-        if (err_code > 0):
+        if (err_code != 0):
             message = ffi.new("char **")
             lib.CeedGetErrorMessage(self._pointer[0], message)
             raise Exception(ffi.string(message[0]).decode("UTF-8"))
-        elif (err_code < 0):
-            sys.exit(ffi.string(message[0]).decode("UTF-8"))
 
     # Get Resource
     def get_resource(self):
