+++ backends/cuda-gen/ceed-cuda-gen-qfunction.c
@@ -16,9 +16,11 @@
 
 #include <ceed.h>
 #include <ceed-backend.h>
+#include <cuda_runtime.h>
 #include <stdio.h>
 #include <string.h>
 #include "ceed-cuda-gen.h"
+#include "../cuda/ceed-cuda.h"
 
 //------------------------------------------------------------------------------
 // Apply QFunction
