+++ backends/cuda-gen/ceed-cuda-gen.c
@@ -16,9 +16,9 @@
 
 #include <ceed.h>
 #include <ceed-backend.h>
-#include <stdarg.h>
 #include <string.h>
 #include "ceed-cuda-gen.h"
+#include "../cuda/ceed-cuda.h"
 
 //------------------------------------------------------------------------------
 // Backend init
