+++ backends/cuda-gen/ceed-cuda-gen-operator-build.cpp
@@ -17,6 +17,7 @@
 
 #include <ceed.h>
 #include <ceed-backend.h>
+#include <cuda_runtime.h>
 #include <iostream>
 #include <sstream>
 #include "ceed-cuda-gen.h"
