+++ backends/cuda/ceed-cuda-qfunction.c
@@ -16,6 +16,7 @@
 
 #include <ceed.h>
 #include <ceed-backend.h>
+#include <cuda.h>
 #include <stdio.h>
 #include <string.h>
 #include "ceed-cuda.h"
