+++ backends/cuda/ceed-cuda.h
@@ -20,7 +20,6 @@
 #include <ceed-backend.h>
 #include <cublas_v2.h>
 #include <cuda.h>
-#include <cuda_runtime.h>
 #include <nvrtc.h>
 
 #define CUDA_MAX_PATH 256
