+++ backends/cuda-shared/ceed-cuda-shared.c
@@ -16,9 +16,10 @@
 
 #include <ceed.h>
 #include <ceed-backend.h>
-#include <stdarg.h>
+#include <stdbool.h>
 #include <string.h>
 #include "ceed-cuda-shared.h"
+#include "../cuda/ceed-cuda.h"
 
 //------------------------------------------------------------------------------
 // Backend init
