+++ backends/hip/ceed-hip-vector.c
@@ -16,6 +16,9 @@
 
 #include <ceed.h>
 #include <ceed-backend.h>
+#include <hip/hip_runtime.h>
+#include <hipblas.h>
+#include <math.h>
 #include <string.h>
 #include "ceed-hip.h"
 
