+++ backends/hip-shared/ceed-hip-shared.c
@@ -16,9 +16,10 @@
 
 #include <ceed.h>
 #include <ceed-backend.h>
-#include <stdarg.h>
+#include <stdbool.h>
 #include <string.h>
 #include "ceed-hip-shared.h"
+#include "../hip/ceed-hip.h"
 
 //------------------------------------------------------------------------------
 // Backend init
