+++ backends/hip-gen/ceed-hip-gen-operator-build.cpp
@@ -18,8 +18,10 @@
 #include <ceed.h>
 #include <ceed-backend.h>
 #include <iostream>
+#include <string>
 #include <sstream>
 #include "ceed-hip-gen.h"
+#include "../hip/ceed-hip.h"
 #include "../hip-shared/ceed-hip-shared.h"
 #include "../hip/ceed-hip-compile.h"
 
