+++ backends/hip/ceed-hip-basis.c
@@ -16,6 +16,7 @@
 
 #include <ceed.h>
 #include <ceed-backend.h>
+#include <hip/hip_runtime.h>
 #include "ceed-hip.h"
 #include "ceed-hip-compile.h"
 
