### 2021-01-29 de2862f
style - drop ceed.h includes in gallery qf source
