+++ gallery/mass2d/ceed-mass2dbuild.h
@@ -20,9 +20,6 @@
 
 #ifndef mass2dbuild_h
 #define mass2dbuild_h
-#ifndef __CUDACC__
-#  include <ceed.h>
-#endif
 
 CEED_QFUNCTION(Mass2DBuild)(void *ctx, const CeedInt Q,
                             const CeedScalar *const *in, CeedScalar *const *out) {
