+++ gallery/mass3d/ceed-mass3dbuild.h
@@ -20,9 +20,6 @@
 
 #ifndef mass3dbuild_h
 #define mass3dbuild_h
-#ifndef __CUDACC__
-#  include <ceed.h>
-#endif
 
 CEED_QFUNCTION(Mass3DBuild)(void *ctx, const CeedInt Q,
                             const CeedScalar *const *in, CeedScalar *const *out) {
