+++ backends/cuda/ceed-cuda.h
@@ -13,6 +13,7 @@
 // the planning and preparation of a capable exascale ecosystem, including
 // software, applications, hardware, advanced system engineering and early
 // testbed platforms, in support of the nation's exascale computing imperative.
+
 #ifndef _ceed_cuda_h
 #define _ceed_cuda_h
 
