+++ backends/hip-gen/ceed-hip-gen.h
@@ -14,6 +14,9 @@
 // software, applications, hardware, advanced system engineering and early
 // testbed platforms, in support of the nation's exascale computing imperative.
 
+#ifndef _ceed_hip_gen_h
+#define _ceed_hip_gen_h
+
 #include <ceed.h>
 #include <ceed-backend.h>
 #include <hip/hip_runtime.h>
@@ -50,3 +53,5 @@ CEED_INTERN int CeedQFunctionCreate_Hip_gen(CeedQFunction qf);
 CEED_INTERN int CeedOperatorCreate_Hip_gen(CeedOperator op);
 
 CEED_INTERN int CeedCompositeOperatorCreate_Hip_gen(CeedOperator op);
+
+#endif // _ceed_hip_gen_h
