+++ backends/memcheck/ceed-memcheck.h
@@ -14,6 +14,9 @@
 // software, applications, hardware, advanced system engineering and early
 // testbed platforms, in support of the nation's exascale computing imperative.
 
+#ifndef _ceed_memcheck_h
+#define _ceed_memcheck_h
+
 #include <ceed.h>
 #include <ceed-backend.h>
 
@@ -24,3 +27,5 @@ typedef struct {
 } CeedQFunction_Memcheck;
 
 CEED_INTERN int CeedQFunctionCreate_Memcheck(CeedQFunction qf);
+
+#endif // _ceed_memcheck_h
