+++ backends/opt/ceed-opt.h
@@ -14,6 +14,9 @@
 // software, applications, hardware, advanced system engineering and early
 // testbed platforms, in support of the nation's exascale computing imperative.
 
+#ifndef _ceed_opt_h
+#define _ceed_opt_h
+
 #include <ceed.h>
 #include <ceed-backend.h>
 #include <stdbool.h>
@@ -43,3 +46,5 @@ typedef struct {
 } CeedOperator_Opt;
 
 CEED_INTERN int CeedOperatorCreate_Opt(CeedOperator op);
+
+#endif // _ceed_opt_h
