+++ backends/hip-gen/ceed-hip-gen-operator-build.cpp
@@ -13,6 +13,7 @@
 // the planning and preparation of a capable exascale ecosystem, including
 // software, applications, hardware, advanced system engineering and early
 // testbed platforms, in support of the nation's exascale computing imperative.
+
 #define CEED_DEBUG_COLOR 12
 
 #include <ceed.h>
