+++ examples/fluids/euler-vortex.h
@@ -44,6 +44,7 @@ struct EulerContext_ {
   CeedScalar etv_mean_velocity[3];
   CeedScalar T_inlet;
   CeedScalar P_inlet;
+  int stabilization;
   int euler_test;
 };
 #endif
@@ -260,7 +261,7 @@ static inline int MMSforce_Euler(CeedInt dim, CeedScalar time,
 
   // Setup
   const CeedScalar gamma = 1.4;
-  const CeedScalar cv = 2.5; // cv computed based on Rd = 1
+  const CeedScalar cv = 2.5; // cv computed based on R = 1
   const CeedScalar x = X[0], y = X[1], z = X[2]; // Coordinates
   // Vortex center
   const CeedScalar xc = center[0] + etv_mean_velocity[0] * time;
@@ -296,7 +297,10 @@ CEED_QFUNCTION(Euler)(void *ctx, CeedInt Q,
   // Context
   const EulerContext context = (EulerContext)ctx;
   const CeedScalar currentTime = context->currentTime;
+  const int stabilization = context->stabilization;
   const CeedScalar gamma  = 1.4;
+  const CeedScalar cv = 2.5;
+  const CeedScalar R = 1.;
 
   CeedPragmaSIMD
   // Quadrature Point Loop
@@ -310,6 +314,25 @@ CEED_QFUNCTION(Euler)(void *ctx, CeedInt Q,
                                     q[3][i] / rho
                                    };
     const CeedScalar E          =   q[4][i];
+    // -- Grad in
+    const CeedScalar drho[3]    =  {dq[0][0][i],
+                                    dq[1][0][i],
+                                    dq[2][0][i]
+                                   };
+    const CeedScalar dU[3][3]   = {{dq[0][1][i],
+                                    dq[1][1][i],
+                                    dq[2][1][i]},
+                                   {dq[0][2][i],
+                                    dq[1][2][i],
+                                    dq[2][2][i]},
+                                   {dq[0][3][i],
+                                    dq[1][3][i],
+                                    dq[2][3][i]}
+                                  };
+    const CeedScalar dE[3]      =  {dq[0][4][i],
+                                    dq[1][4][i],
+                                    dq[2][4][i]
+                                   };
     // -- Interp-to-Interp qdata
     const CeedScalar wdetJ      =   qdata[0][i];
     // -- Interp-to-Grad qdata
@@ -326,6 +349,29 @@ CEED_QFUNCTION(Euler)(void *ctx, CeedInt Q,
                                     qdata[9][i]}
                                   };
     // *INDENT-ON*
+    // -- Grad-to-Grad qdata
+    // dU/dx
+    CeedScalar du[3][3] = {{0}};
+    CeedScalar drhodx[3] = {0};
+    CeedScalar dEdx[3] = {0};
+    CeedScalar dUdx[3][3] = {{0}};
+    CeedScalar dXdxdXdxT[3][3] = {{0}};
+    for (int j=0; j<3; j++) {
+      for (int k=0; k<3; k++) {
+        du[j][k] = (dU[j][k] - drho[k]*u[j]) / rho;
+        drhodx[j] += drho[k] * dXdx[k][j];
[…]
