### 2021-01-29 3d57682
headers - clearify includes to not rely on transitive includes (#701)
* headers - clearify includes to not rely on transitive includes

* style - add header recommendations from 'include-what-you-use'

* style - apply 'include-what-you-use' changes to CUDA backends

* style - 'include-what-you-use' for hip backends

* style - drop ceed.h includes in gallery qf source

* docs - add dev notes for header files

* style - header style and alphabetize
