+++ backends/cuda-gen/ceed-cuda-gen.c
@@ -14,9 +14,11 @@
 // software, applications, hardware, advanced system engineering and early
 // testbed platforms, in support of the nation's exascale computing imperative.
 
+#include <ceed.h>
+#include <ceed-backend.h>
 #include <string.h>
-#include <stdarg.h>
 #include "ceed-cuda-gen.h"
+#include "../cuda/ceed-cuda.h"
 
 //------------------------------------------------------------------------------
 // Backend init
