+++ backends/blocked/ceed-blocked.c
@@ -14,6 +14,10 @@
 // software, applications, hardware, advanced system engineering and early
 // testbed platforms, in support of the nation's exascale computing imperative.
 
+#include <ceed.h>
+#include <ceed-backend.h>
+#include <stdbool.h>
+#include <string.h>
 #include "ceed-blocked.h"
 
 //------------------------------------------------------------------------------
