+++ interface/ceed-fortran.c
@@ -16,10 +16,12 @@
 
 // Fortran interface
 #include <ceed.h>
-#include <ceed-impl.h>
 #include <ceed-backend.h>
+#include <ceed-impl.h>
 #include <ceed-fortran-name.h>
-#include <stdlib.h>
+#include <stdbool.h>
+#include <stdint.h>
+#include <stdio.h>
 #include <string.h>
 
 #define FORTRAN_REQUEST_IMMEDIATE -1
