+++ backends/cuda/ceed-cuda-basis.c
@@ -14,6 +14,10 @@
 // software, applications, hardware, advanced system engineering and early
 // testbed platforms, in support of the nation's exascale computing imperative.
 
+#include <ceed.h>
+#include <ceed-backend.h>
+#include <cuda.h>
+#include <cuda_runtime.h>
 #include "ceed-cuda.h"
 
 //------------------------------------------------------------------------------
