+++ examples/mfem/bp3.hpp
@@ -16,9 +16,9 @@
 
 /// @file
 /// Diffusion operator example using MFEM
+
 #include <ceed.h>
 #include <mfem.hpp>
-
 #include "bp3.h"
 
 /// Wrapper for a diffusion CeedOperator as an mfem::Operator
