+++ backends/blocked/ceed-blocked-operator.c
@@ -14,6 +14,11 @@
 // software, applications, hardware, advanced system engineering and early
 // testbed platforms, in support of the nation's exascale computing imperative.
 
+#include <ceed.h>
+#include <ceed-backend.h>
+#include <stdbool.h>
+#include <stddef.h>
+#include <stdint.h>
 #include "ceed-blocked.h"
 
 //------------------------------------------------------------------------------
