+++ backends/hip/ceed-hip-operator.c
@@ -14,10 +14,14 @@
 // software, applications, hardware, advanced system engineering and early
 // testbed platforms, in support of the nation's exascale computing imperative.
 
+#include <ceed.h>
+#include <ceed-backend.h>
+#include <hip/hip_runtime.h>
+#include <assert.h>
+#include <stdbool.h>
+#include <string.h>
 #include "ceed-hip.h"
 #include "ceed-hip-compile.h"
-#include <string.h>
-#include <math.h>
 
 //------------------------------------------------------------------------------
 // Destroy operator
