+++ backends/hip-shared/ceed-hip-shared-basis.c
@@ -14,7 +14,12 @@
 // software, applications, hardware, advanced system engineering and early
 // testbed platforms, in support of the nation's exascale computing imperative.
 
+#include <ceed.h>
+#include <ceed-backend.h>
+#include <hip/hip_runtime.h>
+#include <stddef.h>
 #include "ceed-hip-shared.h"
+#include "../hip/ceed-hip.h"
 #include "../hip/ceed-hip-compile.h"
 
 //------------------------------------------------------------------------------
