+++ examples/petsc/setup.h
@@ -17,19 +17,19 @@
 #ifndef setup_h
 #define setup_h
 
-#include <stdbool.h>
-#include <string.h>
+#include <ceed.h>
 #include <petsc.h>
 #include <petscdmplex.h>
 #include <petscfe.h>
-#include <ceed.h>
-#include "qfunctions/bps/common.h"
+#include <petscsys.h>
+#include <stdbool.h>
+#include <string.h>
 #include "qfunctions/bps/bp1.h"
 #include "qfunctions/bps/bp2.h"
 #include "qfunctions/bps/bp3.h"
 #include "qfunctions/bps/bp4.h"
+#include "qfunctions/bps/common.h"
 
-#include <petscsys.h>
 #if PETSC_VERSION_LT(3,12,0)
 #ifdef PETSC_HAVE_CUDA
 #include <petsccuda.h>
