+++ gallery/ceed-gallerytemplate.c
@@ -14,8 +14,9 @@
 // software, applications, hardware, advanced system engineering and early
 // testbed platforms, in support of the nation's exascale computing imperative.
 
+#include <ceed.h>
+#include <ceed-backend.h>
 #include <string.h>
-#include "ceed-backend.h"
 #include "ceed-gallerytemplate.h"
 
 /**
