### 2021-01-29 e48ac72
python - use success error code from C bindings
