+++ examples/fluids/euler-vortex.h
@@ -46,6 +46,7 @@ struct EulerContext_ {
   CeedScalar P_inlet;
   int stabilization;
   int euler_test;
+  bool implicit;
 };
 #endif
 
@@ -472,9 +473,9 @@ CEED_QFUNCTION(Euler)(void *ctx, CeedInt Q,
 
       for (int j=0; j<5; j++)
         for (int k=0; k<3; k++)
-          dv[k][j][i] -= stab[j][0] * dXdx[k][0] +
-                         stab[j][1] * dXdx[k][1] +
-                         stab[j][2] * dXdx[k][2];
+          dv[k][j][i] -= wdetJ*(stab[j][0] * dXdx[k][0] +
+                                stab[j][1] * dXdx[k][1] +
+                                stab[j][2] * dXdx[k][2]);
       break;
     case 2:        // SUPG is not implemented for explicit scheme
       break;
@@ -486,6 +487,232 @@ CEED_QFUNCTION(Euler)(void *ctx, CeedInt Q,
   return 0;
 }
 
+// *****************************************************************************
+CEED_QFUNCTION(IFunction_Euler)(void *ctx, CeedInt Q,
+                                const CeedScalar *const *in, CeedScalar *const *out) {
+  // *INDENT-OFF*
+  // Inputs
+  const CeedScalar (*q)[CEED_Q_VLA] = (const CeedScalar(*)[CEED_Q_VLA])in[0],
+                   (*dq)[5][CEED_Q_VLA] = (const CeedScalar(*)[5][CEED_Q_VLA])in[1],
+                   (*qdot)[CEED_Q_VLA] = (const CeedScalar(*)[CEED_Q_VLA])in[2],
+                   (*qdata)[CEED_Q_VLA] = (const CeedScalar(*)[CEED_Q_VLA])in[3],
+                   (*x)[CEED_Q_VLA] = (const CeedScalar(*)[CEED_Q_VLA])in[4];
+  // Outputs
+  CeedScalar (*v)[CEED_Q_VLA] = (CeedScalar(*)[CEED_Q_VLA])out[0],
+             (*dv)[5][CEED_Q_VLA] = (CeedScalar(*)[5][CEED_Q_VLA])out[1];
+  // Context
+  const EulerContext context = (EulerContext)ctx;
+  const CeedScalar currentTime = context->currentTime;
+  const int stabilization = context->stabilization;
+  const CeedScalar gamma  = 1.4;
+  const CeedScalar cv = 2.5;
+  const CeedScalar R = 1.;
+
+  CeedPragmaSIMD
+  // Quadrature Point Loop
+  for (CeedInt i=0; i<Q; i++) {
+    // *INDENT-OFF*
+    // Setup
+    // -- Interp in
+    const CeedScalar rho        =   q[0][i];
+    const CeedScalar u[3]       =  {q[1][i] / rho,
+                                    q[2][i] / rho,
+                                    q[3][i] / rho
+                                   };
+    const CeedScalar E          =   q[4][i];
+    // -- Grad in
+    const CeedScalar drho[3]    =  {dq[0][0][i],
+                                    dq[1][0][i],
+                                    dq[2][0][i]
+                                   };
+    const CeedScalar dU[3][3]   = {{dq[0][1][i],
+                                    dq[1][1][i],
+                                    dq[2][1][i]},
+                                   {dq[0][2][i],
+                                    dq[1][2][i],
+                                    dq[2][2][i]},
[…]
