### 2021-01-29 CEED/libCEED#700 @nbeams
I agree this is a tricky issue for a library like libCEED, which may need to use various task/execution queue models (through backends targeting different architecture), while also potentially being used by other libraries/users.

Currently, the CUDA/HIP/MAGMA backends all use the NULL stream (and only the NULL stream).  For the CUDA/HIP backends, this is baked into the kernel launch functions, see e.g. ` CeedRunKernelCuda`.  For the MAGMA backend, we do store a `magma_queue_t` in the backend data, just to avoid repeatedly creating/destroying the queue -- the queue is required for the MAGMA library calls used in the non-tensor basis routines -- but there's currently no way to set the queue for any stream other than the NULL stream.  (This is for two reasons: 1) there's no libCEED interface to tell the backend which stream you want, and 2) we want to use the same stream as the CUDA/HIP backends, since we delegate to them for QFunction and Operator actions).   So the addition of a stream/queue/command list object could potentially require some far-reaching refactoring across all the CUDA/HIP and MAGMA backends.  Not that that's a reason not to do it -- in fact, I think it could be nice to separate this somewhat from the backends themselves, rather than the implicit setup we have now across CUDA/HIP/MAGMA-- it's just something to keep in mind when envisioning how this setup could/should work.
