### 2021-01-30 e43605a
README: Added Euler Traveling Vortex
