### 2021-01-31 f7f9c5d
examples/fluids: bring back missing VecView function call
