### 2021-01-31 d7a15ae
examples/fluids: bring back missing VecView function call
