### 2021-01-31 2d1f40f
examples/fluids: remove unused context from SetUpDM()
