### 2021-01-31 05f5ecf
examples/fluids: remove unused context from SetUpDM()
