### 2021-01-31 38e1711
examples/fluids: removed vortex_strength from SetupContext
