+++ examples/fluids/advection2d.h
@@ -48,7 +48,6 @@ struct SetupContext_ {
   CeedScalar dc_axis[3];
   CeedScalar wind[3];
   CeedScalar time;
-  CeedScalar vortex_strength;
   int wind_type; // See WindType: 0=ROTATION, 1=TRANSLATION
 };
 #endif
