### 2021-02-01 1b7a89c
examples/fluids: fix broken advection(2d) problem
