### 2021-02-01 c01bdb9
examples/fluids: fix warning about the stab option usage
