### 2021-02-01 58a7032
examples/fluids: fix warning about the stab option usage
