### 2021-02-01 9aabbac
examples/fluids: remove T_inlet and P_inlet cl options and hard code them for each case
