### 2021-02-01 d71a1b5
examples/fluids: remove T_inlet and P_inlet cl options and hard code them for each case
