### 2021-02-01 fea9f2f
ETV: ke -> E_kinetic for consistency
