### 2021-02-02 3ba60be
ETV: some clean-up
