### 2021-02-03 827ffd1
examples/fluids: ETV - removed unused variables
