### 2021-02-03 7ec33f1
examples/fluids: ETV - more clean-up
