### 2021-02-03 ea4b449
examples/fluids: ETV - drop stabilization
