### 2021-02-03 a75e60d
examples/fluids: ETV - drop stabilization
