+++ examples/fluids/euler-vortex.h
@@ -42,7 +42,6 @@ struct EulerContext_ {
   CeedScalar currentTime;
   CeedScalar vortex_strength;
   CeedScalar etv_mean_velocity[3];
-  int stabilization;
   int euler_test;
   bool implicit;
 };
@@ -230,16 +229,6 @@ CEED_QFUNCTION(ICsEuler)(void *ctx, CeedInt Q,
 // Equation of State:
 //   P = (gamma - 1) (E - rho (u u) / 2)
 //
-// Stabilization:
-//   Tau = diag(TauC, TauM, TauM, TauM, TauE)
-//     f1 = rho  sqrt(ui uj gij)
-//     gij = dXi/dX * dXi/dX
-//     TauC = Cc f1 / (8 gii)
-//     TauM = min( 1 , 1 / f1 )
-//     TauE = TauM / (Ce cv)
-//
-//  SU   = Galerkin + grad(v) . ( Ai^T * Tau * (Aj q,j) )
-//
 // Constants:
 //   cv              ,  Specific heat, constant volume
 //   cp              ,  Specific heat, constant pressure
@@ -251,7 +240,6 @@ CEED_QFUNCTION(Euler)(void *ctx, CeedInt Q,
   // *INDENT-OFF*
   // Inputs
   const CeedScalar (*q)[CEED_Q_VLA] = (const CeedScalar(*)[CEED_Q_VLA])in[0],
-                   (*dq)[5][CEED_Q_VLA] = (const CeedScalar(*)[5][CEED_Q_VLA])in[1],
                    (*qdata)[CEED_Q_VLA] = (const CeedScalar(*)[CEED_Q_VLA])in[2];
   // Outputs
   CeedScalar (*v)[CEED_Q_VLA] = (CeedScalar(*)[CEED_Q_VLA])out[0],
@@ -259,7 +247,6 @@ CEED_QFUNCTION(Euler)(void *ctx, CeedInt Q,
   // Context
   const EulerContext context = (EulerContext)ctx;
   const CeedScalar currentTime = context->currentTime;
-  const int stabilization = context->stabilization;
   const CeedScalar gamma  = 1.4;
   const CeedScalar cv = 2.5;
   const CeedScalar R = 1.;
@@ -276,25 +263,7 @@ CEED_QFUNCTION(Euler)(void *ctx, CeedInt Q,
                                     q[3][i] / rho
                                    };
     const CeedScalar E          =   q[4][i];
-    // -- Grad in
-    const CeedScalar drho[3]    =  {dq[0][0][i],
-                                    dq[1][0][i],
-                                    dq[2][0][i]
-                                   };
-    const CeedScalar dU[3][3]   = {{dq[0][1][i],
-                                    dq[1][1][i],
-                                    dq[2][1][i]},
-                                   {dq[0][2][i],
-                                    dq[1][2][i],
-                                    dq[2][2][i]},
-                                   {dq[0][3][i],
-                                    dq[1][3][i],
-                                    dq[2][3][i]}
-                                  };
-    const CeedScalar dE[3]      =  {dq[0][4][i],
-                                    dq[1][4][i],
-                                    dq[2][4][i]
-                                   };
+
     // -- Interp-to-Interp qdata
     const CeedScalar wdetJ      =   qdata[0][i];
[…]
