+++ examples/fluids/conv_test.sh
@@ -20,8 +20,6 @@ declare -A run_flags
     run_flags[problem]=euler_vortex
     run_flags[degree]=2
     run_flags[dm_plex_box_faces]=20,20,1
-    run_flags[P_inlet]=1
-    run_flags[T_inlet]=1
     run_flags[euler_test]=none
     run_flags[lx]=1e3
     run_flags[ly]=1e3
