### 2021-02-03 76ff3eb
examples/fluids: ETV - a better conditional statement for incoming T and P
