### 2021-02-04 bcb2dfa
doc: replace recommonmark with MyST and migrate ReST to MyST
This offers more familiar syntax for most writing, as well as natural
$\pi$ math (versus :math:`\pi`) and markdown syntax for all ReST
features.

The API doc stubs remain in ReST as recommended for autodoc.

https://myst-parser.readthedocs.io/en/latest/sphinx/use.html#use-sphinx-ext-autodoc-in-markdown-files

The change was largely implemented using
https://github.com/executablebooks/rst-to-myst/ with some manual
fix-ups. Tables (both ReST built-in tables and list-table) remain under
eval-rst until they can be manually updated.
