+++ examples/solids/README.md
@@ -1,5 +1,4 @@
-libCEED: Solid Mechanics Example
---------------------------------
+# libCEED: Solid Mechanics Example
 
 This page provides a description of the solid mechanics example for the
 libCEED library, based on PETSc.
@@ -8,21 +7,25 @@ This code solves the steady-state static momentum balance equations using unstru
 In this mini-app, we consider three formulations used in solid mechanics applications: linear elasticity, Neo-Hookean hyperelasticity at small strain, and Neo-Hookean hyperelasticity at finite strain.
 All three of these formulations are for compressible materials.
 
-Build by using::
+Build by using:
 
-   make
+```
+make
+```
 
-and run with::
+and run with:
 
-   ./elasticity -mesh [.exo file] -degree [degree] -nu [nu] -E [E] [boundary options] -problem [problem type] -forcing [forcing] -ceed [ceed]
+```
+./elasticity -mesh [.exo file] -degree [degree] -nu [nu] -E [E] [boundary options] -problem [problem type] -forcing [forcing] -ceed [ceed]
+```
 
-Runtime options
----------------
+## Runtime options
 
-.. inclusion-solids-marker
+% inclusion-solids-marker
 
 The elasticity min-app is controlled via command-line options, the following of which are mandatory.
 
+```{eval-rst}
 .. list-table:: Mandatory Runtime Options
    :header-rows: 1
 
@@ -43,46 +46,50 @@ The elasticity min-app is controlled via command-line options, the following of 
 
    * - :code:`-bc_clamp [int list]`
      - List of face sets on which to displace by :code:`-bc_clamp_[facenumber]_translate [x,y,z]` and/or :code:`bc_clamp_[facenumber]_rotate [rx,ry,rz,c_0,c_1]`
-                                                                                                                   
+
+```
 
 Note: The default for a clamped face is zero displacement. All displacement is with respect to the initial configuration.
 
-   * - :code:`-bc_traction [int list]`
-     - List of face sets on which to set traction boundary conditions with the traction vector :code:`-bc_traction_[facenumber] [tx,ty,tz]`
+> - - {code}`-bc_traction [int list]`
+>   - List of face sets on which to set traction boundary conditions with the traction vector {code}`-bc_traction_[facenumber] [tx,ty,tz]`
 
-.. note::
-
-   This solver can use any mesh format that PETSc's ``DMPlex`` can read (Exodus, Gmsh, Med, etc.).
-   Our tests have primarily been using Exodus meshes created using CUBIT_; sample meshes used for the example runs suggested here can be found in `this repository`_.
-   Note that many mesh formats require PETSc to be configured appropriately; e.g., ``--download-exodusii`` for Exodus support.
-
-.. _CUBIT: https://cubit.sandia.gov/
-.. _this repository: https://github.com/jeremylt/ceedSampleMeshes
+:::{note}
+This solver can use any mesh format that PETSc's `DMPlex` can read (Exodus, Gmsh, Med, etc.).
+Our tests have primarily been using Exodus meshes created using [CUBIT]; sample meshes used for the example runs suggested here can be found in [this repository].
+Note that many mesh formats require PETSc to be configured appropriately; e.g., `--download-exodusii` for Exodus support.
+:::
 
 Consider the specific example of the mesh seen below:
 
-.. image:: https://github.com/jeremylt/ceedSampleMeshes/raw/master/cylinderDiagram.png
+```{image} https://github.com/jeremylt/ceedSampleMeshes/raw/master/cylinderDiagram.png
+```
 
[…]

[…]
