### 2021-02-05 #2945 @thelfer
@diehlpk A colleague and me started writting a review on the paper alone. I need to make this cleaner to share it with the authors. I still haven't installed libCEED and started reviewing the code, the install process and the examples. My agenda is a bit lighter in the coming weeks, so I hope to do that in the near future.
