### 2021-02-06 1a33ffc
tests: specify test name for ceed and nek examples and print the names when there are multiple test arguments
