### 2021-02-06 5e02960
examples/fluids: remove test_ from test names
