### 2021-02-09 17b112a
examples/navierstokes: Use heap memory for ctxSetup
