### 2021-02-09 f12d273
examples/navierstokes: Use heap memory for ctxSetup
