+++ examples/fluids/navierstokes.c
@@ -668,6 +668,7 @@ int main(int argc, char **argv) {
   TSAdapt adapt;
   User user;
   Units units;
+  SetupContext ctxSetupData;
   char ceedresource[4096] = "/cpu/self", problemName[] = "density_current";
   PetscFunctionList problems = NULL;
   PetscInt localNelemVol, lnodes, gnodes, steps;
@@ -747,6 +748,7 @@ int main(int argc, char **argv) {
   ierr = PetscMalloc1(1, &units); CHKERRQ(ierr);
   ierr = PetscCalloc1(1, &problem); CHKERRQ(ierr);
   ierr = PetscCalloc1(1, &bc); CHKERRQ(ierr);
+  ierr = PetscMalloc1(1, &ctxSetupData); CHKERRQ(ierr);
 
   // Register problems to be available on the command line
   ierr = PetscFunctionListAdd(&problems, "density_current", NS_DENSITY_CURRENT);
@@ -959,31 +961,29 @@ int main(int argc, char **argv) {
   const CeedInt dim = problem->dim, ncompx = problem->dim,
                 qdatasizeVol = problem->qdatasizeVol;
   // Set up the libCEED context
-  struct SetupContext_ ctxSetupData = {
-    .theta0 = theta0,
-    .thetaC = thetaC,
-    .P0 = P0,
-    .N = N,
-    .cv = cv,
-    .cp = cp,
-    .Rd = Rd,
-    .g = g,
-    .rc = rc,
-    .lx = lx,
-    .ly = ly,
-    .lz = lz,
-    .center[0] = center[0],
-    .center[1] = center[1],
-    .center[2] = center[2],
-    .dc_axis[0] = dc_axis[0],
-    .dc_axis[1] = dc_axis[1],
-    .dc_axis[2] = dc_axis[2],
-    .wind[0] = wind[0],
-    .wind[1] = wind[1],
-    .wind[2] = wind[2],
-    .time = 0,
-    .wind_type = wind_type,
-  };
+  ctxSetupData->theta0 = theta0;
+  ctxSetupData->thetaC = thetaC;
+  ctxSetupData->P0 = P0;
+  ctxSetupData->N = N;
+  ctxSetupData->cv = cv;
+  ctxSetupData->cp = cp;
+  ctxSetupData->Rd = Rd;
+  ctxSetupData->g = g;
+  ctxSetupData->rc = rc;
+  ctxSetupData->lx = lx;
+  ctxSetupData->ly = ly;
+  ctxSetupData->lz = lz;
+  ctxSetupData->center[0] = center[0];
+  ctxSetupData->center[1] = center[1];
+  ctxSetupData->center[2] = center[2];
+  ctxSetupData->dc_axis[0] = dc_axis[0];
+  ctxSetupData->dc_axis[1] = dc_axis[1];
+  ctxSetupData->dc_axis[2] = dc_axis[2];
+  ctxSetupData->wind[0] = wind[0];
+  ctxSetupData->wind[1] = wind[1];
+  ctxSetupData->wind[2] = wind[2];
+  ctxSetupData->time = 0;
+  ctxSetupData->wind_type = wind_type;
 
   // Create the mesh
   {
@@ -1011,7 +1011,7 @@ int main(int argc, char **argv) {
   // Setup DM
   ierr = DMLocalizeCoordinates(dm); CHKERRQ(ierr);
   ierr = DMSetFromOptions(dm); CHKERRQ(ierr);
-  ierr = SetUpDM(dm, problem, degree, bc, &ctxSetupData); CHKERRQ(ierr);
+  ierr = SetUpDM(dm, problem, degree, bc, ctxSetupData); CHKERRQ(ierr);
 
[…]
