+++ examples/solids/qfunctions/hyperFS.h
@@ -24,16 +24,47 @@
 #  include <math.h>
 #endif
 
+// -----------------------------------------------------------------------------
+// Neo-Hookean context
 #ifndef PHYSICS_STRUCT
 #define PHYSICS_STRUCT
 typedef struct Physics_private *Physics;
-struct Physics_private { // add parameters for MR/GP?
+
+struct Physics_private { 
   CeedScalar   nu;      // Poisson's ratio
   CeedScalar   E;       // Young's Modulus
 };
 #endif
 
+// -----------------------------------------------------------------------------
+// Mooney-Rivlin context
+#ifndef PHYSICS_STRUCT_MR
+#define PHYSICS_STRUCT_MR
+typedef struct Physics_private_MR *Physics_MR;
 
+struct Physics_private_MR { 
+  CeedScalar   nu;      // Poisson's ratio
+  CeedScalar   E;       // Young's Modulus
+  //TO-DO update with MR specific inputs
+};
+#endif
+
+// -----------------------------------------------------------------------------
+// Generalized Polynomial context
+#ifndef PHYSICS_STRUCT_GP
+#define PHYSICS_STRUCT_GP
+typedef struct Physics_private_GP *Physics_GP;
+
+struct Physics_private_GP { 
+  CeedScalar   nu;      // Poisson's ratio
+  CeedScalar   E;       // Young's Modulus
+  //TO-DO update with GP specific inputs
+};
+
+#endif
+// -----------------------------------------------------------------------------
+// end of contexts
+// -----------------------------------------------------------------------------
 
 // -----------------------------------------------------------------------------
 // Series approximation of log1p()
@@ -86,12 +117,53 @@ static inline CeedScalar computeDetCM1(CeedScalar E2work[6]) {
 // -----------------------------------------------------------------------------
 // -----------------------------------------------------------------------------
 // Neo-Hookean model
+
+CeedScalar NH_energyModel(void *ctx, CeedScalar logj, CeedScalar E2[][3]){
+  //requires unpacking the ctx struct again; 
+  const Physics context = (Physics)ctx;
+  const CeedScalar E  = context->E;
+  const CeedScalar nu = context->nu;
+  const CeedScalar TwoMu = E / (1 + nu);
+  const CeedScalar mu = TwoMu / 2;
+  const CeedScalar Kbulk = E / (3*(1 - 2*nu)); // Bulk Modulus
+  const CeedScalar lambda = (3*Kbulk - TwoMu) / 3;
+
+  return lambda*logj*logj/2. - mu*logj + mu*(E2[0][0] + E2[1][1] + E2[2][2])/2.;
+}
+
+
 // -----------------------------------------------------------------------------
 // Mooney-Rivlin model
+CeedScalar MR_energyModel(void *ctx, CeedScalar logj, CeedScalar E2[][3]){
+  //requires unpacking the ctx struct again; 
+  const Physics_MR context = (Physics_MR)ctx;
+  const CeedScalar E  = context->E;
+  const CeedScalar nu = context->nu;
+  const CeedScalar TwoMu = E / (1 + nu);
+  const CeedScalar mu = TwoMu / 2;
+  const CeedScalar Kbulk = E / (3*(1 - 2*nu)); // Bulk Modulus
+  const CeedScalar lambda = (3*Kbulk - TwoMu) / 3;
+  //TO-DO unpack remaining needed for MR
 
+  //TO-DO update with correct energy model
[…]
