+++ examples/fluids/navierstokes.c
@@ -820,22 +820,11 @@ int main(int argc, char **argv) {
                             "1 Kelvin in scaled temperature units",
                             NULL, Kelvin, &Kelvin, NULL); CHKERRQ(ierr);
   Kelvin = fabs(Kelvin);
-  ierr = PetscOptionsScalar("-theta0", "Reference potential temperature",
-                            NULL, theta0, &theta0, NULL); CHKERRQ(ierr);
-  ierr = PetscOptionsScalar("-thetaC", "Perturbation of potential temperature",
-                            NULL, thetaC, &thetaC, NULL); CHKERRQ(ierr);
-  ierr = PetscOptionsScalar("-P0", "Atmospheric pressure",
-                            NULL, P0, &P0, NULL); CHKERRQ(ierr);
+
+
   ierr = PetscOptionsScalar("-E_wind", "Total energy of inflow wind",
                             NULL, E_wind, &E_wind, NULL); CHKERRQ(ierr);
-  ierr = PetscOptionsScalar("-N", "Brunt-Vaisala frequency",
-                            NULL, N, &N, NULL); CHKERRQ(ierr);
-  ierr = PetscOptionsScalar("-cv", "Heat capacity at constant volume",
-                            NULL, cv, &cv, NULL); CHKERRQ(ierr);
-  ierr = PetscOptionsScalar("-cp", "Heat capacity at constant pressure",
-                            NULL, cp, &cp, NULL); CHKERRQ(ierr);
-  ierr = PetscOptionsScalar("-g", "Gravitational acceleration",
-                            NULL, g, &g, NULL); CHKERRQ(ierr);
+
   ierr = PetscOptionsScalar("-lambda",
                             "Stokes hypothesis second viscosity coefficient",
                             NULL, lambda, &lambda, NULL); CHKERRQ(ierr);
@@ -860,37 +849,15 @@ int main(int argc, char **argv) {
   //                     "Warning! Problem density_current does not support -CtauS or -strong_form\n");
   //  CHKERRQ(ierr);
   //}
-  ierr = PetscOptionsScalar("-lx", "Length scale in x direction",
-                            NULL, lx, &lx, NULL); CHKERRQ(ierr);
-  ierr = PetscOptionsScalar("-ly", "Length scale in y direction",
-                            NULL, ly, &ly, NULL); CHKERRQ(ierr);
-  ierr = PetscOptionsScalar("-lz", "Length scale in z direction",
-                            NULL, lz, &lz, NULL); CHKERRQ(ierr);
-  ierr = PetscOptionsScalar("-rc", "Characteristic radius of thermal bubble",
-                            NULL, rc, &rc, NULL); CHKERRQ(ierr);
+
   ierr = PetscOptionsScalar("-resx","Target resolution in x",
                             NULL, resx, &resx, NULL); CHKERRQ(ierr);
   ierr = PetscOptionsScalar("-resy","Target resolution in y",
                             NULL, resy, &resy, NULL); CHKERRQ(ierr);
   ierr = PetscOptionsScalar("-resz","Target resolution in z",
                             NULL, resz, &resz, NULL); CHKERRQ(ierr);
-  n = problem->dim;
-  center[0] = 0.5 * lx;
-  center[1] = 0.5 * ly;
-  center[2] = 0.5 * lz;
-  ierr = PetscOptionsRealArray("-center", "Location of bubble center",
[…]
