+++ examples/fluids/navierstokes.h
@@ -109,9 +109,9 @@ struct Units_ {
 };
 
 // Setup function for each problem
-extern PetscErrorCode NS_DENSITY_CURRENT(problemData *problem);
-extern PetscErrorCode NS_ADVECTION(problemData *problem);
-extern PetscErrorCode NS_ADVECTION2D(problemData *problem);
+extern PetscErrorCode NS_DENSITY_CURRENT(problemData *problem, void **ctxSetupData);
+extern PetscErrorCode NS_ADVECTION(problemData *problem, void **ctxSetupData);
+extern PetscErrorCode NS_ADVECTION2D(problemData *problem, void **ctxSetupData);
 
 // Boundary Condition Functions
 extern PetscErrorCode BC_DENSITY_CURRENT(DM dm, SimpleBC bc,
