### 2021-02-10 3b61df9
examples/fluids: Some clean-up in the post-processing files
Only report the failure of the convergence test if it fails.
