### 2021-02-10 19c4619
examples/fluids: Some clean-up in the post-processing files
Only report the failure of the convergence test if it fails.
