+++ examples/fluids/conv_plot.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+
 # Copyright (c) 2017-2018, Lawrence Livermore National Security, LLC.
 # Produced at the Lawrence Livermore National Laboratory. LLNL-CODE-734707.
 # All Rights reserved. See files LICENSE and NOTICE for details.
@@ -15,7 +16,6 @@
 # software, applications, hardware, advanced system engineering and early
 # testbed platforms, in support of the nation's exascale computing imperative.
 
-import numpy as np
 import pandas as pd
 import argparse
 from pylab import *
@@ -25,7 +25,7 @@ from matplotlib import use
 def plot():
     # Define argparse for the input variables
     parser = argparse.ArgumentParser(description='Get input arguments')
-    parser.add_argument('--conv_result_file',
+    parser.add_argument('-f',
                         dest='conv_result_file',
                         type=str,
                         required=True,
@@ -39,31 +39,26 @@ def plot():
               'gray', 'blue', 'purple', 'pink', 'black']
     res = 'mesh_res'
     fig, ax = plt.subplots()
-    # Arbitrary coefficients
-    C = [2.2e-2, .24e0, .22e0, .7e0, 2.5e0,
-        3e0, 3.5e0, 4e0, 4.5e0, 5e0]
+
     i = 0
     for group in runs.groupby('degree'):
         data = group[1]
         data = data.sort_values('rel_error')
         p = data['degree'].values[0]
         h = 1/data[res]
-        H = C[i] * h**p # H = C h^p
         E = data['rel_error']
-        log_h = np.log10(h)
-        log_H = np.log10(H)
+        h_min = amin(h)
+        E_min = amin(E)
+        H =  E_min * (h/h_min)**p
         ax.loglog(h, E, 'o', color=colors[i])
-        m, b = np.polyfit(log_h, log_H, 1)
-        ax.loglog(h, 10**b * h**m, '--', color=colors[i], label='O(h^' + str(p) + ')')
+        ax.loglog(h, H, '--', color=colors[i], label='O(h$^' + str(p) + '$)')
         i = i + 1
 
     ax.legend(loc='best')
     ax.set_xlabel('h')
     ax.set_ylabel('Relative Error')
     ax.set_title('Convergence by h Refinement')
-    xlim(.03, .3)
-    fig.tight_layout()
-    plt.savefig('conv_plt_h.png', bbox_inches='tight')
+    plt.savefig('conv_plt_h.png')
 
 
 if __name__ == "__main__":
