### 2021-02-13 075be03
tests/junit.py: Add name for tests with multiple test cases
