### 2021-02-17 e8f0441
working on MR and GP frame work. added new physics contexts and energy functions.
