### 2021-02-17 dba3c4e
adding new energy functions for MR
