+++ examples/solids/elasticity.c
@@ -277,6 +277,7 @@ int main(int argc, char **argv) {
                                  ncompu, Ugsz[fineLevel], Ulocsz[fineLevel],
                                  forceCeed); //TO-DO
                                  forceCeed, neumannCeed);
+                                 forceCeed); //TO-DO
     CHKERRQ(ierr);
   }
   // ---- Setup coarse Jacobian evaluator and prolongation/restriction
