### 2021-02-17 f905960
adding new energy functions for MR
