+++ examples/solids/elasticity.c
@@ -290,6 +290,7 @@ int main(int argc, char **argv) {
                                  ncompu, Ugsz[fineLevel], Ulocsz[fineLevel],
                                  forceCeed);
                                  forceCeed, neumannCeed);
+                                 forceCeed); //TO-DO
     CHKERRQ(ierr);
   }
   // ---- Setup coarse Jacobian evaluator and prolongation/restriction
