+++ examples/solids/elasticity.h
@@ -48,8 +48,16 @@ typedef enum {
 static const char *const problemTypes[] = {"linElas",
                                            "hyperSS",
                                            "hyperFS",
+<<<<<<< HEAD
                                            "hyperFS-MR",
+<<<<<<< HEAD
                                            "hyperFS-GP","problemType","ELAS_",0 //TO-DO update with new options; will need to change the structs depending on input
+=======
+                                           "hyperFS-GP",
+=======
+>>>>>>> 1bb95495... adding new energy functions for MR
+                                           "problemType","ELAS_",0 //TO-DO update with new options; will need to change the structs depending on input
+>>>>>>> dbac515c... adding new energy functions for MR
                                           };
 static const char *const problemTypesForDisp[] = {"Linear elasticity",
                                                   "Hyper elasticity small strain",
