+++ examples/solids/elasticity.h
@@ -49,10 +49,15 @@ static const char *const problemTypes[] = {"linElas",
                                            "hyperSS",
                                            "hyperFS",
 <<<<<<< HEAD
+<<<<<<< HEAD
                                            "hyperFS-MR",
                                            "hyperFS-GP",
 =======
 >>>>>>> 1bb95495... adding new energy functions for MR
+=======
+                                           "hyperFS-MR",
+                                           "hyperFS-GP",
+>>>>>>> 8fc4caa8... fising rebase conflicts
                                            "problemType","ELAS_",0 //TO-DO update with new options; will need to change the structs depending on input
                                           };
 static const char *const problemTypesForDisp[] = {"Linear elasticity",
