+++ examples/solids/src/setuplibceed.c
@@ -180,62 +180,185 @@ PetscErrorCode CeedDataDestroy(CeedInt level, CeedData data) {
   PetscFunctionReturn(0);
 };
 
-// Get libCEED restriction data from DMPlex
-PetscErrorCode CreateRestrictionPlex(Ceed ceed, CeedInt P, CeedInt ncomp,
-                                     CeedElemRestriction *Erestrict, DM dm) {
-  PetscInt ierr;
-  PetscInt c, cStart, cEnd, nelem, nnodes, *erestrict, eoffset;
+// Utility function - essential BC dofs are encoded in closure indices as -(i+1)
+PetscInt Involute(PetscInt i) {
+  return i >= 0 ? i : -(i + 1);
+};
+
+// Utility function to create local CEED restriction from DMPlex
+PetscErrorCode CreateRestrictionFromPlex(Ceed ceed, DM dm, CeedInt P,
+    CeedInt height, DMLabel domainLabel, CeedInt value,
+    CeedElemRestriction *Erestrict) {
+
   PetscSection section;
+  PetscInt p, Nelem, Ndof, *erestrict, eoffset, nfields, dim, depth;
+  DMLabel depthLabel;
+  IS depthIS, iterIS;
   Vec Uloc;
+  const PetscInt *iterIndices;
+  PetscErrorCode ierr;
 
   PetscFunctionBeginUser;
 
-  // Get Nelem
-  ierr = DMGetSection(dm, &section); CHKERRQ(ierr);
-  ierr = DMPlexGetHeightStratum(dm, 0, &cStart, &cEnd); CHKERRQ(ierr);
-  nelem = cEnd - cStart;
+  ierr = DMGetDimension(dm, &dim); CHKERRQ(ierr);
+  dim -= height;
+  ierr = DMGetLocalSection(dm, &section); CHKERRQ(ierr);
+  ierr = PetscSectionGetNumFields(section, &nfields); CHKERRQ(ierr);
+  PetscInt ncomp[nfields], fieldoff[nfields+1];
+  fieldoff[0] = 0;
+  for (PetscInt f = 0; f < nfields; f++) {
+    ierr = PetscSectionGetFieldComponents(section, f, &ncomp[f]); CHKERRQ(ierr);
+    fieldoff[f+1] = fieldoff[f] + ncomp[f];
+  }
 
-  // Get indices
-  ierr = PetscMalloc1(nelem*P*P*P, &erestrict); CHKERRQ(ierr);
-  for (c=cStart, eoffset=0; c<cEnd; c++) {
-    PetscInt numindices, *indices, i;
+  ierr = DMPlexGetDepth(dm, &depth); CHKERRQ(ierr);
+  ierr = DMPlexGetDepthLabel(dm, &depthLabel); CHKERRQ(ierr);
+  ierr = DMLabelGetStratumIS(depthLabel, depth - height, &depthIS); CHKERRQ(ierr);
+  if (domainLabel) {
+    IS domainIS;
+    ierr = DMLabelGetStratumIS(domainLabel, value, &domainIS); CHKERRQ(ierr);
+    if (domainIS) { // domainIS is non-empty
+      ierr = ISIntersect(depthIS, domainIS, &iterIS); CHKERRQ(ierr);
+      ierr = ISDestroy(&domainIS); CHKERRQ(ierr);
+    } else { // domainIS is NULL (empty)
+      iterIS = NULL;
+    }
+    ierr = ISDestroy(&depthIS); CHKERRQ(ierr);
+  } else {
+    iterIS = depthIS;
+  }
+  if (iterIS) {
[…]
