### 2021-02-24 72d03b6
WIP: Solids-GivensBC (#702)
* Givens boundary condition added.

Co-authored with: @jedbrown and  @jeremylt

* Comments first letter capitalized
