### 2021-02-24 b5274dc
gitlab - display number processes avaliable for tests
