+++ .gitlab-ci.yml
@@ -9,7 +9,7 @@ noether-rocm:
   script:
 # Compilers
     - export COVERAGE=1 CC=gcc CXX=g++ FC=gfortran HIPCC=hipcc
-    - echo "---------- nproc-----------" && export NPROC_CPU $(nproc) && export NPROC_GPU 8
+    - echo "---------- nproc-----------" && NPROC_CPU=$(nproc) && NPROC_GPU=8
     - echo "---------- CC -------------" && $CC --version
     - echo "---------- CXX ------------" && $CXX --version
     - echo "---------- FC -------------" && $FC --version
