### 2021-03-03 bd05711
add CUDA_ARCH and HIP_ARCH options
