### 2021-03-03 5cb368c
add launch bounds to magma kernels; add macro definition for y-dim of magma basis kernel threadblocks
Co-authored-by: Ahmad Abdelfattah <ahmad@icl.utk.edu>
