+++ backends/magma/ceed-magma-basis.c
@@ -104,7 +104,7 @@ int CeedBasisApply_Magma(CeedBasis basis, CeedInt nelem,
                         impl->dinterp1d, tmode,
                         u, u_elstride, u_compstride,
                         v, v_elstride, v_compstride,
-                        nelem, data->basis_kernel_mode, data->maxthreads,
+                        nelem, data->basis_kernel_mode,
                         data->queue);
     if (ierr != 0) CeedError(ceed, CEED_ERROR_BACKEND,
                                "MAGMA: launch failure detected for magma_interp");
@@ -162,7 +162,7 @@ int CeedBasisApply_Magma(CeedBasis basis, CeedInt nelem,
                        impl->dinterp1d, impl->dgrad1d, tmode,
                        u, u_elstride, u_compstride, u_dimstride,
                        v, v_elstride, v_compstride, v_dimstride,
-                       nelem, data->basis_kernel_mode, data->maxthreads,
+                       nelem, data->basis_kernel_mode,
                        data->queue);
     if (ierr != 0) CeedError(ceed, CEED_ERROR_BACKEND,
                                "MAGMA: launch failure detected for magma_grad");
@@ -177,7 +177,7 @@ int CeedBasisApply_Magma(CeedBasis basis, CeedInt nelem,
     CeedInt Q = Q1d;
     int eldofssize = CeedIntPow(Q, dim);
     ierr = magma_weight(Q, dim, impl->dqweight1d, v, eldofssize, nelem,
-                        data->basis_kernel_mode, data->maxthreads, data->queue);
+                        data->basis_kernel_mode, data->queue);
     if (ierr != 0) CeedError(ceed, CEED_ERROR_BACKEND,
                                "MAGMA: launch failure detected for magma_weight");
   }
