+++ backends/magma/ceed-magma.c
@@ -52,11 +52,6 @@ static int CeedInit_Magma(const char *resource, Ceed ceed) {
   // kernel selection
   data->basis_kernel_mode = MAGMA_KERNEL_DIM_SPECIFIC;
 
-  // kernel max threads per thread-block
-  data->maxthreads[0] = 128;  // for 1D kernels
-  data->maxthreads[1] = 128;  // for 2D kernels
-  data->maxthreads[2] =  64;  // for 3D kernels
-
   // get/set device ID
   const char *device_spec = strstr(resource, ":device_id=");
   const int deviceID = (device_spec) ? atoi(device_spec+11) : -1;

[…]
