### 2021-03-03 52fd280
add CUDA_ARCH and HIP_ARCH to CONFIG_VARS
