### 2021-03-04 d24be11
examples/ceed/ex1-volume.c: make the naming of the tests consistent
