### 2021-03-04 657ae46
tests/junit.py: a more compact if statement
helped-by: @jeremylt
