+++ examples/solids/qfunctions/hyperFS.h
@@ -62,8 +62,8 @@ struct Physics_private_GP {
   CeedScalar   nu;      // Poisson's ratio rm
   CeedScalar   E;       // Young's Modulus rm
   //material properties for GP
-  CeedScalar C_mat; // 2D matrix
-  CeedScalar K; // 1D array
+  CeedScalar C_mat[6][6]; // 2D matrix
+  CeedScalar K[6]; // 1D array
   CeedScalar N; // max value of the sum; usually 1 or 2
 };
 
@@ -124,7 +124,7 @@ static inline CeedScalar computeDetCM1(CeedScalar E2work[6]) {
 // -----------------------------------------------------------------------------
 // Neo-Hookean model
 
-CeedScalar NH_energyModel(void *ctx, CeedScalar logj, CeedScalar E2[][3]){
+CeedScalar NH_energyModel(void *ctx, const CeedScalar detC_m1, CeedScalar E2[][3], CeedScalar wdetJ){
   //requires unpacking the ctx struct again; 
   const Physics context = (Physics)ctx;
   const CeedScalar E  = context->E;
@@ -133,92 +133,80 @@ CeedScalar NH_energyModel(void *ctx, CeedScalar logj, CeedScalar E2[][3]){
   const CeedScalar mu = TwoMu / 2;
   const CeedScalar Kbulk = E / (3*(1 - 2*nu)); // Bulk Modulus
   const CeedScalar lambda = (3*Kbulk - TwoMu) / 3;
+  
+  CeedScalar logj = log1p_series_shifted(detC_m1)/2.;
 
-  return lambda*logj*logj/2. - mu*logj + mu*(E2[0][0] + E2[1][1] + E2[2][2])/2.;
+  return (lambda*logj*logj/2. - mu*logj + mu*(E2[0][0] + E2[1][1] + E2[2][2])/2.)* wdetJ;
 }
 
 // -----------------------------------------------------------------------------
 // Mooney-Rivlin model
-CeedScalar MR_energyModel(void *ctx, CeedScalar logj, CeedScalar E2[][3]){
+CeedScalar MR_energyModel(void *ctx, CeedScalar detC_m1, CeedScalar E2[][3], CeedScalar wdetJ){
   //requires unpacking the ctx struct again; 
   const Physics_MR context = (Physics_MR)ctx;
-  const CeedScalar E  = context->E;
-  const CeedScalar nu = context->nu;
-  const CeedScalar TwoMu = E / (1 + nu);
-  const CeedScalar mu = TwoMu / 2;
-  const CeedScalar Kbulk = E / (3*(1 - 2*nu)); // Bulk Modulus
-  const CeedScalar lambda = (3*Kbulk - TwoMu) / 3;
+  // const CeedScalar E  = context->E;
+  // const CeedScalar nu = context->nu;
+  // const CeedScalar TwoMu = E / (1 + nu);
+  // const CeedScalar mu = TwoMu / 2;
+  // const CeedScalar Kbulk = E / (3*(1 - 2*nu)); // Bulk Modulus
+  // const CeedScalar lambda = (3*Kbulk - TwoMu) / 3;
   //TO-DO unpack remaining needed for MR
   const CeedScalar mu_1 = context -> mu_1; // material constant mu_1
   const CeedScalar mu_2 = context -> mu_2; // material constant mu_2
   const CeedScalar k_1 = context -> k_1; // material constant k_1
 
-  //TO-DO update with correct energy model:
+  //energy model:
+  const CeedScalar C[3][3] = {{1 + E2[0][0], E2[0][1], E2[0][2]},
+                              {E2[0][1], 1 + E2[1][1], E2[1][2]},
[…]
