### 2021-03-09 660e55a
WIP: Trying to assign values to context data in each problem function
