### 2021-03-09 29c4482
WIP: Trying to assign values to context data in each problem function
