+++ examples/fluids/navierstokes.c
@@ -669,6 +669,7 @@ int main(int argc, char **argv) {
   User user;
   Units units;
   SetupContext ctxSetupData;
+  NSContext ctxNSData;
   char ceedresource[4096] = "/cpu/self", problemName[] = "density_current";
   PetscFunctionList problems = NULL;
   PetscInt localNelemVol, lnodes, gnodes, steps;
@@ -749,6 +750,7 @@ int main(int argc, char **argv) {
   ierr = PetscCalloc1(1, &problem); CHKERRQ(ierr);
   ierr = PetscCalloc1(1, &bc); CHKERRQ(ierr);
   ierr = PetscMalloc1(1, &ctxSetupData); CHKERRQ(ierr);
+  ierr = PetscMalloc1(1, &ctxNSData); CHKERRQ(ierr);
 
   // Register problems to be available on the command line
   ierr = PetscFunctionListAdd(&problems, "density_current", NS_DENSITY_CURRENT);
@@ -817,17 +819,16 @@ int main(int argc, char **argv) {
                             NULL, Kelvin, &Kelvin, NULL); CHKERRQ(ierr);
   Kelvin = fabs(Kelvin);
 
-
   ierr = PetscOptionsScalar("-E_wind", "Total energy of inflow wind",
                             NULL, E_wind, &E_wind, NULL); CHKERRQ(ierr);
 
-  ierr = PetscOptionsScalar("-lambda",
-                            "Stokes hypothesis second viscosity coefficient",
-                            NULL, lambda, &lambda, NULL); CHKERRQ(ierr);
-  ierr = PetscOptionsScalar("-mu", "Shear dynamic viscosity coefficient",
-                            NULL, mu, &mu, NULL); CHKERRQ(ierr);
-  ierr = PetscOptionsScalar("-k", "Thermal conductivity",
-                            NULL, k, &k, NULL); CHKERRQ(ierr);
+  //ierr = PetscOptionsScalar("-lambda",
+  //                          "Stokes hypothesis second viscosity coefficient",
+  //                          NULL, lambda, &lambda, NULL); CHKERRQ(ierr);
+  //ierr = PetscOptionsScalar("-mu", "Shear dynamic viscosity coefficient",
+  //                          NULL, mu, &mu, NULL); CHKERRQ(ierr);
+  //ierr = PetscOptionsScalar("-k", "Thermal conductivity",
+  //                          NULL, k, &k, NULL); CHKERRQ(ierr);
   ierr = PetscOptionsScalar("-CtauS",
                             "Scale coefficient for tau (nondimensional)",
                             NULL, CtauS, &CtauS, NULL); CHKERRQ(ierr);
@@ -883,46 +884,45 @@ int main(int argc, char **argv) {
 
   {
     // Choose the problem from the list of registered problems
-    PetscErrorCode (*p)(problemData *, void **);
+    PetscErrorCode (*p)(problemData *, void **, void **);
     ierr = PetscFunctionListFind(problems, problemName, &p); CHKERRQ(ierr);
     if (!p) SETERRQ1(PETSC_COMM_SELF, 1, "Problem '%s' not found", problemName);
-    ierr = (*p)(problem, &ctxSetupData); CHKERRQ(ierr);
+    ierr = (*p)(problem, &ctxSetupData, &units); CHKERRQ(ierr);
   }
 
   // Set up the libCEED context
   ctxSetupData->time = 0;
 
   // Define derived units
[…]
