+++ examples/fluids/navierstokes.h
@@ -110,9 +110,11 @@ struct Units_ {
 
 // Setup function for each problem
 extern PetscErrorCode NS_DENSITY_CURRENT(problemData *problem,
-    void **ctxSetupData);
-extern PetscErrorCode NS_ADVECTION(problemData *problem, void **ctxSetupData);
-extern PetscErrorCode NS_ADVECTION2D(problemData *problem, void **ctxSetupData);
+    void **ctxSetupData, void **ctx);
+extern PetscErrorCode NS_ADVECTION(problemData *problem,
+    void **ctxSetupData, void *ctx);
+extern PetscErrorCode NS_ADVECTION2D(problemData *problem,
+    void **ctxSetupData, void *ctx);
 
 // Boundary Condition Functions
 extern PetscErrorCode BC_DENSITY_CURRENT(DM dm, SimpleBC bc,
