### 2021-03-09 abdd747
examples/solids: Add residual and Jacobian in current configuration
There are some benefits to evaluating residuals and/or Jacobians in
current configuration. This includes simpler representation of body
forces and contact, as well as lower storage and lower arithmetic in the
Jacobian action. We add these implementations as well documentation to
derive their representations via linearize-then-pushforward and
pushforward-then-linearize.

Co-authored-by: Arash Mehraban <arashm81@gmail.com>
Co-authored-by: Jed Brown <jed@jedbrown.org>
Co-authored-by: Jeremy L. Thompson <jeremy.thompson@colorado.e
[…]
