+++ doc/sphinx/source/references.bib
@@ -99,3 +99,14 @@
   publisher={ACM New York, NY, USA},
   doi={10.1145/1498765.1498785}
 }
+
+@article{davydov2020matrix,
+  title={A matrix-free approach for finite-strain hyperelastic problems using geometric multigrid},
+  author={Davydov, Denis and Pelteret, Jean-Paul and Arndt, Daniel and Kronbichler, Martin and Steinmann, Paul},
+  journal={International Journal for Numerical Methods in Engineering},
+  volume={121},
+  number={13},
+  pages={2874--2895},
+  year={2020},
+  publisher={Wiley Online Library}
+}

[…]
