### 2021-03-09 5d6e984
added no body force with new HyperFScur
