### 2021-03-09 878690b
added necessary constituents for the new QFunction
