### 2021-03-09 e648102
new QFunction with tesnor b & tau
