### 2021-03-09 59f499b
bug: there is a problem in pointer to struct
