+++ examples/fluids/navierstokes.h
@@ -108,9 +108,16 @@ struct Units_ {
   PetscScalar Joule;
 };
 
+// Setup Context for QFunctions
+typedef struct QFContext_ *QFContext;
+struct QFContext_ {
+  NSContext ctxNSData;
+  AdvectionContext ctxAdvectionData;
+};
+
 // Setup function for each problem
 extern PetscErrorCode NS_DENSITY_CURRENT(problemData *problem,
-    void **ctxSetupData, void **ctx);
+    void **ctxSetupData, void **ctx, void **ctxQF);
 extern PetscErrorCode NS_ADVECTION(problemData *problem,
     void **ctxSetupData, void *ctx);
 extern PetscErrorCode NS_ADVECTION2D(problemData *problem,
