### 2021-03-09 3dc4012
bug: there is a problem in pointer to struct
