### 2021-03-09 495dc06
doc/solids: Add doc for current configuration
