### 2021-03-09 568e5c5
SWE: fixed the double prototyping
