+++ examples/solids/qfunctions/hyperFSCUR.h
@@ -79,42 +79,29 @@ static inline CeedScalar computeDetCM1cur(CeedScalar E2work[6]) {
          E2work[4]*E2work[4] - E2work[3]*E2work[3];
 };
 
-
-// -----------------------------------------------------------------------------
-// Compute det F
-// -----------------------------------------------------------------------------
-static inline CeedScalar computeDetF(CeedScalar F[3][3]) {
-  CeedScalar a1 = F[0][0] * F[1][1] * F[2][2]; 
-  CeedScalar a2 = F[0][1] * F[1][2] * F[2][0]; 
-  CeedScalar a3 = F[0][2] * F[1][0] * F[2][1]; 
-  CeedScalar a4 = F[0][2] * F[1][1] * F[2][0]; 
-  CeedScalar a5 = F[0][0] * F[1][2] * F[2][1]; 
-  CeedScalar a6 = F[0][1] * F[1][0] * F[2][2];
-  
-  return (a1+a2+a3) - (a4+a5+a6); 
-};
-
 // -----------------------------------------------------------------------------
 // Common computations between Ftau and dFtau
 // -----------------------------------------------------------------------------
 static inline int commonFtau(const CeedScalar lambda, const CeedScalar mu,
-                           const CeedScalar gradu[3][3], CeedScalar F_inv[3][3], 
-			   CeedScalar tau_work[6], CeedScalar b_invwork[6], 
-			   CeedScalar *J, CeedScalar *llnj) {
+                             const CeedScalar gradu[3][3], CeedScalar F_inv[3][3],
+                             CeedScalar tau_work[6], CeedScalar b_invwork[6],
+                             CeedScalar *J, CeedScalar *llnj) {
 
-  
+
   // Compute The Deformation Gradient : F = I3 + gradu
-  CeedScalar F[3][3] =  {{gradu[0][0] + 1,
-                          gradu[0][1],
-                          gradu[0][2]},
-                         {gradu[1][0],
-                          gradu[1][1] + 1,
-                          gradu[1][2]},
-                         {gradu[2][0],
-                          gradu[2][1],
-                          gradu[2][2] + 1}
+  // *INDENT-OFF*
+  CeedScalar F[3][3] =  {{gradu[0][0][i] + 1,
+                          gradu[0][1][i],
+                          gradu[0][2][i]},
+                         {gradu[1][0][i],
+                          gradu[1][1][i] + 1,
+                          gradu[1][2][i]},
+                         {gradu[2][0][i],
+                          gradu[2][1][i],
+                          gradu[2][2][i] + 1}
                         };
-  // b - left Cauchy-Green 
+  // *INDENT-ON*
+  // b - left Cauchy-Green
   // b =  F*F^T
   CeedScalar b[3][3];
   for (CeedInt j = 0; j < 3; j++)
@@ -123,13 +110,26 @@ static inline int commonFtau(const CeedScalar lambda, const CeedScalar mu,
       for (CeedInt m = 0; m < 3; m++)
         b[j][k] += F[j][m] * F[k][m]; // F * F^T
     }
-    
+
+  // E - Green-Lagrange strain tensor
+  //     E = 1/2 (gradu + gradu^T + gradu^T*gradu)
+  const CeedInt indj[6] = {0, 1, 2, 1, 0, 0}, indk[6] = {0, 1, 2, 2, 2, 1};
+  CeedScalar E2work[6];
+  for (CeedInt m = 0; m < 6; m++) {
[…]
