### 2021-03-10 365f29b
avoid to compute log1p twice in commonFtau function
