### 2021-03-10 fbd5dd4
removed llnj and logJ from commonFtau output
