### 2021-03-10 34858c3
examples/fluids/conv_plot.py: clean-up
