### 2021-03-10 bd52b23
comment out stuff that have been moved to densitycurrent.c
