### 2021-03-10 0382d39
comment out stuff that have been moved to densitycurrent.c
