+++ examples/fluids/navierstokes.c
@@ -687,8 +687,8 @@ int main(int argc, char **argv) {
   CeedOperator op_setupVol, op_ics;
   CeedScalar Rd;
   CeedMemType memtyperequested;
-  PetscScalar WpermK, Pascal, JperkgK, mpersquareds, kgpercubicm,
-              kgpersquaredms, Joulepercubicm, Joule;
+  //PetscScalar WpermK, Pascal, JperkgK, mpersquareds, kgpercubicm,
+  //            kgpersquaredms, Joulepercubicm, Joule;
   problemData *problem = NULL;
   WindType wind_type;
   StabilizationType stab;
@@ -709,37 +709,37 @@ int main(int argc, char **argv) {
   // *INDENT-ON*
 
   // Create the libCEED contexts
-  PetscScalar meter      = 1e-2;     // 1 meter in scaled length units
-  PetscScalar second     = 1e-2;     // 1 second in scaled time units
-  PetscScalar kilogram   = 1e-6;     // 1 kilogram in scaled mass units
-  PetscScalar Kelvin     = 1;        // 1 Kelvin in scaled temperature units
-  CeedScalar theta0      = 300.;     // K
-  CeedScalar thetaC      = -15.;     // K
-  CeedScalar P0          = 1.e5;     // Pa
+  //PetscScalar meter      = 1e-2;     // 1 meter in scaled length units
+  //PetscScalar second     = 1e-2;     // 1 second in scaled time units
+  //PetscScalar kilogram   = 1e-6;     // 1 kilogram in scaled mass units
+  //PetscScalar Kelvin     = 1;        // 1 Kelvin in scaled temperature units
+  //CeedScalar theta0      = 300.;     // K
+  //CeedScalar thetaC      = -15.;     // K
+  //CeedScalar P0          = 1.e5;     // Pa
   CeedScalar E_wind      = 1.e6;     // J
-  CeedScalar N           = 0.01;     // 1/s
-  CeedScalar cv          = 717.;     // J/(kg K)
-  CeedScalar cp          = 1004.;    // J/(kg K)
-  CeedScalar g           = 9.81;     // m/s^2
-  CeedScalar lambda      = -2./3.;   // -
-  CeedScalar mu          = 75.;      // Pa s, dynamic viscosity
+  //CeedScalar N           = 0.01;     // 1/s
+  //CeedScalar cv          = 717.;     // J/(kg K)
+  //CeedScalar cp          = 1004.;    // J/(kg K)
+  //CeedScalar g           = 9.81;     // m/s^2
+  //CeedScalar lambda      = -2./3.;   // -
+  //CeedScalar mu          = 75.;      // Pa s, dynamic viscosity
   // mu = 75 is not physical for air, but is good for numerical stability
-  CeedScalar k           = 0.02638;  // W/(m K)
+  //CeedScalar k           = 0.02638;  // W/(m K)
   CeedScalar CtauS       = 0.;       // dimensionless
   CeedScalar strong_form = 0.;       // [0,1]
-  PetscScalar lx         = 8000.;    // m
-  PetscScalar ly         = 8000.;    // m
-  PetscScalar lz         = 4000.;    // m
-  CeedScalar rc          = 1000.;    // m (Radius of bubble)
-  PetscScalar resx       = 1000.;    // m (resolution in x)
-  PetscScalar resy       = 1000.;    // m (resolution in y)
-  PetscScalar resz       = 1000.;    // m (resolution in z)
+  //PetscScalar lx         = 8000.;    // m
+  //PetscScalar ly         = 8000.;    // m
+  //PetscScalar lz         = 4000.;    // m
+  //CeedScalar rc          = 1000.;    // m (Radius of bubble)
[…]
