### 2021-03-10 e377c63
WIP: The first two regression test pass. There is a problem with stabilization.
