### 2021-03-10 41cd400
WIP: The first two regression test pass. There is a problem with stabilization.
