+++ examples/fluids/navierstokes.h
@@ -119,9 +119,9 @@ struct QFContext_ {
 extern PetscErrorCode NS_DENSITY_CURRENT(problemData *problem,
     void **ctxSetupData, void **ctx, void **ctxQF);
 extern PetscErrorCode NS_ADVECTION(problemData *problem,
-    void **ctxSetupData, void *ctx);
+                                   void **ctxSetupData, void *ctx);
 extern PetscErrorCode NS_ADVECTION2D(problemData *problem,
-    void **ctxSetupData, void *ctx);
+                                     void **ctxSetupData, void *ctx);
 
 // Boundary Condition Functions
 extern PetscErrorCode BC_DENSITY_CURRENT(DM dm, SimpleBC bc,
