### 2021-03-11 5554b4c
Add CeedOperatorLinearFullAssemble() to explicitly assemble a CeedOperator
The output format is a stream of (row, column, value) triples, which client
code can use to assemble a sparse matrix by adding the values in the
appropriate places.
