+++ interface/ceed-operator.c
@@ -296,6 +296,37 @@ int CeedOperatorSingleView(CeedOperator op, bool sub, FILE *stream) {
 }
 
 /**
+  @brief Find the active vector ElemRestriction for a CeedOperator
+
+  @param[in] op            CeedOperator to find active basis for
+  @param[out] activerstr   ElemRestriction for active input vector
+
+  @return An error code: 0 - success, otherwise - failure
+
+  @ref Utility
+**/
+static int CeedOperatorGetActiveElemRestriction(CeedOperator op,
+    CeedElemRestriction *activerstr) {
+  *activerstr = NULL;
+  for (int i = 0; i < op->qf->numinputfields; i++)
+    if (op->inputfields[i]->vec == CEED_VECTOR_ACTIVE) {
+      *activerstr = op->inputfields[i]->Erestrict;
+      break;
+    }
+
+  if (!*activerstr) {
+    // LCOV_EXCL_START
+    int ierr;
+    Ceed ceed;
+    ierr = CeedOperatorGetCeed(op, &ceed); CeedChk(ierr);
+    return CeedError(ceed, 1,
+                     "No active ElemRestriction found!");
+    // LCOV_EXCL_STOP
+  }
+  return 0;
+}
+
+/**
   @brief Find the active vector basis for a CeedOperator
 
   @param[in] op            CeedOperator to find active basis for
@@ -1281,6 +1312,378 @@ int CeedOperatorLinearAssembleAddPointBlockDiagonal(CeedOperator op,
            assembled, request); CeedChk(ierr);
   }
   return CEED_ERROR_SUCCESS;
+}
+
+/**
+   @brief Perform matrix-matrix multiply C = A B
+
+   Given m by n matrix A (column-major order) and n by p matrix B
+   (also column-major) return m by p matrix C (which should already
+   be allocated when you call this).
+
+   This could, probably should, be a GPU kernel.
+
+   @ref Utility
+**/
+static int MatrixMatrixMultiply(CeedInt m, CeedInt n, CeedInt p,
+                                const CeedScalar *A,
+                                const CeedScalar *B,
+                                CeedScalar *C) {
+  for (int i = 0; i < m; ++i) {
+    for (int j = 0; j < p; ++j) {
+      // C[j*m + i] = 0.0;
+      for (int k = 0; k < n; ++k) {
+        C[j*m + i] += A[k*m + i] * B[j*n + k];
+      }
+    }
+  }
+
+  return 0;
+}
+
+/**
+   @brief Fully assemble a non-composite CeedOperator
+
+   Users should generally call CeedOperatorLinearFullAssemble() instead.
+
+   offset is a running count; this function puts entries into rows, cols, vals
+   starting at rows[offset] until it is finished with op.
+
+   @ref Developer
+**/
+int CeedSingleOperatorFullAssemble(CeedOperator op, CeedInt offset,
+                                   CeedInt *rows, CeedInt *cols, CeedScalar *vals) {
+  int ierr;
+  Ceed ceed = op->ceed;;
+
+  // Assemble QFunction
+  CeedQFunction qf;
+  ierr = CeedOperatorGetQFunction(op, &qf); CeedChk(ierr);
+  CeedInt numinputfields, numoutputfields;
+  ierr= CeedQFunctionGetNumArgs(qf, &numinputfields, &numoutputfields);
+  CeedChk(ierr);
+  CeedVector assembledqf;
+  CeedElemRestriction rstr_q;
+  ierr = CeedOperatorLinearAssembleQFunction(
[…]
