+++ tests/t561-operator.c
@@ -1,0 +1,169 @@
+/// @file
+/// Test full assembly of Poisson operator
+/// \test Test full assembly of Poisson operator
+#include <ceed.h>
+#include <stdlib.h>
+#include <math.h>
+#include "t534-operator.h"
+
+int main(int argc, char **argv) {
+  Ceed ceed;
+  CeedElemRestriction Erestrictx, Erestrictu,
+                      Erestrictui, Erestrictqi;
+  CeedBasis bx, bu;
+  CeedQFunction qf_setup, qf_diff;
+  CeedOperator op_setup, op_diff;
+  CeedVector qdata, X, U, V;
+  CeedInt P = 3, Q = 4, dim = 2;
+  CeedInt nx = 3, ny = 2;
+  CeedInt nelem = nx * ny;
+  CeedInt ndofs = (nx*2+1)*(ny*2+1), nqpts = nelem*Q*Q;
+  CeedInt indx[nelem*P*P];
+  CeedScalar assembled[ndofs*ndofs];
+  CeedScalar x[dim*ndofs], assembledTrue[ndofs*ndofs];
+  CeedScalar *u;
+  const CeedScalar *v;
+
+  CeedInit(argv[1], &ceed);
+
+  // DoF Coordinates
+  for (CeedInt i=0; i<nx*2+1; i++)
+    for (CeedInt j=0; j<ny*2+1; j++) {
+      x[i+j*(nx*2+1)+0*ndofs] = (CeedScalar) i / (2*nx);
+      x[i+j*(nx*2+1)+1*ndofs] = (CeedScalar) j / (2*ny);
+    }
+  CeedVectorCreate(ceed, dim*ndofs, &X);
+  CeedVectorSetArray(X, CEED_MEM_HOST, CEED_USE_POINTER, x);
+
+  // Qdata Vector
+  CeedVectorCreate(ceed, nqpts*dim*(dim+1)/2, &qdata);
+
+  // Element Setup
+  for (CeedInt i=0; i<nelem; i++) {
+    CeedInt col, row, offset;
+    col = i % nx;
+    row = i / nx;
+    offset = col*(P-1) + row*(nx*2+1)*(P-1);
+    for (CeedInt j=0; j<P; j++)
+      for (CeedInt k=0; k<P; k++)
+        indx[P*(P*i+k)+j] = offset + k*(nx*2+1) + j;
+  }
+
+  // Restrictions
+  CeedElemRestrictionCreate(ceed, nelem, P*P, dim, ndofs, dim*ndofs,
+                            CEED_MEM_HOST, CEED_USE_POINTER, indx, &Erestrictx);
+
+  CeedElemRestrictionCreate(ceed, nelem, P*P, 1, 1, ndofs, CEED_MEM_HOST,
+                            CEED_USE_POINTER, indx, &Erestrictu);
+  CeedInt stridesu[3] = {1, Q*Q, Q*Q};
+  CeedElemRestrictionCreateStrided(ceed, nelem, Q*Q, 1, nqpts, stridesu,
+                                   &Erestrictui);
+
+  CeedInt stridesqd[3] = {1, Q*Q, Q *Q *dim *(dim+1)/2};
+  CeedElemRestrictionCreateStrided(ceed, nelem, Q*Q, dim*(dim+1)/2,
+                                   dim*(dim+1)/2*nqpts,
+                                   stridesqd, &Erestrictqi);
+
+  // Bases
+  CeedBasisCreateTensorH1Lagrange(ceed, dim, dim, P, Q, CEED_GAUSS, &bx);
+  CeedBasisCreateTensorH1Lagrange(ceed, dim, 1, P, Q, CEED_GAUSS, &bu);
+
[…]
