### 2021-03-11 e978d27
Add full assembly test for triangular mesh, t563-operator.c, which is analogous to t536-operator.c
