+++ tests/t563-operator.c
@@ -1,0 +1,215 @@
+/// @file
+/// Test full assembly of mass and Poisson operator (see t536)
+/// \test Test full assembly of mass and Poisson operator
+#include <ceed.h>
+#include <stdlib.h>
+#include <math.h>
+#include "t320-basis.h"
+#include "t535-operator.h"
+
+int main(int argc, char **argv) {
+  Ceed ceed;
+  CeedElemRestriction Erestrictx, Erestrictu,
+                      Erestrictui, Erestrictqi;
+  CeedBasis bx, bu;
+  CeedQFunction qf_setup_mass, qf_setup_diff, qf_apply;
+  CeedOperator op_setup_mass, op_setup_diff, op_apply;
+  CeedVector qdata_mass, qdata_diff, X, U, V;
+  CeedInt nelem = 12, dim = 2, P = 6, Q = 4;
+  CeedInt nx = 3, ny = 2;
+  CeedInt row, col, offset;
+  CeedInt ndofs = (nx*2+1)*(ny*2+1), nqpts = nelem*Q;
+  CeedInt indx[nelem*P*P];
+  CeedScalar assembled[ndofs*ndofs];
+  CeedScalar x[dim*ndofs], assembledTrue[ndofs*ndofs];
+  CeedScalar qref[dim*Q], qweight[Q];
+  CeedScalar interp[P*Q], grad[dim*P*Q];
+  CeedScalar *u;
+  const CeedScalar *v;
+
+  CeedInit(argv[1], &ceed);
+
+  // DoF Coordinates
+  for (CeedInt i=0; i<ndofs; i++) {
+    x[i] = (1. / (nx*2)) * (CeedScalar) (i % (nx*2+1));
+    x[i+ndofs] = (1. / (ny*2)) * (CeedScalar) (i / (nx*2+1));
+  }
+  CeedVectorCreate(ceed, dim*ndofs, &X);
+  CeedVectorSetArray(X, CEED_MEM_HOST, CEED_USE_POINTER, x);
+
+  // Qdata Vectors
+  CeedVectorCreate(ceed, nqpts, &qdata_mass);
+  CeedVectorCreate(ceed, nqpts*dim*(dim+1)/2, &qdata_diff);
+
+  // Element Setup
+  for (CeedInt i=0; i<nelem/2; i++) {
+    col = i % nx;
+    row = i / nx;
+    offset = col*2 + row*(nx*2+1)*2;
+
+    indx[i*2*P +  0] =  2 + offset;
+    indx[i*2*P +  1] =  9 + offset;
+    indx[i*2*P +  2] = 16 + offset;
+    indx[i*2*P +  3] =  1 + offset;
+    indx[i*2*P +  4] =  8 + offset;
+    indx[i*2*P +  5] =  0 + offset;
+
+    indx[i*2*P +  6] = 14 + offset;
+    indx[i*2*P +  7] =  7 + offset;
+    indx[i*2*P +  8] =  0 + offset;
+    indx[i*2*P +  9] = 15 + offset;
+    indx[i*2*P + 10] =  8 + offset;
+    indx[i*2*P + 11] = 16 + offset;
+  }
+
+  // Restrictions
+  CeedElemRestrictionCreate(ceed, nelem, P, dim, ndofs, dim*ndofs,
+                            CEED_MEM_HOST, CEED_USE_POINTER, indx, &Erestrictx);
+
+  CeedElemRestrictionCreate(ceed, nelem, P, 1, 1, ndofs, CEED_MEM_HOST,
+                            CEED_USE_POINTER, indx, &Erestrictu);
+  CeedInt stridesu[3] = {1, Q, Q};
+  CeedElemRestrictionCreateStrided(ceed, nelem, Q, 1, nqpts, stridesu,
[…]
