### 2021-03-15 af734c2
SWE: fix the weak form in the explicit scheme
