### 2021-03-15 816e79c
assembly: change interface to separate nonzero structure from values
