+++ tests/t563-operator.c
@@ -151,12 +151,16 @@ int main(int argc, char **argv) {
   CeedInt nentries;
   CeedInt *rows;
   CeedInt *cols;
-  CeedScalar *vals;
-  CeedOperatorLinearFullAssemble(op_apply, &nentries, &rows, &cols,
-                                 &vals);
+  CeedVector values;
+  CeedOperatorLinearAssembleSymbolic(op_apply, &nentries, &rows, &cols);
+  CeedVectorCreate(ceed, nentries, &values);
+  CeedOperatorLinearAssemble(op_apply, values);
+  const CeedScalar *vals;
+  CeedVectorGetArrayRead(values, CEED_MEM_HOST, &vals);
   for (int k=0; k<nentries; ++k) {
     assembled[rows[k]*ndofs + cols[k]] += vals[k];
   }
+  CeedVectorRestoreArrayRead(values, &vals);
 
   // Manually assemble operator
   CeedVectorCreate(ceed, ndofs, &U);
@@ -192,7 +196,7 @@ int main(int argc, char **argv) {
   // Cleanup
   free(rows);
   free(cols);
-  free(vals);
+  CeedVectorDestroy(&values);
   CeedQFunctionDestroy(&qf_setup_mass);
   CeedQFunctionDestroy(&qf_setup_diff);
   CeedQFunctionDestroy(&qf_apply);
