+++ interface/ceed-operator.c
@@ -1315,11 +1315,12 @@ int CeedOperatorLinearAssembleAddPointBlockDiagonal(CeedOperator op,
 }
 
 /**
-   @brief Perform matrix-matrix multiply C = A B
+   @brief Perform matrix-matrix multiply C += A B
 
-   Given m by n matrix A (column-major order) and n by p matrix B
-   (also column-major) return m by p matrix C (which should already
-   be allocated when you call this).
+   Given m by n matrix A and n by p matrix B, add the matrix-matrix
+   product AB into the existing matrix C.
+
+   All matrices are stored in column-major order.
 
    This could, probably should, be a GPU kernel.
 
@@ -1331,7 +1332,6 @@ static int MatrixMatrixMultiply(CeedInt m, CeedInt n, CeedInt p,
                                 CeedScalar *C) {
   for (int i = 0; i < m; ++i) {
     for (int j = 0; j < p; ++j) {
-      // C[j*m + i] = 0.0;
       for (int k = 0; k < n; ++k) {
         C[j*m + i] += A[k*m + i] * B[j*n + k];
       }
@@ -1342,19 +1342,104 @@ static int MatrixMatrixMultiply(CeedInt m, CeedInt n, CeedInt p,
 }
 
 /**
-   @brief Fully assemble a non-composite CeedOperator
+   @brief Build nonzero pattern for non-composite operator.
 
-   Users should generally call CeedOperatorLinearFullAssemble() instead.
-
-   offset is a running count; this function puts entries into rows, cols, vals
-   starting at rows[offset] until it is finished with op.
+   Users should generally use CeedOperatorLinearAssembleSymbolic()
 
    @ref Developer
 **/
-int CeedSingleOperatorFullAssemble(CeedOperator op, CeedInt offset,
-                                   CeedInt *rows, CeedInt *cols, CeedScalar *vals) {
+int CeedSingleOperatorAssembleSymbolic(CeedOperator op, CeedInt offset,
+                                       CeedInt *rows, CeedInt *cols) {
+  int ierr;
+  Ceed ceed = op->ceed;
+  if (op->composite) {
+    // LCOV_EXCL_START
+    return CeedError(ceed, 1, "Not suitable for composite operator");
+    // LCOV_EXCL_STOP
+  }
+
+  CeedElemRestriction rstrin;
+  ierr = CeedOperatorGetActiveElemRestriction(op, &rstrin); CeedChk(ierr);
+  CeedInt nelem, elemsize, nnodes, ncomp;
+  ierr = CeedElemRestrictionGetNumElements(rstrin, &nelem); CeedChk(ierr);
+  ierr = CeedElemRestrictionGetElementSize(rstrin, &elemsize); CeedChk(ierr);
+  ierr = CeedElemRestrictionGetLVectorSize(rstrin, &nnodes); CeedChk(ierr);
+  ierr = CeedElemRestrictionGetNumComponents(rstrin, &ncomp); CeedChk(ierr);
+  CeedInt layout_er[3];
+  ierr = CeedElemRestrictionGetELayout(rstrin, &layout_er); CeedChk(ierr);
+
+  CeedInt local_nentries = elemsize*ncomp * elemsize*ncomp * nelem;
+
+  // Determine elem_dof relation
+  CeedVector index_vec;
+  ierr = CeedVectorCreate(ceed, nnodes, &index_vec); CeedChk(ierr);
+  CeedScalar *array;
+  ierr = CeedVectorGetArray(index_vec, CEED_MEM_HOST, &array); CeedChk(ierr);
+  for (CeedInt i = 0; i < nnodes; ++i) {
+    array[i] = i;
+  }
+  ierr = CeedVectorRestoreArray(index_vec, &array); CeedChk(ierr);
[…]
