+++ examples/solids/src/cloptions.c
@@ -232,7 +232,7 @@ PetscErrorCode ProcessCommandLineOptions(MPI_Comm comm, AppCtx appCtx) {
 };
 
 // Process physics options
-PetscErrorCode ProcessPhysics(MPI_Comm comm, Physics phys, Units units) {
+PetscErrorCode ProcessPhysics(MPI_Comm comm, Physics phys, Units units) { //TO-DO
   PetscErrorCode ierr;
   PetscBool nuFlag = PETSC_FALSE;
   PetscBool YoungFlag = PETSC_FALSE;
@@ -284,6 +284,71 @@ PetscErrorCode ProcessPhysics(MPI_Comm comm, Physics phys, Units units) {
 
   // Scale E to Pa
   phys->E *= units->Pascal;
+
+  PetscFunctionReturn(0);
+};
+
+// Process physics options
+PetscErrorCode ProcessPhysics_MR(MPI_Comm comm, Physics_MR phys_MR, Units units) { //TO-DO
+  PetscErrorCode ierr;
+  PetscBool mu_1_Flag = PETSC_FALSE;
+  PetscBool mu_2_Flag = PETSC_FALSE;
+  PetscBool k_1_Flag = PETSC_FALSE;
+  phys_MR->mu_1 = 0;
+  phys_MR->mu_2 = 0;
+  phys_MR->k_1 = 0;
+  units->meter     = 1;        // 1 meter in scaled length units
+  units->second    = 1;        // 1 second in scaled time units
+  units->kilogram  = 1;        // 1 kilogram in scaled mass units
+
+  PetscFunctionBeginUser;
+
+  ierr = PetscOptionsBegin(comm, NULL,
+                           "Elasticity / Hyperelasticity in PETSc with libCEED",
+                           NULL); CHKERRQ(ierr);
+
+  ierr = PetscOptionsScalar("-mu_1", "Material Property mu_1", NULL, phys_MR->mu_1, &phys_MR->mu_1,
+                            &mu_1_Flag); CHKERRQ(ierr);
+
+  ierr = PetscOptionsScalar("-mu_2", "Material Property mu_2", NULL, phys_MR->mu_2, &phys_MR->mu_2,
+                            &mu_2_Flag); CHKERRQ(ierr);
+
+  ierr = PetscOptionsScalar("-K", "Material Property K_1", NULL, phys_MR->k_1, &phys_MR->k_1,
+                            &k_1_Flag); CHKERRQ(ierr);
+
+  ierr = PetscOptionsScalar("-units_meter", "1 meter in scaled length units",
+                            NULL, units->meter, &units->meter, NULL);
+  CHKERRQ(ierr);
+  units->meter = fabs(units->meter);
+
+  ierr = PetscOptionsScalar("-units_second", "1 second in scaled time units",
+                            NULL, units->second, &units->second, NULL);
+  CHKERRQ(ierr);
+  units->second = fabs(units->second);
+
+  ierr = PetscOptionsScalar("-units_kilogram", "1 kilogram in scaled mass units",
+                            NULL, units->kilogram, &units->kilogram, NULL);
+  CHKERRQ(ierr);
+  units->kilogram = fabs(units->kilogram);
+
+  ierr = PetscOptionsEnd(); CHKERRQ(ierr); // End of setting Physics
+
+  // Check for all required options to be set
+  if (!mu_1_Flag) {
+    SETERRQ(PETSC_COMM_SELF, PETSC_ERR_SUP, "-mu_1 option needed");
+  }
+  if (!mu_2_Flag) {
+    SETERRQ(PETSC_COMM_SELF, PETSC_ERR_SUP, "-mu_2 option needed");
+  }
[…]
