### 2021-03-17 da12e01
SWE: some comments on the div operations
