### 2021-03-17 d99129b
Command line option for viz in Fluids and Solids examples
examples/fluids: change the cl option -of to -output_dir for the output folder

examples/solids: add cl option -output_dir for the viz output

change: output folder -> output directory

Co-authored-by: Jed Brown <jed@jedbrown.org>

examples: change variable name outputfolder -> outputdir

examples: update the README files
