### 2021-03-18 2c2833d
examples/fluids: drop the broken supg test of density_current for now
The stabilization enum is not passed to density_current QFunction hence
the test doesn't work. We need to create a new test when the code is
debugged.
