### 2021-03-18 f61a6f6
assembly: use CeedMatrixMultiply() instead of hand-rolled duplicated code
This also involves switching some arrays from column-major ordering to
row-major ordering.
