### 2021-03-18 161bc06
Remove CeedMatrixMultiply from ceed-impl.h because it is already in ceed-backend.h
