### 2021-03-20 bd4df46
Python: fix parse/skip of multi-line CEED_VERSION_GE macro
