+++ python/build_ceed_cffi.py
@@ -19,6 +19,8 @@ import re
 from cffi import FFI
 ffibuilder = FFI()
 
+ceed_version_ge = re.compile('\s+\(!?CEED_VERSION.*')
+
 # ------------------------------------------------------------------------------
 # Provide C definitions to CFFI
 # ------------------------------------------------------------------------------
@@ -28,7 +30,8 @@ with open(os.path.abspath("include/ceed.h")) as f:
              not line.startswith("  static") and
              "CeedErrorImpl" not in line and
              "const char *, ...);" not in line and
-             not line.startswith("CEED_EXTERN const char *const")]
+             not line.startswith("CEED_EXTERN const char *const") and
+             not ceed_version_ge.match(line)]
     lines = [line.replace("CEED_EXTERN", "extern") for line in lines]
     header = '\n'.join(lines)
     header = header.split("static inline CeedInt CeedIntPow", 1)[0]
