### 2021-03-20 e15f9bd
Error Handling Improvement [fix #696] (#691)
* Operator - add operator/qfunction field compatibility checks

* QFunction - do not allow adding fields to QFunction in use with an operator

* Examples - add some extra exclusion markers in ceed example

* error - add error enum

* error - update error enum names and numbering

* error - use CEED_ERROR_BACKEND in all backend errors

* error - begin classifying interface errors

* error - update backends to use CEED_ERROR_SUCCESS and CeedChkBackend

* error - use new errors in gallery

* error - add some unsaved modifications

* error - improve doc
[…]
