+++ backends/xsmm/ceed-xsmm-blocked.c
@@ -28,21 +28,22 @@ static int CeedInit_Xsmm_Blocked(const char *resource, Ceed ceed) {
   if (strcmp(resource, "/cpu/self") && strcmp(resource, "/cpu/self/xsmm")
       && strcmp(resource, "/cpu/self/xsmm/blocked"))
     // LCOV_EXCL_START
-    return CeedError(ceed, 1, "blocked libXSMM backend cannot use resource: %s",
+    return CeedError(ceed, CEED_ERROR_BACKEND,
+                     "blocked libXSMM backend cannot use resource: %s",
                      resource);
   // LCOV_EXCL_STOP
-  ierr = CeedSetDeterministic(ceed, true); CeedChk(ierr);
+  ierr = CeedSetDeterministic(ceed, true); CeedChkBackend(ierr);
 
   // Create reference CEED that implementation will be dispatched
   //   through unless overridden
   Ceed ceedref;
   CeedInit("/cpu/self/opt/blocked", &ceedref);
-  ierr = CeedSetDelegate(ceed, ceedref); CeedChk(ierr);
+  ierr = CeedSetDelegate(ceed, ceedref); CeedChkBackend(ierr);
 
   ierr = CeedSetBackendFunction(ceed, "Ceed", ceed, "TensorContractCreate",
-                                CeedTensorContractCreate_Xsmm); CeedChk(ierr);
+                                CeedTensorContractCreate_Xsmm); CeedChkBackend(ierr);
 
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 //------------------------------------------------------------------------------
