+++ backends/cuda/ceed-cuda-qfunctioncontext.c
@@ -26,7 +26,7 @@
 static inline size_t bytes(const CeedQFunctionContext ctx) {
   int ierr;
   size_t ctxsize;
-  ierr = CeedQFunctionContextGetContextSize(ctx, &ctxsize); CeedChk(ierr);
+  ierr = CeedQFunctionContextGetContextSize(ctx, &ctxsize); CeedChkBackend(ierr);
   return ctxsize;
 }
 
@@ -37,13 +37,13 @@ static inline int CeedQFunctionContextSyncH2D_Cuda(
   const CeedQFunctionContext ctx) {
   int ierr;
   Ceed ceed;
-  ierr = CeedQFunctionContextGetCeed(ctx, &ceed); CeedChk(ierr);
+  ierr = CeedQFunctionContextGetCeed(ctx, &ceed); CeedChkBackend(ierr);
   CeedQFunctionContext_Cuda *impl;
-  ierr = CeedQFunctionContextGetBackendData(ctx, &impl); CeedChk(ierr);
+  ierr = CeedQFunctionContextGetBackendData(ctx, &impl); CeedChkBackend(ierr);
 
   ierr = cudaMemcpy(impl->d_data, impl->h_data, bytes(ctx),
                     cudaMemcpyHostToDevice); CeedChk_Cu(ceed, ierr);
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 //------------------------------------------------------------------------------
@@ -53,13 +53,13 @@ static inline int CeedQFunctionContextSyncD2H_Cuda(
   const CeedQFunctionContext ctx) {
   int ierr;
   Ceed ceed;
-  ierr = CeedQFunctionContextGetCeed(ctx, &ceed); CeedChk(ierr);
+  ierr = CeedQFunctionContextGetCeed(ctx, &ceed); CeedChkBackend(ierr);
   CeedQFunctionContext_Cuda *impl;
-  ierr = CeedQFunctionContextGetBackendData(ctx, &impl); CeedChk(ierr);
+  ierr = CeedQFunctionContextGetBackendData(ctx, &impl); CeedChkBackend(ierr);
 
   ierr = cudaMemcpy(impl->h_data, impl->d_data, bytes(ctx),
                     cudaMemcpyDeviceToHost); CeedChk_Cu(ceed, ierr);
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 //------------------------------------------------------------------------------
@@ -70,28 +70,28 @@ static int CeedQFunctionContextSetDataHost_Cuda(const CeedQFunctionContext ctx,
     CeedScalar *data) {
   int ierr;
   CeedQFunctionContext_Cuda *impl;
-  ierr = CeedQFunctionContextGetBackendData(ctx, &impl); CeedChk(ierr);
+  ierr = CeedQFunctionContextGetBackendData(ctx, &impl); CeedChkBackend(ierr);
 
   switch (cmode) {
   case CEED_COPY_VALUES: {
     if(!impl->h_data) {
-      ierr = CeedMalloc(bytes(ctx), &impl->h_data_allocated); CeedChk(ierr);
+      ierr = CeedMalloc(bytes(ctx), &impl->h_data_allocated); CeedChkBackend(ierr);
       impl->h_data = impl->h_data_allocated;
     }
     memcpy(impl->h_data, data, bytes(ctx));
   } break;
   case CEED_OWN_POINTER:
-    ierr = CeedFree(&impl->h_data_allocated); CeedChk(ierr);
+    ierr = CeedFree(&impl->h_data_allocated); CeedChkBackend(ierr);
     impl->h_data_allocated = data;
     impl->h_data = data;
     break;
   case CEED_USE_POINTER:
-    ierr = CeedFree(&impl->h_data_allocated); CeedChk(ierr);
[…]
