+++ backends/occa/ceed-occa-qfunction.cpp
@@ -83,6 +83,7 @@ namespace ceed {
       // CEED defines
       props["defines/CeedPragmaSIMD"] = "";
       props["defines/CEED_Q_VLA"] = "OCCA_Q";
+      props["defines/CEED_ERROR_SUCCESS"] = 0;
 
       std::stringstream ss;
       ss << "#define CEED_QFUNCTION(FUNC_NAME) \\" << std::endl
@@ -112,7 +113,7 @@ namespace ceed {
         );
       }
 
-      return 0;
+      return CEED_ERROR_SUCCESS;
     }
 
     std::string QFunction::getKernelSource(const std::string &kernelName,
@@ -212,7 +213,7 @@ namespace ceed {
 
       qFunctionKernel.run();
 
-      return 0;
+      return CEED_ERROR_SUCCESS;
     }
 
     //---[ Ceed Callbacks ]-----------
@@ -236,7 +237,7 @@ namespace ceed {
       CeedOccaRegisterFunction(qf, "Apply", QFunction::ceedApply);
       CeedOccaRegisterFunction(qf, "Destroy", QFunction::ceedDestroy);
 
-      return 0;
+      return CEED_ERROR_SUCCESS;
     }
 
     int QFunction::ceedApply(CeedQFunction qf, CeedInt Q,
@@ -251,7 +252,7 @@ namespace ceed {
 
     int QFunction::ceedDestroy(CeedQFunction qf) {
       delete QFunction::from(qf);
-      return 0;
+      return CEED_ERROR_SUCCESS;
     }
   }
 }
