+++ gallery/mass1d/ceed-massapply.c
@@ -30,7 +30,8 @@ static int CeedQFunctionInit_MassApply(Ceed ceed, const char *requested,
   const char *name = "MassApply";
   if (strcmp(name, requested))
     // LCOV_EXCL_START
-    return CeedError(ceed, 1, "QFunction '%s' does not match requested name: %s",
+    return CeedError(ceed, CEED_ERROR_UNSUPPORTED,
+                     "QFunction '%s' does not match requested name: %s",
                      name, requested);
   // LCOV_EXCL_STOP
 
@@ -39,7 +40,7 @@ static int CeedQFunctionInit_MassApply(Ceed ceed, const char *requested,
   ierr = CeedQFunctionAddInput(qf, "qdata", 1, CEED_EVAL_NONE); CeedChk(ierr);
   ierr = CeedQFunctionAddOutput(qf, "v", 1, CEED_EVAL_INTERP); CeedChk(ierr);
 
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 /**
