### 2021-03-23 308c0cb
assembly: modify use of fallback resource
