+++ examples/fluids/advection2d.c
@@ -1,7 +1,7 @@
 #include "navierstokes.h"
 
 PetscErrorCode NS_ADVECTION2D(problemData *problem, void **ctxSetupData,
-                              void *ctx) {
+                                  void **ctx, void **ctxPhys) {
 
   PetscFunctionBeginUser;
   problem->dim                       = 2;
