+++ examples/fluids/navierstokes.h
@@ -120,9 +120,9 @@ struct Physics_ {
 extern PetscErrorCode NS_DENSITY_CURRENT(problemData *problem,
     void **ctxSetupData, void **ctx, void **ctxPhys);
 extern PetscErrorCode NS_ADVECTION(problemData *problem,
-                                   void **ctxSetupData, void *ctx);
+    void **ctxSetupData, void **ctx, void **ctxPhys);
 extern PetscErrorCode NS_ADVECTION2D(problemData *problem,
-                                     void **ctxSetupData, void *ctx);
+    void **ctxSetupData, void **ctx, void **ctxPhys);
 
 // Boundary Condition Functions
 extern PetscErrorCode BC_DENSITY_CURRENT(DM dm, SimpleBC bc,
