+++ examples/fluids/advection.c
@@ -2,8 +2,18 @@
 
 PetscErrorCode NS_ADVECTION(problemData *problem, void **ctxSetupData,
                                   void **ctx, void **ctxPhys) {
+  PetscInt ierr;
+  MPI_Comm comm = PETSC_COMM_WORLD;
+  StabilizationType stab;
+  SetupContext ctxSetup = *(SetupContext *)ctxSetupData;
+  Units units = *(Units *)ctx;
+  Physics ctxPhysData = *(Physics *)ctxPhys;
+  ierr = PetscMalloc1(1, &ctxPhysData->ctxAdvectionData); CHKERRQ(ierr);
 
   PetscFunctionBeginUser;
+  // ------------------------------------------------------
+  //               SET UP ADVECTION
+  // ------------------------------------------------------
   problem->dim                       = 3;
   problem->qdatasizeVol              = 10;
   problem->qdatasizeSur              = 4;
@@ -21,6 +31,106 @@ PetscErrorCode NS_ADVECTION(problemData *problem, void **ctxSetupData,
   problem->applySur_loc              = Advection_Sur_loc;
   problem->bc                        = BC_ADVECTION;
   problem->non_zero_time             = PETSC_FALSE;
+
+  // ------------------------------------------------------
+  //             Create the libCEED context
+  // ------------------------------------------------------
+  PetscScalar lx         = 8000.;       // m
+  PetscScalar ly         = 8000.;       // m
+  PetscScalar lz         = 4000.;       // m
+  CeedScalar rc          = 1000.;       // m (Radius of bubble)
+  CeedScalar CtauS       = 0.;          // dimensionless
+  CeedScalar strong_form = 0.;          // [0,1]
+  CeedScalar E_wind      = 1.e6;        // J
+  PetscReal wind[3]      = {1., 0, 0};  // m/s
+
+  // ------------------------------------------------------
+  //             Create the PETSc context
+  // ------------------------------------------------------
+  PetscScalar meter    = 1e-2;  // 1 meter in scaled length units
+  PetscScalar kilogram = 1e-6;  // 1 kilogram in scaled mass units
+  PetscScalar second   = 1e-2;  // 1 second in scaled time units
+  PetscScalar Joule;
+
+  // ------------------------------------------------------
+  //              Command line Options
+  // ------------------------------------------------------
+  ierr = PetscOptionsBegin(comm, NULL, "Options for ADVECTION",
+                           NULL); CHKERRQ(ierr);
+  // -- Physics
+  ierr = PetscOptionsScalar("-lx", "Length scale in x direction",
+                            NULL, lx, &lx, NULL); CHKERRQ(ierr);
+  ierr = PetscOptionsScalar("-ly", "Length scale in y direction",
+                            NULL, ly, &ly, NULL); CHKERRQ(ierr);
+  ierr = PetscOptionsScalar("-lz", "Length scale in z direction",
+                            NULL, lz, &lz, NULL); CHKERRQ(ierr);
+  ierr = PetscOptionsScalar("-rc", "Characteristic radius of thermal bubble",
+                            NULL, rc, &rc, NULL); CHKERRQ(ierr);
+  PetscInt n = problem->dim;
+  PetscBool userWind;
+  ierr = PetscOptionsRealArray("-problem_advection_wind_translation",
+                               "Constant wind vector",
+                               NULL, wind, &n, &userWind); CHKERRQ(ierr);
+  ierr = PetscOptionsScalar("-CtauS",
+                            "Scale coefficient for tau (nondimensional)",
[…]
