+++ examples/fluids/navierstokes.h
@@ -113,6 +113,7 @@ typedef struct Physics_ *Physics;
 struct Physics_ {
   NSContext ctxNSData;
   AdvectionContext ctxAdvectionData;
+  WindType wind_type;
   StabilizationType stab;
 };
 
