+++ examples/fluids/advection.c
@@ -150,14 +150,18 @@ PetscErrorCode BC_ADVECTION(DM dm, SimpleBC bc, WindType wind_type,
   MPI_Comm comm = PETSC_COMM_WORLD;
 
   // Default boundary conditions
-  //   slip bc on all faces and no wall bc
-  bc->nslip[0] = bc->nslip[1] = bc->nslip[2] = 2;
-  bc->slips[0][0] = 5;
-  bc->slips[0][1] = 6;
-  bc->slips[1][0] = 3;
-  bc->slips[1][1] = 4;
-  bc->slips[2][0] = 1;
-  bc->slips[2][1] = 2;
+  if (wind_type == ADVECTION_WIND_TRANSLATION) {
+    bc->nwall = 0;
+    bc->nslip[0] = bc->nslip[1] = bc->nslip[2] = 0;
+  } else { // Default boundary conditions
+    bc->nslip[0] = bc->nslip[1] = bc->nslip[2] = 2;
+    bc->slips[0][0] = 5;
+    bc->slips[0][1] = 6;
+    bc->slips[1][0] = 3;
+    bc->slips[1][1] = 4;
+    bc->slips[2][0] = 1;
+    bc->slips[2][1] = 2;
+  }
 
   PetscFunctionBeginUser;
   // Parse command line options
