### 2021-03-25 74c2e4b
doc: add livehtml target for sphinx-autobuild
