### 2021-03-26 ec3da8b
Install install backend headers under include/ceed/
This makes it possible to distribute source plugins that provide
additional backends. It's also used in MFEM, perhaps temporarily.

Deprecate ceed-backend.h, which was not previously installed, but some
users accessed it from an in-place build.

Also install CUDA and HIP headers that allow users to provide CUfunction
and hipFunction_t.

Co-authored-by: Jeremy L. Thompson <jeremy.thompson@colorado.edu>
Requested-by: Andrew T. Barker <barker29@llnl.gov>
