+++ backends/opt/ceed-opt.h
@@ -17,8 +17,8 @@
 #ifndef _ceed_opt_h
 #define _ceed_opt_h
 
-#include <ceed.h>
-#include <ceed-backend.h>
+#include <ceed/ceed.h>
+#include <ceed/backend.h>
 #include <stdbool.h>
 #include <stdint.h>
 
