+++ backends/hip/ceed-hip-basis.c
@@ -14,8 +14,8 @@
 // software, applications, hardware, advanced system engineering and early
 // testbed platforms, in support of the nation's exascale computing imperative.
 
-#include <ceed.h>
-#include <ceed-backend.h>
+#include <ceed/ceed.h>
+#include <ceed/backend.h>
 #include <hip/hip_runtime.h>
 #include "ceed-hip.h"
 #include "ceed-hip-compile.h"
