+++ backends/memcheck/ceed-memcheck-qfunction.c
@@ -14,8 +14,8 @@
 // software, applications, hardware, advanced system engineering and early
 // testbed platforms, in support of the nation's exascale computing imperative.
 
-#include <ceed.h>
-#include <ceed-backend.h>
+#include <ceed/ceed.h>
+#include <ceed/backend.h>
 #include <valgrind/memcheck.h>
 #include "ceed-memcheck.h"
 
