+++ backends/magma/kernels/hip/interp_2d.hip.cpp
@@ -138,7 +138,6 @@ magma_interp_1d_ncomp_q(
     return launch_failed;
 }
 
-
 //////////////////////////////////////////////////////////////////////////////////////////
 static magma_int_t 
 magma_interp_2d_ncomp_q_p(
@@ -194,7 +193,6 @@ magma_interp_2d_ncomp_q_p(
     }
     return launch_failed;
 }
-
 
 
 //////////////////////////////////////////////////////////////////////////////////////////
