### 2021-03-26 bac76c5
Fluids - Refactor: drop the supg regression test for now
We need to debug SU/SUPG for density_current. The test does not work.
