+++ examples/fluids/navierstokes.h
@@ -119,11 +119,11 @@ struct Physics_ {
 
 // Setup function for each problem
 extern PetscErrorCode NS_DENSITY_CURRENT(problemData *problem,
-    void **ctxSetupData, void **ctx, void **ctxPhys);
+    void *ctxSetupData, void *ctx, void *ctxPhys);
 extern PetscErrorCode NS_ADVECTION(problemData *problem,
-                                   void **ctxSetupData, void **ctx, void **ctxPhys);
+                                   void *ctxSetupData, void *ctx, void *ctxPhys);
 extern PetscErrorCode NS_ADVECTION2D(problemData *problem,
-                                     void **ctxSetupData, void **ctx, void **ctxPhys);
+                                     void *ctxSetupData, void *ctx, void *ctxPhys);
 
 // Boundary Condition Functions
 extern PetscErrorCode BC_DENSITY_CURRENT(DM dm, SimpleBC bc, WindType wind_type,
