### 2021-03-28 e1e9e29
Stricter field checking and better error messages
Add CEED_ERROR_INCOMPATIBLE because CEED_ERROR_MINOR is not very descriptive.

operator - minor error for CEED_EVAL_WEIGHT + non-trivial restriction

Co-authored-by: Jeremy L. Thompson <jeremy.thompson@colorado.edu>
