### 2021-03-31 d5fada8
operator - use new error codes in full assembly
