+++ interface/ceed-operator.c
@@ -319,11 +319,11 @@ static int CeedOperatorGetActiveElemRestriction(CeedOperator op,
     int ierr;
     Ceed ceed;
     ierr = CeedOperatorGetCeed(op, &ceed); CeedChk(ierr);
-    return CeedError(ceed, 1,
+    return CeedError(ceed, CEED_ERROR_INCOMPLETE,
                      "No active ElemRestriction found!");
     // LCOV_EXCL_STOP
   }
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 /**
@@ -1327,7 +1327,8 @@ int CeedSingleOperatorAssembleSymbolic(CeedOperator op, CeedInt offset,
   Ceed ceed = op->ceed;
   if (op->composite) {
     // LCOV_EXCL_START
-    return CeedError(ceed, 1, "Not suitable for composite operator");
+    return CeedError(ceed, CEED_ERROR_UNSUPPORTED,
+                     "Composite operator not supported");
     // LCOV_EXCL_STOP
   }
 
@@ -1388,12 +1389,12 @@ int CeedSingleOperatorAssembleSymbolic(CeedOperator op, CeedInt offset,
   }
   if (count != local_nentries)
     // LCOV_EXCL_START
-    return CeedError(ceed, 1, "Bad calculation of entries!");
+    return CeedError(ceed, CEED_ERROR_MAJOR, "Error computing assembled entries");
   // LCOV_EXCL_STOP
   ierr = CeedVectorRestoreArrayRead(elem_dof, &elem_dof_a); CeedChk(ierr);
   ierr = CeedVectorDestroy(&elem_dof); CeedChk(ierr);
 
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 /**
@@ -1409,7 +1410,8 @@ int CeedSingleOperatorAssemble(CeedOperator op, CeedInt offset,
   Ceed ceed = op->ceed;;
   if (op->composite) {
     // LCOV_EXCL_START
-    return CeedError(ceed, 1, "Not suitable for composite operator");
+    return CeedError(ceed, CEED_ERROR_UNSUPPORTED,
+                     "Composite operator not supported");
     // LCOV_EXCL_STOP
   }
 
@@ -1571,7 +1573,7 @@ int CeedSingleOperatorAssemble(CeedOperator op, CeedInt offset,
                   gradin[(din*nqpts+q) * elemsize + n];
               } else {
                 // LCOV_EXCL_START
-                return CeedError(ceed, 1, "Not implemented!");
+                return CeedError(ceed, CEED_ERROR_UNSUPPORTED, "Not implemented!");
                 // LCOV_EXCL_STOP
               }
             }
@@ -1586,7 +1588,7 @@ int CeedSingleOperatorAssemble(CeedOperator op, CeedInt offset,
                   gradin[(dout*nqpts+q) * elemsize + n];
               } else {
                 // LCOV_EXCL_START
-                return CeedError(ceed, 1, "Not implemented!");
+                return CeedError(ceed, CEED_ERROR_UNSUPPORTED, "Not implemented!");
                 // LCOV_EXCL_STOP
               }
             }
@@ -1630,7 +1632,7 @@ int CeedSingleOperatorAssemble(CeedOperator op, CeedInt offset,
   }
   if (count != local_nentries)
     // LCOV_EXCL_START
-    return CeedError(ceed, 1, "Bad calculation of entries!");
[…]
