### 2021-03-31 38ed57a
operator - switch default fallback to none
