+++ interface/ceed-operator.c
@@ -169,14 +169,17 @@ static int CeedOperatorCheckField(Ceed ceed, CeedQFunctionField qfield,
 /**
   @brief Check if a CeedOperator is ready to be used.
 
-  @param[in] ceed Ceed object for error handling
   @param[in] op   CeedOperator to check
 
   @return An error code: 0 - success, otherwise - failure
 
   @ref Developer
 **/
-static int CeedOperatorCheckReady(Ceed ceed, CeedOperator op) {
+static int CeedOperatorCheckReady(CeedOperator op) {
+  int ierr;
+  Ceed ceed;
+  ierr = CeedOperatorGetCeed(op, &ceed); CeedChk(ierr);
+
   if (op->interfacesetup)
     return CEED_ERROR_SUCCESS;
 
@@ -1110,23 +1113,20 @@ int CeedOperatorLinearAssembleQFunction(CeedOperator op, CeedVector *assembled,
                                         CeedElemRestriction *rstr,
                                         CeedRequest *request) {
   int ierr;
-  Ceed ceed = op->ceed;
-  ierr = CeedOperatorCheckReady(ceed, op); CeedChk(ierr);
+  ierr = CeedOperatorCheckReady(op); CeedChk(ierr);
 
   // Backend version
   if (op->LinearAssembleQFunction) {
-    ierr = op->LinearAssembleQFunction(op, assembled, rstr, request);
-    CeedChk(ierr);
+    return op->LinearAssembleQFunction(op, assembled, rstr, request);
   } else {
     // Fallback to reference Ceed
     if (!op->opfallback) {
       ierr = CeedOperatorCreateFallback(op); CeedChk(ierr);
     }
     // Assemble
-    ierr = op->opfallback->LinearAssembleQFunction(op->opfallback, assembled,
-           rstr, request); CeedChk(ierr);
+    return CeedOperatorLinearAssembleQFunction(op->opfallback, assembled,
+           rstr, request);
   }
-  return CEED_ERROR_SUCCESS;
 }
 
 /**
@@ -1149,12 +1149,11 @@ int CeedOperatorLinearAssembleQFunction(CeedOperator op, CeedVector *assembled,
 int CeedOperatorLinearAssembleDiagonal(CeedOperator op, CeedVector assembled,
                                        CeedRequest *request) {
   int ierr;
-  Ceed ceed = op->ceed;
-  ierr = CeedOperatorCheckReady(ceed, op); CeedChk(ierr);
+  ierr = CeedOperatorCheckReady(op); CeedChk(ierr);
 
   // Use backend version, if available
   if (op->LinearAssembleDiagonal) {
-    ierr = op->LinearAssembleDiagonal(op, assembled, request); CeedChk(ierr);
+    return op->LinearAssembleDiagonal(op, assembled, request);
   } else if (op->LinearAssembleAddDiagonal) {
     ierr = CeedVectorSetValue(assembled, 0.0); CeedChk(ierr);
     return CeedOperatorLinearAssembleAddDiagonal(op, assembled, request);
@@ -1164,15 +1163,8 @@ int CeedOperatorLinearAssembleDiagonal(CeedOperator op, CeedVector assembled,
       ierr = CeedOperatorCreateFallback(op); CeedChk(ierr);
     }
     // Assemble
-    if (op->opfallback->LinearAssembleDiagonal) {
-      ierr = op->opfallback->LinearAssembleDiagonal(op->opfallback, assembled,
-             request); CeedChk(ierr);
-    } else {
[…]
