### 2021-03-31 de7534d
operator - fix fallback for full assembly
