### 2021-03-31 343d1d2
gitlab - set NPROC_GPU to min of nproc, 8
