### 2021-03-31 ab25172
gitlab - display number processes avaliable for tests (#706)
* gitlab - display number processes avaliable for tests

* gitlab - seprate CPU and GPU nproc

* gitlab -update nproc export

* gitlab - set NPROC_GPU to min of nproc, 8
