+++ .gitlab-ci.yml
@@ -9,6 +9,7 @@ noether-rocm:
   script:
 # Compilers
     - export COVERAGE=1 CC=gcc CXX=g++ FC=gfortran HIPCC=hipcc
+    - echo "---------- nproc-----------" && NPROC_CPU=$(nproc) && NPROC_GPU=$(($(nproc)<8?$(nproc):8))
     - echo "---------- CC -------------" && $CC --version
     - echo "---------- CXX ------------" && $CXX --version
     - echo "---------- FC -------------" && $FC --version
@@ -24,17 +25,21 @@ noether-rocm:
     - cd .. && export OCCA_VERSION=occa-1.1.0 OCCA_OPENCL_ENABLED=0 && { [[ -d $OCCA_VERSION ]] || { git clone --depth 1 --branch v1.1.0 https://github.com/libocca/occa.git $OCCA_VERSION && make -C $OCCA_VERSION -j$(nproc); }; } && make -C $OCCA_VERSION info && export OCCA_DIR=$PWD/$OCCA_VERSION && cd libCEED
 # libCEED
     - make info
-    - make -j$(nproc)
+    - BACKENDS_CPU=$(make info-backends | grep -o '/cpu[^ ]*') && BACKENDS_GPU=$(make info-backends | grep -o '/gpu[^ ]*')
+    - make -j$NPROC_CPU
     - echo '[{"subject":"/","metrics":[{"name":"Transfer Size (KB)","value":"19.5","desiredSize":"smaller"},{"name":"Speed Index","value":0,"desiredSize":"smaller"},{"name":"Total Score","value":92,"desiredSize":"larger"},{"name":"Requests","value":4,"desiredSize":"smaller"}]}]' > performance.json
-    - make -k -j$(nproc) junit realsearch=%
+    - make -k -j$NPROC_CPU BACKENDS="$BACKENDS_CPU" junit realsearch=%
+    - make -k -j$NPROC_GPU BACKENDS="$BACKENDS_GPU" junit realsearch=%
 # MFEM v4.2
     - cd .. && export MFEM_VERSION=mfem-4.2 && { [[ -d $MFEM_VERSION ]] || { git clone --depth 1 --branch v4.2 https://github.com/mfem/mfem.git $MFEM_VERSION && make -C $MFEM_VERSION -j$(nproc) serial CXXFLAGS="-O -std=c++11"; }; } && export MFEM_DIR=$PWD/$MFEM_VERSION && cd libCEED
     - make -C $MFEM_DIR info
-    - make -k -j$(nproc) junit search=mfem
+    - make -k -j$NPROC_CPU BACKENDS="$BACKENDS_CPU" junit search=mfem
+    - make -k -j$NPROC_GPU BACKENDS="$BACKENDS_GPU" junit search=mfem
 # Nek5000 v19.0
     - export COVERAGE=0
     - cd .. && export NEK5K_VERSION=Nek5000-19.0 && { [[ -d $NEK5K_VERSION ]] || { git clone --depth 1 --branch v19.0 https://github.com/Nek5000/Nek5000.git $NEK5K_VERSION && cd $NEK5K_VERSION/tools && ./maketools genbox genmap reatore2 && cd ../..; }; } && git -C $NEK5K_VERSION describe --tags && export NEK5K_DIR=$PWD/$NEK5K_VERSION && export PATH=$NEK5K_DIR/bin:$PATH MPI=0 && cd libCEED
-    - make -k -j$(nproc) junit search=nek
+    - make -k -j$NPROC_CPU BACKENDS="$BACKENDS_CPU" junit search=nek
+    - make -k -j$NPROC_GPU BACKENDS="$BACKENDS_GPU" junit search=nek
 # Report status
     - echo "SUCCESS" > .job_status
   after_script:
