### 2021-03-31 e6bb70c
operator - fix fallback usage
