### 2021-03-31 d48c403
operator - fix fallback for digonal assembly
