### 2021-03-31 dc8efd8
Fluids - More Regression Tests (#687)
* examples/fluids: Regression tests: Changed the reference files' names and
the flags to associate with the corresponding example (so far
density_current only)

Changed the files' names

Changed the flags

* examples/fluids: added new regression tests for advection(2d)

* examples/fluids: regression tests: removed the enums and passed defined command-line options for filepath and test tolerance.
Removed solver controls and assigned testtol=1E-11 for explicit cases.

style

examples/fluids: regression tests: fixed the pathfile issue (it needs to be relative
[…]
