+++ backends/avx/ceed-avx-blocked.c
@@ -36,7 +36,7 @@ static int CeedInit_Avx(const char *resource, Ceed ceed) {
   // Create reference CEED that implementation will be dispatched
   //   through unless overridden
   Ceed ceedref;
-  CeedInit("/cpu/self/ref/blocked", &ceedref);
+  CeedInit("/cpu/self/opt/blocked", &ceedref);
   ierr = CeedSetDelegate(ceed, ceedref); CeedChkBackend(ierr);
 
   ierr = CeedSetBackendFunction(ceed, "Ceed", ceed, "TensorContractCreate",
