### 2021-03-31 e2f0418
Full assembly of CeedOperator (#712)
* Add CeedOperatorLinearFullAssemble() to explicitly assemble a CeedOperator

The output format is a stream of (row, column, value) triples, which client
code can use to assemble a sparse matrix by adding the values in the
appropriate places.

* Add full assembly test for triangular mesh, t563-operator.c, which is analogous to t536-operator.c

* assembly: change interface to separate nonzero structure from values

* assembly: add potential backends for CeedOperatorLinearAssemble, CeedOperatorLinearAssemblySymbolic

* assembly: use CeedMatrixMultiply() inste
[…]
