+++ tests/t565-operator.c
@@ -1,0 +1,206 @@
+/// @file
+/// Test full assembly of composite operator (see t538)
+/// \test Test full assembly of composite operator
+#include <ceed.h>
+#include <stdlib.h>
+#include <math.h>
+
+int main(int argc, char **argv) {
+  Ceed ceed;
+  CeedElemRestriction Erestrictx, Erestrictu,
+                      Erestrictui, ErestrictqiMass, ErestrictqiDiff;
+  CeedBasis bx, bu;
+  CeedQFunction qf_setupMass, qf_mass, qf_setupDiff, qf_diff;
+  CeedOperator op_setupMass, op_mass, op_setupDiff, op_diff, op_apply;
+  CeedVector qdataMass, qdataDiff, X, U, V;
+  CeedInt P = 3, Q = 4, dim = 2;
+  CeedInt nx = 3, ny = 2;
+  CeedInt nelem = nx * ny;
+  CeedInt ndofs = (nx*2+1)*(ny*2+1), nqpts = nelem*Q*Q;
+  CeedInt indx[nelem*P*P];
+  CeedScalar assembled[ndofs*ndofs];
+  CeedScalar x[dim*ndofs], assembledTrue[ndofs*ndofs];
+  CeedScalar *u;
+  const CeedScalar *v;
+
+  CeedInit(argv[1], &ceed);
+
+  // DoF Coordinates
+  for (CeedInt i=0; i<nx*2+1; i++)
+    for (CeedInt j=0; j<ny*2+1; j++) {
+      x[i+j*(nx*2+1)+0*ndofs] = (CeedScalar) i / (2*nx);
+      x[i+j*(nx*2+1)+1*ndofs] = (CeedScalar) j / (2*ny);
+    }
+  CeedVectorCreate(ceed, dim*ndofs, &X);
+  CeedVectorSetArray(X, CEED_MEM_HOST, CEED_USE_POINTER, x);
+
+  // Qdata Vectors
+  CeedVectorCreate(ceed, nqpts, &qdataMass);
+  CeedVectorCreate(ceed, nqpts*dim*(dim+1)/2, &qdataDiff);
+
+  // Element Setup
+  for (CeedInt i=0; i<nelem; i++) {
+    CeedInt col, row, offset;
+    col = i % nx;
+    row = i / nx;
+    offset = col*(P-1) + row*(nx*2+1)*(P-1);
+    for (CeedInt j=0; j<P; j++)
+      for (CeedInt k=0; k<P; k++)
+        indx[P*(P*i+k)+j] = offset + k*(nx*2+1) + j;
+  }
+
+  // Restrictions
+  CeedElemRestrictionCreate(ceed, nelem, P*P, dim, ndofs, dim*ndofs,
+                            CEED_MEM_HOST, CEED_USE_POINTER, indx, &Erestrictx);
+
+  CeedElemRestrictionCreate(ceed, nelem, P*P, 1, 1, ndofs, CEED_MEM_HOST,
+                            CEED_USE_POINTER, indx, &Erestrictu);
+  CeedInt stridesu[3] = {1, Q*Q, Q*Q};
+  CeedElemRestrictionCreateStrided(ceed, nelem, Q*Q, 1, nqpts, stridesu,
+                                   &Erestrictui);
+
+  CeedInt stridesqdMass[3] = {1, Q*Q, Q*Q};
+  CeedElemRestrictionCreateStrided(ceed, nelem, Q*Q, 1, nqpts,
+                                   stridesqdMass, &ErestrictqiMass);
+  CeedInt stridesqdDiff[3] = {1, Q*Q, Q*Q*dim*(dim+1)/2}; /* *NOPAD* */
[…]
