### 2021-03-31 7b46028
doc - improve release notes (#719)
