### 2021-03-31 cd1c839
separate kershaw function
