### 2021-03-31 5857240
examples/petsc/multigrid: error handling and style fix
