### 2021-03-31 cfa59c5
examples/petsc: Kershaw mesh transformation (#715)
* add kershaw transformation

* separate kershaw function

* examples/petsc/multigrid: error handling and style fix

Co-authored-by: Jed Brown <jed@jedbrown.org>
