### 2021-03-31 aa8cce7
make configure: add OCCA_DIR
