### 2021-04-01 952aa24
Makefile: fix typo
Reported-by: Yohann Dudouit <yohann.dudouit@gmail.com>
