### 2021-04-01 efe57be
Update doc/sphinx/source/releasenotes.rst
Co-authored-by: Jed Brown <jed@jedbrown.org>
