+++ doc/sphinx/source/releasenotes.rst
@@ -32,7 +32,7 @@ Examples
 * :ref:`example-petsc-elasticity` example updated with traction boundary conditions and improved Dirichlet boundary conditions.
 * :ref:`example-petsc-elasticity` example updated with Neo-Hookean hyperelasticity in current configuration as well as improved Neo-Hookean hyperelasticity exploring storage vs computation tradeoffs.
 * :ref:`example-petsc-navier-stokes` example updated with Traveling Vortex test case which is the analytical solution to the Euler equations.
-* :ref:`example-petsc-navier-stokes` example updated with a support for performing convergence study and plotting the obtained results.
+* :ref:`example-petsc-navier-stokes` example updated with support for performing convergence study and plotting order of convergence by polynomial degree.
 
 .. _v0.7:
 
@@ -548,4 +548,3 @@ Examples available in this release:
 +-------------------------+-----------------------------------+
 | ``petsc``               | BP1 (scalar mass operator)        |
 +-------------------------+-----------------------------------+
-
