### 2021-04-01 288c66a
Update doc/sphinx/source/releasenotes.rst
Co-authored-by: Jed Brown <jed@jedbrown.org>
