### 2021-04-01 f45e5ea
update release notes (#723)
* doc: update release notes

Co-authored-by: Jed Brown <jed@jedbrown.org>
