### 2021-04-01 80132a9
tests/python: fix double free remnant after switch to context manager
