+++ rust/Cargo.toml
@@ -6,7 +6,7 @@ authors = [
 ]
 build = "build.rs"
 name = "libceed"
-version = "0.7.0"
+version = "0.8.0"
 links = "libceed"
 edition = "2018"
 
