### 2021-04-05 466791e
Merge main after v0.8 (#723)
