+++ examples/fluids/src/euler-vortex.c
@@ -1,178 +1,149 @@
 #include "../navierstokes.h"
 
-PetscErrorCode NS_ADVECTION2D(problemData *problem, void *ctxSetupData,
-                              void *ctx, void *ctxPhys) {
+PetscErrorCode NS_EULER_VORTEX(problemData *problem, void *ctxSetupData,
+                               void *ctx, void *ctxPhys) {
   PetscInt ierr;
   MPI_Comm comm = PETSC_COMM_WORLD;
-  WindType wind_type;
-  StabilizationType stab;
+  EulerTestType eulertest;
   PetscBool implicit;
-  PetscBool hasCurrentTime = PETSC_FALSE;
+  PetscBool hasCurrentTime = PETSC_TRUE;
+  PetscBool hasNeumann = PETSC_TRUE;
   SetupContext ctxSetup = *(SetupContext *)ctxSetupData;
   Units units = *(Units *)ctx;
   Physics ctxPhysData = *(Physics *)ctxPhys;
-  ierr = PetscMalloc1(1, &ctxPhysData->ctxAdvectionData); CHKERRQ(ierr);
+  ierr = PetscMalloc1(1, &ctxPhysData->ctxEulerData); CHKERRQ(ierr);
+
 
   PetscFunctionBeginUser;
   // ------------------------------------------------------
-  //               SET UP ADVECTION2D
+  //               SET UP DENSITY_CURRENT
   // ------------------------------------------------------
-  problem->dim                       = 2;
-  problem->qdatasizeVol              = 5;
-  problem->qdatasizeSur              = 3;
-  problem->setupVol                  = Setup2d;
-  problem->setupVol_loc              = Setup2d_loc;
-  problem->setupSur                  = SetupBoundary2d;
-  problem->setupSur_loc              = SetupBoundary2d_loc;
-  problem->ics                       = ICsAdvection2d;
-  problem->ics_loc                   = ICsAdvection2d_loc;
-  problem->applyVol_rhs              = Advection2d;
-  problem->applyVol_rhs_loc          = Advection2d_loc;
-  problem->applyVol_ifunction        = IFunction_Advection2d;
-  problem->applyVol_ifunction_loc    = IFunction_Advection2d_loc;
-  problem->applySur                  = Advection2d_Sur;
-  problem->applySur_loc              = Advection2d_Sur_loc;
-  problem->bc                        = Exact_Advection2d;
-  problem->bc_fnc                    = BC_ADVECTION2D;
+  problem->dim                       = 3;
+  problem->qdatasizeVol              = 10;
+  problem->qdatasizeSur              = 4;
+  problem->setupVol                  = Setup;
+  problem->setupVol_loc              = Setup_loc;
+  problem->setupSur                  = SetupBoundary;
+  problem->setupSur_loc              = SetupBoundary_loc;
+  problem->ics                       = ICsEuler;
+  problem->ics_loc                   = ICsEuler_loc;
+  problem->applyVol_rhs              = Euler;
+  problem->applyVol_rhs_loc          = Euler_loc;
+  problem->applyVol_ifunction        = IFunction_Euler;
+  problem->applyVol_ifunction_loc    = IFunction_Euler_loc;
+  problem->applySur                  = Euler_Sur;
+  problem->applySur_loc              = Euler_Sur_loc;
+  problem->bc                        = Exact_Euler;
+  problem->bc_fnc                    = BC_EULER_VORTEX;
   problem->non_zero_time             = PETSC_TRUE;
 
   // ------------------------------------------------------
   //             Create the libCEED context
[…]
