### 2021-04-05 6dbfb41
Update device ID selection for HIP/CUDA backends; add for MAGMA backends
