+++ backends/magma/ceed-magma-det.c
@@ -16,17 +16,37 @@
 
 #include <ceed/ceed.h>
 #include <ceed/backend.h>
+#include <string.h>
+#include <stdlib.h>
 #include "ceed-magma.h"
 
 CEED_INTERN int CeedInit_Magma_Det(const char *resource, Ceed ceed) {
   int ierr;
-  if (strcmp(resource, "/gpu/cuda/magma/det")
-      && strcmp(resource, "/gpu/hip/magma/det"))
+  const int nrc = 18; // number of characters in resource
+  if (strncmp(resource, "/gpu/cuda/magma/det", nrc)
+      && strncmp(resource, "/gpu/hip/magma/det", nrc))
     // LCOV_EXCL_START
     return CeedError(ceed, CEED_ERROR_BACKEND,
                      "Magma backend cannot use resource: %s", resource);
   // LCOV_EXCL_STOP
   ierr = CeedSetDeterministic(ceed, true); CeedChkBackend(ierr);
+
+  Ceed_Magma *data;
+  ierr = CeedCalloc(sizeof(Ceed_Magma), &data); CeedChkBackend(ierr);
+  ierr = CeedSetData(ceed, data); CeedChkBackend(ierr);
+
+  // get/set device ID
+  const char *device_spec = strstr(resource, ":device_id=");
+  const int deviceID = (device_spec) ? atoi(device_spec+11) : -1;
+
+  int currentDeviceID;
+  magma_getdevice(&currentDeviceID);
+  if (deviceID >= 0 && currentDeviceID != deviceID) {
+    magma_setdevice(deviceID);
+    currentDeviceID = deviceID;
+  }
+  // create a queue that uses the null stream
+  data->device = currentDeviceID;
 
   // Create reference CEED that implementation will be dispatched
   //   through unless overridden
