### 2021-04-06 e83e87a
petsc - refactor examples to reduce redundant code
