+++ examples/petsc/multigrid.c
@@ -38,8 +38,7 @@
 /// CEED BPs 1-6 multigrid example using PETSc
 const char help[] = "Solve CEED BPs using p-multigrid with PETSc and DMPlex\n";
 
-#define multigrid
-#include "setup.h"
+#include "bps.h"
 
 // Transition from a value of "a" for x=0, to a value of "b" for x=1.  Optionally
 // smooth -- see the commented versions at the end.
@@ -278,7 +277,10 @@ int main(int argc, char **argv) {
     ierr = DMClone(dmorig, &dm[i]); CHKERRQ(ierr);
     ierr = DMGetVecType(dmorig, &vectype); CHKERRQ(ierr);
     ierr = DMSetVecType(dm[i], vectype); CHKERRQ(ierr);
-    ierr = SetupDMByDegree(dm[i], leveldegrees[i], ncompu, bpchoice);
+    PetscInt dim;
+    ierr = DMGetDimension(dm[i], &dim); CHKERRQ(ierr);
+    ierr = SetupDMByDegree(dm[i], leveldegrees[i], ncompu, dim,
+                           bpOptions[bpchoice].enforcebc, bpOptions[bpchoice].bcsfunc);
     CHKERRQ(ierr);
 
     // Create vectors
@@ -366,7 +368,7 @@ int main(int argc, char **argv) {
     }
     ierr = PetscMalloc1(1, &ceeddata[i]); CHKERRQ(ierr);
     ierr = SetupLibceedByDegree(dm[i], ceed, leveldegrees[i], dim, qextra,
-                                ncompu, gsize[i], xlsize[i], bpchoice,
+                                dim, ncompu, gsize[i], xlsize[i], bpOptions[bpchoice],
                                 ceeddata[i], i==(fineLevel), rhsceed, &target);
     CHKERRQ(ierr);
   }
@@ -385,9 +387,8 @@ int main(int argc, char **argv) {
                               &qfprolong);
 
   // Set up libCEED level transfer operators
-  ierr = CeedLevelTransferSetup(ceed, numlevels, ncompu, bpchoice, ceeddata,
-                                leveldegrees, qfrestrict, qfprolong);
-  CHKERRQ(ierr);
+  ierr = CeedLevelTransferSetup(ceed, numlevels, ncompu, ceeddata, leveldegrees,
+                                qfrestrict, qfprolong); CHKERRQ(ierr);
 
   // Create the error QFunction
   CeedQFunctionCreateInterior(ceed, 1, bpOptions[bpchoice].error,
@@ -413,12 +414,12 @@ int main(int argc, char **argv) {
     // CEED vector
     ierr = VecZeroEntries(Xloc[i]); CHKERRQ(ierr);
     ierr = VecGetArray(Xloc[i], &x); CHKERRQ(ierr);
-    CeedVectorSetArray(ceeddata[i]->xceed, CEED_MEM_HOST, CEED_USE_POINTER, x);
+    CeedVectorSetArray(ceeddata[i]->Xceed, CEED_MEM_HOST, CEED_USE_POINTER, x);
 
     // Multiplicity
     CeedElemRestrictionGetMultiplicity(ceeddata[i]->Erestrictu,
-                                       ceeddata[i]->xceed);
-    CeedVectorSyncArray(ceeddata[i]->xceed, CEED_MEM_HOST);
+                                       ceeddata[i]->Xceed);
+    CeedVectorSyncArray(ceeddata[i]->Xceed, CEED_MEM_HOST);
 
     // Restore vector
[…]
