+++ examples/petsc/qfunctions/bps/bp4sphere.h
@@ -24,10 +24,8 @@
 #  include <math.h>
 #endif
 
-// *****************************************************************************
+// -----------------------------------------------------------------------------
 // This QFunction sets up the rhs and true solution for the problem
-// *****************************************************************************
-
 // -----------------------------------------------------------------------------
 CEED_QFUNCTION(SetupDiffRhs3)(void *ctx, const CeedInt Q,
                              const CeedScalar *const *in,
@@ -74,7 +72,7 @@ CEED_QFUNCTION(SetupDiffRhs3)(void *ctx, const CeedInt Q,
   return 0;
 }
 
-// *****************************************************************************
+// -----------------------------------------------------------------------------
 // This QFunction applies the diffusion operator for a vector field of 3 components.
 //
 // Inputs:
@@ -84,8 +82,6 @@ CEED_QFUNCTION(SetupDiffRhs3)(void *ctx, const CeedInt Q,
 // Output:
 //   vJ     - Output vector (test functions) Jacobian at quadrature points
 //
-// *****************************************************************************
-
 // -----------------------------------------------------------------------------
 CEED_QFUNCTION(Diff3)(void *ctx, const CeedInt Q,
                       const CeedScalar *const *in, CeedScalar *const *out) {
