+++ examples/fluids/navierstokes.c
@@ -39,6 +39,11 @@
 //TESTARGS(name="adv2d_rotation_explicit_strong") -ceed {ceed_resource} -test -problem advection2d -strong_form 1 -degree 3 -dm_plex_box_faces 1,1,2 -units_kilogram 1e-9 -lx 125 -ly 125 -lz 250 -rc 100. -ts_dt 1e-3 -compare_final_state_atol 1E-11 -compare_final_state_filename examples/fluids/tests-output/fluids-navierstokes-adv2d-rotation-explicit-strong.bin
 //TESTARGS(name="adv2d_rotation_implicit_stab_supg") -ceed {ceed_resource} -test -problem advection2d -CtauS .3 -stab supg -degree 3 -dm_plex_box_faces 1,1,2 -units_kilogram 1e-9 -lx 125 -ly 125 -lz 250 -rc 100. -ksp_atol 1e-4 -ksp_rtol 1e-3 -ksp_type bcgs -snes_atol 1e-3 -snes_lag_jacobian 100 -snes_lag_jacobian_persists -snes_mf_operator -ts_dt 1e-3 -implicit -ts_type alpha -compare_final_state_atol 5E-4 -compare_final_state_filename examples/fluids/tests-output/fluids-navierstokes-adv2d-rotation-implicit-stab-supg.bin
 //TESTARGS(name="adv2d_translation_implicit_stab_su") -ceed {ceed_resource} -test -problem advection2d -CtauS .3 -stab su -degree 3 -dm_plex_box_faces 1,1,2 -units_kilogram 1e-9 -lx 125 -ly 125 -lz 250 -rc 100. -ksp_atol 1e-4 -ksp_rtol 1e-3 -ksp_type bcgs -snes_atol 1e-3 -snes_lag_jacobian 100 -snes_lag_jacobian_persists -snes_mf_operator -ts_dt 1e-3 -implicit -ts_type alpha -problem_advection_wind translation -problem_advection_wind_translation .53,-1.33,0 -compare_final_state_atol 5E-4 -compare_final_state_filename examples/fluids/tests-output/fluids-navierstokes-adv2d-translation-implicit-stab-su.bin
+//TESTARGS(name="euler_t1") -ceed {ceed_resource} -test -problem euler_vortex -dm_plex_box_faces 1,1,2 -units_meter 1e-3 -lx 125 -ly 125 -lz 1 -ts_dt 1e-3 -problem_euler_mean_velocity -.4,1.6,0 -vortex_strength 3 -euler_test t1 -compare_final_state_atol 1E-11 -compare_final_state_filename examples/fluids/tests-output/fluids-navierstokes-euler-t1.bin
+//TESTARGS(name="euler_t2") -ceed {ceed_resource} -test -problem euler_vortex -dm_plex_box_faces 1,1,2 -units_meter 1e-3 -lx 125 -ly 125 -lz 1 -ts_dt 1e-3 -problem_euler_mean_velocity -.4,1.6,0 -vortex_strength 3 -euler_test t2 -compare_final_state_atol 1E-11 -compare_final_state_filename examples/fluids/tests-output/fluids-navierstokes-euler-t2.bin
+//TESTARGS(name="euler_t3") -ceed {ceed_resource} -test -problem euler_vortex -dm_plex_box_faces 1,1,2 -units_meter 1e-3 -lx 125 -ly 125 -lz 1 -ts_dt 1e-3 -problem_euler_mean_velocity -.4,1.6,0 -vortex_strength 3 -euler_test t3 -compare_final_state_atol 1E-11 -compare_final_state_filename examples/fluids/tests-output/fluids-navierstokes-euler-t3.bin
[…]
