### 2021-04-06 2bbc7fe
interface - improved resource matching check
