+++ tests/tap.sh
@@ -124,6 +124,15 @@ for ((i=0;i<${#backends[@]};++i)); do
         continue
     fi
 
+    # grep to pass test t006, t007 on error
+    if grep -F -q -e 'No suitable backend:' ${output}.err \
+            && [[ "$1" = "t006"* || "$1" = "t007"* ]] ; then
+        printf "ok $i0 PASS - expected failure $1 $backend\n"
+        printf "ok $i1 PASS - expected failure $1 $backend stdout\n"
+        printf "ok $i2 PASS - expected failure $1 $backend stderr\n"
+        continue
+    fi
+
     # grep to pass test t11* on error
     if grep -F -q -e 'access' ${output}.err \
             && [[ "$1" = "t11"* ]] ; then
