+++ tests/t007-ceed.c
@@ -9,8 +9,10 @@ int main(int argc, char **argv) {
   size_t end_index ;
   for (end_index = 0; argv[1][end_index]; end_index++) {}
   argv[1][end_index - 1] -= 1;
-  CeedInit(argv[1], &ceed);
-  CeedDestroy(&ceed);
 
+  CeedInit(argv[1], &ceed);
+  // LCOV_EXCL_START
+  CeedDestroy(&ceed);
   return 0;
+  // LCOV_EXCL_STOP
 }
