### 2021-04-07 22e4421
interface - add help option for CeedInit and function to access arrays of resources + priorities
