+++ tests/t007-ceed.c
@@ -6,7 +6,7 @@
 int main(int argc, char **argv) {
   Ceed ceed;
 
-  size_t end_index ;
+  size_t end_index;
   for (end_index = 0; argv[1][end_index]; end_index++) {}
   argv[1][end_index - 1] -= 1;
 
