+++ tests/tap.sh
@@ -133,6 +133,15 @@ for ((i=0;i<${#backends[@]};++i)); do
         continue
     fi
 
+    # grep to pass test t008 on output in stderr
+    if grep -F -q -e 'Avaliable backend resources:' ${output}.err \
+            && [[ "$1" = "t008"* ]] ; then
+        printf "ok $i0 PASS - expected failure $1 $backend\n"
+        printf "ok $i1 PASS - expected failure $1 $backend stdout\n"
+        printf "ok $i2 PASS - expected failure $1 $backend stderr\n"
+        continue
+    fi
+
     # grep to pass test t11* on error
     if grep -F -q -e 'access' ${output}.err \
             && [[ "$1" = "t11"* ]] ; then
