### 2021-04-07 9b07255
style - update to new style for petsc examples
