+++ examples/petsc/multigrid.c
@@ -44,30 +44,30 @@ int main(int argc, char **argv) {
   PetscInt ierr;
   MPI_Comm comm;
   char filename[PETSC_MAX_PATH_LEN],
-       ceedresource[PETSC_MAX_PATH_LEN] = "/cpu/self";
+       ceed_resource[PETSC_MAX_PATH_LEN] = "/cpu/self";
   double my_rt_start, my_rt, rt_min, rt_max;
-  PetscInt degree = 3, qextra, *lsize, *xlsize, *gsize, dim = 3, fineLevel,
-           melem[3] = {3, 3, 3}, ncompu = 1, numlevels = degree, *leveldegrees;
+  PetscInt degree = 3, q_extra, *l_size, *xl_size, *g_size, dim = 3, fine_level,
+           mesh_elem[3] = {3, 3, 3}, num_comp_u = 1, num_levels = degree, *level_degrees;
   PetscScalar *r;
   PetscScalar eps = 1.0;
   PetscBool test_mode, benchmark_mode, read_mesh, write_solution;
-  PetscLogStage solvestage;
-  DM  *dm, dmorig;
-  SNES snesdummy;
+  PetscLogStage solve_stage;
+  DM  *dm, dm_orig;
+  SNES snes_dummy;
   KSP ksp;
   PC pc;
-  Mat *matO, *matPR, matcoarse;
-  Vec *X, *Xloc, *mult, rhs, rhsloc;
-  PetscMemType memtype;
-  UserO *userO;
-  UserProlongRestr *userPR;
+  Mat *mat_O, *mat_pr, mat_coarse;
+  Vec *X, *X_loc, *mult, rhs, rhs_loc;
+  PetscMemType mem_type;
+  UserO *user_O;
+  UserProlongRestr *user_pr;
   Ceed ceed;
-  CeedData *ceeddata;
-  CeedVector rhsceed, target;
-  CeedQFunction qferror, qfrestrict, qfprolong;
-  CeedOperator operror;
-  bpType bpchoice;
-  coarsenType coarsen;
+  CeedData *ceed_data;
+  CeedVector rhs_ceed, target;
+  CeedQFunction qf_error, qf_restrict, qf_prolong;
+  CeedOperator op_error;
+  BPType bp_choice;
+  CoarsenType coarsen;
 
   ierr = PetscInitialize(&argc, &argv, NULL, help);
   if (ierr) return ierr;
@@ -75,12 +75,12 @@ int main(int argc, char **argv) {
 
   // Parse command line options
   ierr = PetscOptionsBegin(comm, NULL, "CEED BPs in PETSc", NULL); CHKERRQ(ierr);
-  bpchoice = CEED_BP3;
+  bp_choice = CEED_BP3;
   ierr = PetscOptionsEnum("-problem",
                           "CEED benchmark problem to solve", NULL,
-                          bpTypes, (PetscEnum)bpchoice, (PetscEnum *)&bpchoice,
+                          bp_types, (PetscEnum)bp_choice, (PetscEnum *)&bp_choice,
                           NULL); CHKERRQ(ierr);
-  ncompu = bpOptions[bpchoice].ncompu;
+  num_comp_u = bp_options[bp_choice].num_comp_u;
   test_mode = PETSC_FALSE;
   ierr = PetscOptionsBool("-test",
                           "Testing mode (do not print unless error is large)",
@@ -105,16 +105,16 @@ int main(int argc, char **argv) {
                          NULL, degree, &degree, NULL); CHKERRQ(ierr);
   if (degree < 1) SETERRQ1(PETSC_COMM_WORLD, PETSC_ERR_ARG_OUTOFRANGE,
[…]
