### 2021-04-07 9ff8684
interface - update CeedRegistryGetList to use malloc directly, make priorities array optional
