+++ examples/solids/qfunctions/hyperFS.h
@@ -182,49 +182,19 @@ CeedScalar GP_energyModel(void *ctx, CeedScalar detC_m1, CeedScalar E2[][3], Cee
 }
 // -----------------------------------------------------------------------------
 // -----------------------------------------------------------------------------
-// Energy derivations S for models CALL IN COMMONFS WHERE COMMENTED IN
+// Energy derivations S for models; PASS IN FOR SWORK_FUNC IN COMMONFS_GENERIC
 // -----------------------------------------------------------------------------
 // Neo-Hookean model
-CeedScalar NH_2nd_PK(){
-  return 0;
-}
-// -----------------------------------------------------------------------------
-// Mooney-Rivlin model
-CeedScalar MR_2nd_PK(){
-  // 
-  return 0;
-}
-// -----------------------------------------------------------------------------
-// Generalized Polynomial model
-CeedScalar GP_2nd_PK(){
-  return 0;
-}
-// -----------------------------------------------------------------------------
-// -----------------------------------------------------------------------------
-
-// -----------------------------------------------------------------------------
-// Common computations between FS and dFS
-// -----------------------------------------------------------------------------
-static inline int commonFS(const CeedScalar lambda, const CeedScalar mu,
-                           const CeedScalar gradu[3][3], CeedScalar Swork[6],
-                           CeedScalar Cinvwork[6], CeedScalar *detC_m1,
-                           CeedScalar *llnj) {
-  // E - Green-Lagrange strain tensor
-  //     E = 1/2 (gradu + gradu^T + gradu^T*gradu)
-  const CeedInt indj[6] = {0, 1, 2, 1, 0, 0}, indk[6] = {0, 1, 2, 2, 2, 1};
-  CeedScalar E2work[6];
-  for (CeedInt m = 0; m < 6; m++) {
-    E2work[m] = gradu[indj[m]][indk[m]] + gradu[indk[m]][indj[m]];
-    for (CeedInt n = 0; n < 3; n++)
-      E2work[m] += gradu[n][indj[m]]*gradu[n][indk[m]];
-  }
-  // *INDENT-OFF*
-  CeedScalar E2[3][3] = {{E2work[0], E2work[5], E2work[4]},
-                         {E2work[5], E2work[1], E2work[3]},
-                         {E2work[4], E2work[3], E2work[2]}
-                        };
-  // *INDENT-ON*
-  (*detC_m1) = computeDetCM1(E2work);
+CeedScalar NH_2nd_PK(void *ctx, CeedScalar Swork[6], CeedScalar Cinvwork[6], CeedScalar *detC_m1,
+                           CeedScalar *llnj, const CeedInt indj[6], const CeedInt indk[6], CeedScalar E2[][3]){
+  // unpack ctx here
+  const Physics context = (Physics)ctx;
+  const CeedScalar E  = context->E;
+  const CeedScalar nu = context->nu;
+  const CeedScalar TwoMu = E / (1 + nu);
+  const CeedScalar mu = TwoMu / 2;
+  const CeedScalar Kbulk = E / (3*(1 - 2*nu)); // Bulk Modulus
+  const CeedScalar lambda = (3*Kbulk - TwoMu) / 3;
 
[…]
