+++ tests/junit.py
@@ -101,7 +101,7 @@ def run(test, backends):
                 if test[:4] in 't006 t007'.split():
                     check_required_failure(case, proc.stderr, 'No suitable backend:')
                 if test[:4] in 't008'.split():
-                    check_required_failure(case, proc.stderr, 'Avaliable backend resources:')
+                    check_required_failure(case, proc.stderr, 'Available backend resources:')
                 if test[:4] in 't110 t111 t112 t113 t114'.split():
                     check_required_failure(case, proc.stderr, 'Cannot grant CeedVector array access')
                 if test[:4] in 't115'.split():
