+++ tests/tap.sh
@@ -134,7 +134,7 @@ for ((i=0;i<${#backends[@]};++i)); do
     fi
 
     # grep to pass test t008 on output in stderr
-    if grep -F -q -e 'Avaliable backend resources:' ${output}.err \
+    if grep -F -q -e 'Available backend resources:' ${output}.err \
             && [[ "$1" = "t008"* ]] ; then
         printf "ok $i0 PASS - expected failure $1 $backend\n"
         printf "ok $i1 PASS - expected failure $1 $backend stdout\n"
