### 2021-04-07 71415c5
build - register exact names of backends in CeedRegister_Weak for better error messages
