+++ backends/ceed-backend-weak.c
@@ -1,17 +1,33 @@
 #include <ceed/backend.h>
+#include <limits.h>
 #include <stdlib.h>
 #include <stdio.h>
 
-// This function provides a debug target for weak symbols
 // LCOV_EXCL_START
-static int CeedRegister_Weak(const char *name) {
+// This function provides improved error messages for uncompiled backends
+static int CeedInit_Weak(const char *resource, Ceed ceed) {
+  return CeedError(ceed, CEED_ERROR_UNSUPPORTED,
+                   "Backend not currently compiled: %s", resource);
+}
+
+// This function provides a debug target for weak symbols
+static int CeedRegister_Weak(const char *name, int num_prefixes, ...) {
   if (getenv("CEED_DEBUG")) fprintf(stderr, "Weak %s\n", name);
+
+  va_list prefixes;
+  va_start(prefixes, num_prefixes);
+  int ierr;
+  for (int i=0; i<num_prefixes; i++) {
+    ierr = CeedRegister(va_arg(prefixes, const char*), CeedInit_Weak, UINT_MAX);
+    CeedChk(ierr);
+  }
+  va_end(prefixes);
   return CEED_ERROR_SUCCESS;
 }
 // LCOV_EXCL_STOP
 
-#define MACRO(name)                                                     \
+#define MACRO(name,num_prefixes, ...)                                                     \
   CEED_INTERN int name(void) __attribute__((weak));                     \
-  int name(void) { return CeedRegister_Weak(__func__); }
+  int name(void) { return CeedRegister_Weak(__func__,num_prefixes, ## __VA_ARGS__); }
 #include "ceed-backend-list.h"
 #undef MACRO
