+++ examples/solids/src/setuplibceed.c
@@ -69,13 +69,13 @@ problemData problemOptions[5] = {
   [ELAS_HYPER_FS] = { 
     .qdatasize = 10,
     .setupgeo = SetupGeo,
-    .apply = HyperFSF,
-    .jacob = HyperFSdF,
+    .apply = HyperFSF_NH,
+    .jacob = HyperFSdF_NH,
     .energy = HyperFSEnergy_NH,
     .diagnostic = HyperFSDiagnostic,
     .setupgeofname = SetupGeo_loc,
-    .applyfname = HyperFSF_loc,
-    .jacobfname = HyperFSdF_loc,
+    .applyfname = HyperFSF_NH_loc,
+    .jacobfname = HyperFSdF_NH_loc,
     .energyfname = HyperFSEnergy_NH_loc,
     .diagnosticfname = HyperFSDiagnostic_loc,
     .qmode = CEED_GAUSS
@@ -84,13 +84,13 @@ problemData problemOptions[5] = {
   [ELAS_HYPER_FS_MR] = { 
     .qdatasize = 10,
     .setupgeo = SetupGeo,
-    .apply = HyperFSF,
-    .jacob = HyperFSdF,
+    .apply = HyperFSF_MR,
+    .jacob = HyperFSdF_MR,
     .energy = HyperFSEnergy_MR,
     .diagnostic = HyperFSDiagnostic,
     .setupgeofname = SetupGeo_loc,
-    .applyfname = HyperFSF_loc,
-    .jacobfname = HyperFSdF_loc,
+    .applyfname = HyperFSF_MR_loc,
+    .jacobfname = HyperFSdF_MR_loc,
     .energyfname = HyperFSEnergy_MR_loc,
     .diagnosticfname = HyperFSDiagnostic_loc,
     .qmode = CEED_GAUSS
@@ -99,13 +99,13 @@ problemData problemOptions[5] = {
   [ELAS_HYPER_FS_GP] = { 
     .qdatasize = 10,
     .setupgeo = SetupGeo,
-    .apply = HyperFSF,
-    .jacob = HyperFSdF,
+    .apply = HyperFSF_GP,
+    .jacob = HyperFSdF_GP,
     .energy = HyperFSEnergy_GP,
     .diagnostic = HyperFSDiagnostic,
     .setupgeofname = SetupGeo_loc,
-    .applyfname = HyperFSF_loc,
-    .jacobfname = HyperFSdF_loc,
+    .applyfname = HyperFSF_GP_loc,
+    .jacobfname = HyperFSdF_GP_loc,
     .energyfname = HyperFSEnergy_GP_loc,
     .diagnosticfname = HyperFSDiagnostic_loc,
     .qmode = CEED_GAUSS
