+++ examples/solids/qfunctions/hyperFS.h
@@ -185,7 +185,7 @@ CeedScalar GP_energyModel(void *ctx, CeedScalar detC_m1, CeedScalar E2[][3], Cee
 // Energy derivations S for models; PASS IN FOR SWORK_FUNC IN COMMONFS_GENERIC
 // -----------------------------------------------------------------------------
 // Neo-Hookean model
-CeedScalar NH_2nd_PK(void *ctx, CeedScalar Swork[6], CeedScalar Cinvwork[6], CeedScalar *detC_m1,
+CeedScalar NH_2nd_PK(void *ctx, CeedScalar Swork[6], CeedScalar Cinvwork[6], const CeedScalar *detC_m1,
                            CeedScalar *llnj, const CeedInt indj[6], const CeedInt indk[6], CeedScalar E2[][3]){
   // unpack ctx here
   const Physics context = (Physics)ctx;
@@ -233,7 +233,7 @@ CeedScalar NH_2nd_PK(void *ctx, CeedScalar Swork[6], CeedScalar Cinvwork[6], Cee
 }
 // -----------------------------------------------------------------------------
 // Mooney-Rivlin model
-CeedScalar MR_2nd_PK(void *ctx, CeedScalar Swork[6], CeedScalar Cinvwork[6], CeedScalar *detC_m1,
+CeedScalar MR_2nd_PK(void *ctx, CeedScalar Swork[6], CeedScalar Cinvwork[6], const CeedScalar *detC_m1,
                            CeedScalar *llnj, const CeedInt indj[6], const CeedInt indk[6], CeedScalar E2[][3]){
   // unoack context
   const Physics_MR context = (Physics_MR)ctx;
@@ -286,7 +286,7 @@ CeedScalar MR_2nd_PK(void *ctx, CeedScalar Swork[6], CeedScalar Cinvwork[6], Cee
 }
 // -----------------------------------------------------------------------------
 // Generalized Polynomial model
-CeedScalar GP_2nd_PK(void *ctx, CeedScalar Swork[6], CeedScalar Cinvwork[6], CeedScalar *detC_m1,
+CeedScalar GP_2nd_PK(void *ctx, CeedScalar Swork[6], CeedScalar Cinvwork[6], const CeedScalar *detC_m1,
                            CeedScalar *llnj, const CeedInt indj[6], const CeedInt indk[6], CeedScalar E2[][3]){
   // // C : right Cauchy-Green tensor
   // // C = I + 2E
@@ -394,8 +394,10 @@ static inline int commonFS_generic(void *ctx, const CeedScalar gradu[][3], CeedS
 // -----------------------------------------------------------------------------
 // Residual evaluation for hyperelasticity, finite strain
 // -----------------------------------------------------------------------------
-CEED_QFUNCTION(HyperFSF)(void *ctx, CeedInt Q, const CeedScalar *const *in,
-                         CeedScalar *const *out) { // add helper to call specific model
+static inline int HyperFSF_Generic(void *ctx, CeedInt Q, const CeedScalar *const *in,
+                         CeedScalar *const *out, 
+                         CeedScalar (*Swork_func)(void *ctx, CeedScalar Swork[6], CeedScalar Cinvwork[6], const CeedScalar *detC_m1,
+                         CeedScalar *llnj, const CeedInt indj[6], const CeedInt indk[6], CeedScalar E2[][3])) { // add helper to call specific model
   // *INDENT-OFF*
   // Inputs
[…]
