### 2021-04-08 d0c91ce
ceed - create CEED_MAX_BACKEND_PRIORITY, use to repair registry list reporting
