+++ examples/fluids/navierstokes.c
@@ -46,618 +46,7 @@
 
 const char help[] = "Solve Navier-Stokes using PETSc and libCEED\n";
 
-#include <ceed.h>
-#include <petscdmplex.h>
-#include <petscsys.h>
-#include <petscts.h>
-#include <stdbool.h>
-
 #include "navierstokes.h"
-
-#if PETSC_VERSION_LT(3,14,0)
-#  define DMPlexGetClosureIndices(a,b,c,d,e,f,g,h,i) DMPlexGetClosureIndices(a,b,c,d,f,g,i)
-#  define DMPlexRestoreClosureIndices(a,b,c,d,e,f,g,h,i) DMPlexRestoreClosureIndices(a,b,c,d,f,g,i)
-#endif
-
-#if PETSC_VERSION_LT(3,14,0)
-#  define DMAddBoundary(a,b,c,d,e,f,g,h,i,j,k,l,m,n) DMAddBoundary(a,b,c,e,h,i,j,k,f,g,m)
-#elif PETSC_VERSION_LT(3,16,0)
-#  define DMAddBoundary(a,b,c,d,e,f,g,h,i,j,k,l,m,n) DMAddBoundary(a,b,c,e,h,i,j,k,l,f,g,m)
-#else
-#  define DMAddBoundary(a,b,c,d,e,f,g,h,i,j,k,l,m,n) DMAddBoundary(a,b,c,d,f,g,h,i,j,k,l,m,n)
-#endif
-
-// MemType Options
-static const char *const memTypes[] = {
-  "host",
-  "device",
-  "memType", "CEED_MEM_", NULL
-};
-
-// Essential BC dofs are encoded in closure indices as -(i+1).
-static PetscInt Involute(PetscInt i) {
-  return i >= 0 ? i : -(i+1);
-}
-
-// Utility function to create local CEED restriction
-static PetscErrorCode CreateRestrictionFromPlex(Ceed ceed, DM dm, CeedInt P,
-    CeedInt height, DMLabel domainLabel, CeedInt value,
-    CeedElemRestriction *Erestrict) {
-
-  PetscSection section;
-  PetscInt p, Nelem, Ndof, *erestrict, eoffset, nfields, dim, depth;
-  DMLabel depthLabel;
-  IS depthIS, iterIS;
-  Vec Uloc;
-  const PetscInt *iterIndices;
-  PetscErrorCode ierr;
-
-  PetscFunctionBeginUser;
-  ierr = DMGetDimension(dm, &dim); CHKERRQ(ierr);
-  dim -= height;
-  ierr = DMGetLocalSection(dm, &section); CHKERRQ(ierr);
-  ierr = PetscSectionGetNumFields(section, &nfields); CHKERRQ(ierr);
-  PetscInt ncomp[nfields], fieldoff[nfields+1];
-  fieldoff[0] = 0;
-  for (PetscInt f=0; f<nfields; f++) {
-    ierr = PetscSectionGetFieldComponents(section, f, &ncomp[f]); CHKERRQ(ierr);
-    fieldoff[f+1] = fieldoff[f] + ncomp[f];
-  }
-
-  ierr = DMPlexGetDepth(dm, &depth); CHKERRQ(ierr);
-  ierr = DMPlexGetDepthLabel(dm, &depthLabel); CHKERRQ(ierr);
-  ierr = DMLabelGetStratumIS(depthLabel, depth - height, &depthIS); CHKERRQ(ierr);
-  if (domainLabel) {
-    IS domainIS;
-    ierr = DMLabelGetStratumIS(domainLabel, value, &domainIS); CHKERRQ(ierr);
-    if (domainIS) { // domainIS is non-empty
[…]
