+++ examples/fluids/src/setuplibceed.c
@@ -1,0 +1,252 @@
+
+#include "../navierstokes.h"
+
+// -----------------------------------------------------------------------------
+// libCEED Functions
+// -----------------------------------------------------------------------------
+
+// Utility function - essential BC dofs are encoded in closure indices as -(i+1).
+PetscInt Involute(PetscInt i) {
+  return i >= 0 ? i : -(i+1);
+}
+
+// Utility function to create local CEED restriction
+PetscErrorCode CreateRestrictionFromPlex(Ceed ceed, DM dm, CeedInt P,
+    CeedInt height, DMLabel domainLabel,
+    CeedInt value, CeedElemRestriction *Erestrict) {
+  PetscSection section;
+  PetscInt p, Nelem, Ndof, *erestrict, eoffset, nfields, dim, depth;
+  DMLabel depthLabel;
+  IS depthIS, iterIS;
+  Vec Uloc;
+  const PetscInt *iterIndices;
+  PetscErrorCode ierr;
+
+  PetscFunctionBeginUser;
+  ierr = DMGetDimension(dm, &dim); CHKERRQ(ierr);
+  dim -= height;
+  ierr = DMGetLocalSection(dm, &section); CHKERRQ(ierr);
+  ierr = PetscSectionGetNumFields(section, &nfields); CHKERRQ(ierr);
+  PetscInt ncomp[nfields], fieldoff[nfields+1];
+  fieldoff[0] = 0;
+  for (PetscInt f=0; f<nfields; f++) {
+    ierr = PetscSectionGetFieldComponents(section, f, &ncomp[f]); CHKERRQ(ierr);
+    fieldoff[f+1] = fieldoff[f] + ncomp[f];
+  }
+
+  ierr = DMPlexGetDepth(dm, &depth); CHKERRQ(ierr);
+  ierr = DMPlexGetDepthLabel(dm, &depthLabel); CHKERRQ(ierr);
+  ierr = DMLabelGetStratumIS(depthLabel, depth - height, &depthIS); CHKERRQ(ierr);
+  if (domainLabel) {
+    IS domainIS;
+    ierr = DMLabelGetStratumIS(domainLabel, value, &domainIS); CHKERRQ(ierr);
+    if (domainIS) { // domainIS is non-empty
+      ierr = ISIntersect(depthIS, domainIS, &iterIS); CHKERRQ(ierr);
+      ierr = ISDestroy(&domainIS); CHKERRQ(ierr);
+    } else { // domainIS is NULL (empty)
+      iterIS = NULL;
+    }
+    ierr = ISDestroy(&depthIS); CHKERRQ(ierr);
+  } else {
+    iterIS = depthIS;
+  }
+  if (iterIS) {
+    ierr = ISGetLocalSize(iterIS, &Nelem); CHKERRQ(ierr);
+    ierr = ISGetIndices(iterIS, &iterIndices); CHKERRQ(ierr);
+  } else {
+    Nelem = 0;
+    iterIndices = NULL;
+  }
+  ierr = PetscMalloc1(Nelem*PetscPowInt(P, dim), &erestrict); CHKERRQ(ierr);
+  for (p=0,eoffset=0; p<Nelem; p++) {
+    PetscInt c = iterIndices[p];
+    PetscInt numindices, *indices, nnodes;
+    ierr = DMPlexGetClosureIndices(dm, section, section, c, PETSC_TRUE,
+                                   &numindices, &indices, NULL, NULL);
+    CHKERRQ(ierr);
+    bool flip = false;
+    if (height > 0) {
+      PetscInt numCells, numFaces, start = -1;
+      const PetscInt *orients, *faces, *cells;
[…]

[…]
