### 2021-04-09 b7b4221
Fluids - removed CreateVectorFromPetscVec() since it was only used once
