+++ examples/fluids/src/misc.c
@@ -4,16 +4,6 @@
 // Miscellaneous utility functions
 // -----------------------------------------------------------------------------
 
-int CreateVectorFromPetscVec(Ceed ceed, Vec p, CeedVector *v) {
-  PetscErrorCode ierr;
-  PetscInt m;
-
-  PetscFunctionBeginUser;
-  ierr = VecGetLocalSize(p, &m); CHKERRQ(ierr);
-  ierr = CeedVectorCreate(ceed, m, v); CHKERRQ(ierr);
-  PetscFunctionReturn(0);
-}
-
 int VectorPlacePetscVec(CeedVector c, Vec p) {
   PetscErrorCode ierr;
   PetscInt mceed, mpetsc;
