### 2021-04-09 8725033
minor - improve error message for uncompiled backends
