### 2021-04-09 d1d35e2
style - updating C variable style
