+++ backends/opt/ceed-opt-serial.c
@@ -47,9 +47,9 @@ static int CeedInit_Opt_Serial(const char *resource, Ceed ceed) {
 
   // Create reference CEED that implementation will be dispatched
   //   through unless overridden
-  Ceed ceedref;
-  CeedInit("/cpu/self/ref/serial", &ceedref);
-  ierr = CeedSetDelegate(ceed, ceedref); CeedChkBackend(ierr);
+  Ceed ceed_ref;
+  CeedInit("/cpu/self/ref/serial", &ceed_ref);
+  ierr = CeedSetDelegate(ceed, ceed_ref); CeedChkBackend(ierr);
 
   // Set fallback CEED resource for advanced operator functionality
   const char fallbackresource[] = "/cpu/self/ref/serial";
@@ -64,7 +64,7 @@ static int CeedInit_Opt_Serial(const char *resource, Ceed ceed) {
   // Set blocksize
   Ceed_Opt *data;
   ierr = CeedCalloc(1, &data); CeedChkBackend(ierr);
-  data->blksize = 1;
+  data->blk_size = 1;
   ierr = CeedSetData(ceed, data); CeedChkBackend(ierr);
 
   return CEED_ERROR_SUCCESS;
