+++ tests/t402-qfunction.c
@@ -1,6 +1,6 @@
 /// @file
-/// Test viewing of qfunction
-/// \test Test viewing of qfunction
+/// Test viewing of QFunction
+/// \test Test viewing of QFunction
 #include <ceed.h>
 #include "t400-qfunction.h"
 

[…]
