### 2021-04-09 333d076
docs - update header file notes with example
