### 2021-04-10 f8da22b
Fluids - created AppCtx for cl option variables and moved cl options to a cloption.c
