### 2021-04-10 9b69176
Fluids: change style of AppCtx
