+++ examples/fluids/src/cloptions.c
@@ -1,9 +1,28 @@
+// Copyright (c) 2017, Lawrence Livermore National Security, LLC. Produced at
+// the Lawrence Livermore National Laboratory. LLNL-CODE-734707. All Rights
+// reserved. See files LICENSE and NOTICE for details.
+//
+// This file is part of CEED, a collection of benchmarks, miniapps, software
+// libraries and APIs for efficient high-order finite element and spectral
+// element discretizations for exascale applications. For more information and
+// source code availability see http://github.com/ceed.
+//
+// The CEED research is supported by the Exascale Computing Project 17-SC-20-SC,
+// a collaborative effort of two U.S. Department of Energy organizations (Office
+// of Science and the National Nuclear Security Administration) responsible for
+// the planning and preparation of a capable exascale ecosystem, including
+// software, applications, hardware, advanced system engineering and early
+// testbed platforms, in support of the nation's exascale computing imperative.
+
+/// @file
+/// Command line option processing for fluid mechanics example using PETSc
 
 #include "../navierstokes.h"
 
 // -----------------------------------------------------------------------------
 // Process command line options
 // -----------------------------------------------------------------------------
+
 // Process general command line options
 PetscErrorCode ProcessCommandLineOptions(MPI_Comm comm, AppCtx app_ctx) {
 
@@ -19,29 +38,30 @@ PetscErrorCode ProcessCommandLineOptions(MPI_Comm comm, AppCtx app_ctx) {
   ierr = PetscOptionsString("-ceed", "CEED resource specifier",
                             NULL, app_ctx->ceed_resource, app_ctx->ceed_resource,
                             sizeof(app_ctx->ceed_resource), &ceed_flag); CHKERRQ(ierr);
-  
+
   app_ctx->test_mode = PETSC_FALSE;
   ierr = PetscOptionsBool("-test", "Run in test mode",
                           NULL, app_ctx->test_mode, &app_ctx->test_mode, NULL); CHKERRQ(ierr);
-  
+
   app_ctx->test_tol = 1E-11;
   ierr = PetscOptionsScalar("-compare_final_state_atol",
                             "Test absolute tolerance",
                             NULL, app_ctx->test_tol, &app_ctx->test_tol, NULL); CHKERRQ(ierr);
-  
+
   ierr = PetscOptionsString("-compare_final_state_filename", "Test filename",
                             NULL, app_ctx->file_path, app_ctx->file_path,
                             sizeof(app_ctx->file_path), NULL); CHKERRQ(ierr);
-  
-  ierr = PetscOptionsFList("-problem", "Problem to solve", NULL, app_ctx->problems,
+
+  ierr = PetscOptionsFList("-problem", "Problem to solve", NULL,
+                           app_ctx->problems,
                            app_ctx->problem_name, app_ctx->problem_name, sizeof(app_ctx->problem_name),
                            &problem_flag); CHKERRQ(ierr);
-  
+
   app_ctx->viz_refine = 0;
   ierr = PetscOptionsInt("-viz_refine",
                          "Regular refinement levels for visualization",
                          NULL, app_ctx->viz_refine, &app_ctx->viz_refine, NULL); CHKERRQ(ierr);
-  
+
   app_ctx->output_freq = 10;
   ierr = PetscOptionsInt("-output_freq",
                          "Frequency of output, in number of steps",
[…]
