+++ examples/fluids/Makefile
@@ -33,10 +33,10 @@ LDFLAGS += $(patsubst -L%, $(call pkgconf, --variable=ldflag_rpath $(PETSc.pc))%
 LDLIBS = $(call pkgconf, --libs-only-l $(PETSc.pc) $(ceed.pc)) -lm
 
 OBJDIR := build
-SRCDIR := src
+UTILSDIR := utils
 PROBLEMDIR := problems
 
-src.c := navierstokes.c $(sort $(wildcard $(PROBLEMDIR)/*.c)) $(sort $(wildcard $(SRCDIR)/*.c))
+src.c := navierstokes.c $(sort $(wildcard $(PROBLEMDIR)/*.c)) $(sort $(wildcard $(UTILSDIR)/*.c))
 src.o = $(src.c:%.c=$(OBJDIR)/%.o)
 
 all: navierstokes
