### 2021-04-10 46ee82d
Created CreateDistributedDM() and some style
