+++ examples/fluids/utils/cloptions.c
@@ -25,24 +25,24 @@
 // Register problems to be available on the command line
 PetscErrorCode RegisterProblem(AppCtx app_ctx) {
 
-    PetscErrorCode   ierr;
-    app_ctx->problems = NULL;
+  PetscErrorCode   ierr;
+  app_ctx->problems = NULL;
 
-    PetscFunctionBeginUser;
+  PetscFunctionBeginUser;
 
-    ierr = PetscFunctionListAdd(&app_ctx->problems, "density_current",
-                                NS_DENSITY_CURRENT); CHKERRQ(ierr);
+  ierr = PetscFunctionListAdd(&app_ctx->problems, "density_current",
+                              NS_DENSITY_CURRENT); CHKERRQ(ierr);
 
-    ierr = PetscFunctionListAdd(&app_ctx->problems, "euler_vortex",
-                                NS_EULER_VORTEX); CHKERRQ(ierr);
+  ierr = PetscFunctionListAdd(&app_ctx->problems, "euler_vortex",
+                              NS_EULER_VORTEX); CHKERRQ(ierr);
 
-    ierr = PetscFunctionListAdd(&app_ctx->problems, "advection", 
-                                NS_ADVECTION); CHKERRQ(ierr);
+  ierr = PetscFunctionListAdd(&app_ctx->problems, "advection",
+                              NS_ADVECTION); CHKERRQ(ierr);
 
-    ierr = PetscFunctionListAdd(&app_ctx->problems, "advection2d", 
-                                NS_ADVECTION2D); CHKERRQ(ierr);
+  ierr = PetscFunctionListAdd(&app_ctx->problems, "advection2d",
+                              NS_ADVECTION2D); CHKERRQ(ierr);
 
-    PetscFunctionReturn(0);
+  PetscFunctionReturn(0);
 }
 
 // Process general command line options
