### 2021-04-10 aed13ae
solid/MR: fixed the residual evaluation section
