### 2021-04-10 b6c62e0
added computeJM1 function instead of computeDetCM1 function
