+++ examples/solids/qfunctions/FSInitial-NH1.h
@@ -70,18 +70,18 @@ static inline CeedScalar log1p_series_shifted(CeedScalar x) {
 #endif
 
 // -----------------------------------------------------------------------------
-// Compute det C - 1
+// Compute det F - 1
 // -----------------------------------------------------------------------------
-#ifndef DETCM1
-#define DETCM1
-static inline CeedScalar computeDetCM1(CeedScalar E2work[6]) {
-  return E2work[0]*(E2work[1]*E2work[2]-E2work[3]*E2work[3]) +
-         E2work[5]*(E2work[4]*E2work[3]-E2work[5]*E2work[2]) +
-         E2work[4]*(E2work[5]*E2work[3]-E2work[4]*E2work[1]) +
-         E2work[0] + E2work[1] + E2work[2] +
-         E2work[0]*E2work[1] + E2work[0]*E2work[2] +
-         E2work[1]*E2work[2] - E2work[5]*E2work[5] -
-         E2work[4]*E2work[4] - E2work[3]*E2work[3];
+#ifndef DETJM1
+#define DETJM1
+static inline CeedScalar computeJM1(const CeedScalar grad_u[3][3]) {
+  return grad_u[0][0]*(grad_u[1][1]*grad_u[2][2]-grad_u[1][2]*grad_u[2][1]) +
+         grad_u[0][1]*(grad_u[1][2]*grad_u[2][0]-grad_u[1][0]*grad_u[2][2]) +
+         grad_u[0][2]*(grad_u[1][0]*grad_u[2][1]-grad_u[2][0]*grad_u[1][1]) +
+         grad_u[0][0] + grad_u[1][1] + grad_u[2][2] +
+         grad_u[0][0]*grad_u[1][1] + grad_u[0][0]*grad_u[2][2] +
+         grad_u[1][1]*grad_u[2][2] - grad_u[0][1]*grad_u[1][0] -
+         grad_u[0][2]*grad_u[2][0] - grad_u[1][2]*grad_u[2][1];
 };
 #endif
 
@@ -90,7 +90,7 @@ static inline CeedScalar computeDetCM1(CeedScalar E2work[6]) {
 // -----------------------------------------------------------------------------
 static inline int commonFS(const CeedScalar lambda, const CeedScalar mu,
                            const CeedScalar grad_u[3][3], CeedScalar Swork[6],
-                           CeedScalar Cinvwork[6], CeedScalar *detC_m1,
+                           CeedScalar Cinvwork[6], CeedScalar *Jm1,
                            CeedScalar *llnj) {
   // E - Green-Lagrange strain tensor
   //     E = 1/2 (grad_u + grad_u^T + grad_u^T*grad_u)
@@ -107,7 +107,8 @@ static inline int commonFS(const CeedScalar lambda, const CeedScalar mu,
                          {E2work[4], E2work[3], E2work[2]}
                         };
   // *INDENT-ON*
-  (*detC_m1) = computeDetCM1(E2work);
+  // J-1
+  (*Jm1) = computeJM1(grad_u);
 
   // C : right Cauchy-Green tensor
   // C = I + 2E
@@ -127,7 +128,7 @@ static inline int commonFS(const CeedScalar lambda, const CeedScalar mu,
                      C[0][2]*C[2][1] - C[0][1]*C[2][2] /* *NOPAD* */
                     };
   for (CeedInt m = 0; m < 6; m++)
-    Cinvwork[m] = A[m] / (*detC_m1 + 1.);
+    Cinvwork[m] = A[m] / ((*Jm1 + 1.)*(*Jm1 + 1.));
 
[…]
