### 2021-04-10 cea8f1d
delete computeDetF in FSCurrent-NH1and2
