### 2021-04-11 b3fe387
Update ceed.c
Fix function name in docs
