+++ examples/solids/index.rst
@@ -338,14 +338,14 @@ Carrying through the differentiation :math:numref:`strain-energy-grad` for the m
    Note that the product :math:`\bm C^{-1} \bm E` is also symmetric, and that :math:`\bm E` should be computed using :math:numref:`eq-green-lagrange-strain`.
 
    Similarly, it is preferable to compute :math:`\log J` using ``log1p``, especially in case of nearly incompressible materials.
-   To sketch this idea, suppose we have the :math:`2\times 2` symmetric matrix :math:`C = \left( \begin{smallmatrix} 1 + e_{00} & e_{01} \\ e_{01} & 1 + e_{11} \end{smallmatrix} \right)`.
+   To sketch this idea, suppose we have the :math:`2\times 2` non-symmetric matrix :math:`\bm{F} = \left( \begin{smallmatrix} 1 + u_{0,0} & u_{0,1} \\ u_{1,0} & 1 + u_{1,1} \end{smallmatrix} \right)`.
    Then we compute
 
    .. math::
-      \log \sqrt{\lvert C \rvert} = \frac 1 2 \mathtt{log1p}(e_{00} + e_{11} + e_{00} e_{11} - e_{01}^2).
+      \log J = \mathtt{log1p}(u_{0,0} + u_{1,1} + u_{0,0} u_{1,1} - u_{0,1} - u_{1,0}).
 
-   which gives accurate results even in the limit when the entries :math:`e_{ij}` are very small.
-   For example, if :math:`e_{ij} \sim 10^{-8}`, then naive computation of :math:`\bm I_3 - \bm C^{-1}` and :math:`\log J` will have a relative accuracy of order :math:`10^{-8}` in double precision and no correct digits in single precision.
+   which gives accurate results even in the limit when the entries :math:`u_{i,j}` are very small.
+   For example, if :math:`u_{i,j} \sim 10^{-8}`, then naive computation of :math:`\bm I_3 - \bm C^{-1}` and :math:`\log J` will have a relative accuracy of order :math:`10^{-8}` in double precision and no correct digits in single precision.
    When using the stable choices above, these quantities retain full :math:`\varepsilon_{\text{machine}}` relative accuracy.
 
 .. note::
