### 2021-04-12 34359f1
interface - add Ceed*Reference functions
