### 2021-04-12 9560d06
interface - add reference copying functions Co-authored-by: Jed Brown <jed@jedbrown.org>
