+++ interface/ceed-vector.c
@@ -130,7 +130,7 @@ int CeedVectorSetData(CeedVector vec, void *data) {
 
   @ref Backend
 **/
-int CeedVectorIncrementRefCounter(CeedVector vec) {
+int CeedVectorReference(CeedVector vec) {
   vec->ref_count++;
   return CEED_ERROR_SUCCESS;
 }
@@ -174,11 +174,35 @@ int CeedVectorCreate(Ceed ceed, CeedInt length, CeedVector *vec) {
 
   ierr = CeedCalloc(1, vec); CeedChk(ierr);
   (*vec)->ceed = ceed;
-  ierr = CeedIncrementRefCounter(ceed); CeedChk(ierr);
+  ierr = CeedReference(ceed); CeedChk(ierr);
   (*vec)->ref_count = 1;
   (*vec)->length = length;
   (*vec)->state = 0;
   ierr = ceed->VectorCreate(length, *vec); CeedChk(ierr);
+  return CEED_ERROR_SUCCESS;
+}
+
+/**
+  @brief Copy the pointer to a CeedVector. Both pointers should
+           be destroyed with `CeedVectorDestroy()`;
+           Note: If `*vec_copy` is non-NULL, then it is assumed that
+           `*vec_copy` is a pointer to a CeedVector. This
+           CeedVector will be destroyed if `*vec_copy` is the only
+           reference to this CeedVector.
+
+  @param vec            CeedVector to copy reference to
+  @param[out] vec_copy  Variable to store copied reference
+
+  @return An error code: 0 - success, otherwise - failure
+
+  @ref User
+**/
+int CeedVectorReferenceCopy(CeedVector vec, CeedVector *vec_copy) {
+  int ierr;
+
+  ierr = CeedVectorReference(vec); CeedChk(ierr);
+  ierr = CeedVectorDestroy(vec_copy); CeedChk(ierr);
+  *vec_copy = vec;
   return CEED_ERROR_SUCCESS;
 }
 

[…]
