### 2021-04-12 735987e
Update examples/solids/index.rst
Co-authored-by: Jed Brown <jed@jedbrown.org>
