### 2021-04-12 f1a52cc
refactored MR to work with current version of main. Updated doc with dS for MR
