+++ examples/solids/src/setuplibceed.c
@@ -23,12 +23,20 @@
 #include "../qfunctions/Linear.h"            // Linear elasticity
 #include "../qfunctions/SS-NH.h"             // Hyperelasticity small strain
 // Hyperelasticity finite strain
+<<<<<<< HEAD
 #include "../qfunctions/FSInitial-NH1.h"     // -- Initial config 1 w/ dXref_dxinit, Grad(u) storage  / Neo-Hookean 
 #include "../qfunctions/FSInitial-NH2.h"     // -- Initial config 2 w/ dXref_dxinit, Grad(u), Cinv, constant storage  / Neo-Hookean 
 #include "../qfunctions/FSCurrent-NH1.h"     // -- Current config 1 w/ dXref_dxinit, Grad(u) storage  / Neo-Hookean 
 #include "../qfunctions/FSCurrent-NH2.h"     // -- Current config 2 w/ dXref_dxcurr, tau, constant storage / Neo-Hookean 
 #include "../qfunctions/FSInitial-MR1.h"     // -- Initial config 1 w/ dXref_dxinit, Grad(u) storage / Mooney-Rivlin
 #include "../qfunctions/FSInitial-MRc.h"     // -- Initial config (old version) for Neo-Hookean and Mooney-Rivlin (and start of GP)
+=======
+#include "../qfunctions/FSInitial-NH1.h"     // -- Initial config 1 w/ dXref_dxinit, Grad(u) storage
+#include "../qfunctions/FSInitial-NH2.h"     // -- Initial config 2 w/ dXref_dxinit, Grad(u), Cinv, constant storage
+#include "../qfunctions/FSCurrent-NH1.h"     // -- Current config 1 w/ dXref_dxinit, Grad(u) storage
+#include "../qfunctions/FSCurrent-NH2.h"     // -- Current config 2 w/ dXref_dxcurr, tau, constant storage
+#include "../qfunctions/FSInitial-MR.h"     // -- Initial config (old version) for Neo-Hookean and Mooney-Rivlin (and start of GP)
+>>>>>>> refactored MR to work with current version of main. Updated doc with dS for MR
 #include "../qfunctions/constantForce.h"     // Constant forcing function
 #include "../qfunctions/manufacturedForce.h" // Manufactured solution forcing
 #include "../qfunctions/manufacturedTrue.h"  // Manufactured true solution
@@ -42,7 +50,11 @@
 // Problem options
 // -----------------------------------------------------------------------------
 // Data specific to each problem option
+<<<<<<< HEAD
 problemData problem_options[8] = {
+=======
+problemData problem_options[9] = {
+>>>>>>> refactored MR to work with current version of main. Updated doc with dS for MR
   [ELAS_LINEAR] = {
     .q_data_size = 10, // For linear elasticity, 6 would be sufficient
     .setup_geo = SetupGeo,
@@ -127,6 +139,7 @@ problemData problem_options[8] = {
     .diagnostic_loc = ElasFSCurrentNH2Diagnostic_loc,
     .quad_mode = CEED_GAUSS
   },
+<<<<<<< HEAD
   [ELAS_FSInitial_MR1] = {
     .q_data_size = 10,
     .setup_geo = SetupGeo,
@@ -153,6 +166,51 @@ problemData problem_options[8] = {
     .jacob_loc = ElasFSInitialMRcdF_loc,
     .energy_loc = ElasFSInitialMRcEnergy_loc,
[…]
