### 2021-04-12 378d457
refactored MR to work with current version of main. Updated doc with dS for MR
