### 2021-04-12 593c7ab
Fluids - create VizRefineDM() for viz_refine
