+++ doc/sphinx/source/references.bib
@@ -134,3 +134,11 @@
   year={2011},
   pages={807--827}
 }
+
+@book{bower2010applied,
+  title={Applied mechanics of solids},
+  author={Bower, Allan F},
+  year={2010},
+  publisher={CRC press},
+  isbn={978-1-4398-0247-2}
+}
