+++ doc/sphinx/source/references.bib
@@ -141,6 +141,7 @@
   year={2010},
   publisher={CRC press},
   isbn={978-1-4398-0247-2}
+<<<<<<< HEAD
 }
 
 
@@ -158,4 +159,6 @@
   number    = "4/5",
   pages     = "489--490",
   year      =  2002
+=======
+>>>>>>> fixed documentation formatting
 }
