### 2021-04-12 1aa3f08
Fluids - remove the second ceed initialization
