+++ examples/fluids/navierstokes.c
@@ -56,6 +56,7 @@ int main(int argc, char **argv) {
   TSAdapt adapt;
   User user;
   Units units;
+  CeedData ceed_data;
   SetupContext ctxSetupData;
   Physics ctxPhysData;
   AppCtx app_ctx;
@@ -66,12 +67,7 @@ int main(int argc, char **argv) {
   PetscScalar ftime;
   Vec Q, Qloc, Xloc;
   Ceed ceed;
-  CeedVector xcorners, qdata, q0ceed;
-  CeedBasis basisx, basisxc, basisq;
-  CeedElemRestriction restrictx, restrictq, restrictqdi;
-  CeedQFunction qf_setupVol, qf_ics, qf_rhsVol, qf_ifunctionVol;
   CeedQFunctionContext ctxSetup, ctxNS, ctxAdvection, ctxEuler;
-  CeedOperator op_setupVol, op_ics;
   PetscInt Xloc_size;
   SimpleBC bc;
   double start, cpu_time_used;
@@ -96,7 +92,7 @@ int main(int argc, char **argv) {
   if (ierr) return ierr;
 
   // ---------------------------------------------------------------------------
-  // Allocate PETSc context
+  // Allocate contexts
   // ---------------------------------------------------------------------------
   ierr = PetscCalloc1(1, &user); CHKERRQ(ierr);
   ierr = PetscMalloc1(1, &units); CHKERRQ(ierr);
@@ -105,6 +101,7 @@ int main(int argc, char **argv) {
   ierr = PetscMalloc1(1, &ctxSetupData); CHKERRQ(ierr);
   ierr = PetscMalloc1(1, &ctxPhysData); CHKERRQ(ierr);
   ierr = PetscCalloc1(1, &app_ctx); CHKERRQ(ierr);
+  ierr = PetscCalloc1(1, &ceed_data); CHKERRQ(ierr);
 
   // ---------------------------------------------------------------------------
   // Process command line options
@@ -134,6 +131,7 @@ int main(int argc, char **argv) {
                 qdatasizeVol = problem->qdatasizeVol,
                 numP = app_ctx->degree + 1, // Set number of 1D nodes and quadrature points
                 numQ = numP + app_ctx->q_extra;
+  user->phys = ctxPhysData;
 
   // ---------------------------------------------------------------------------
   // Setup DM
@@ -229,174 +227,17 @@ int main(int argc, char **argv) {
   // ---------------------------------------------------------------------------
   // Set up libCEED
   // ---------------------------------------------------------------------------
-  // CEED Bases
-  CeedBasisCreateTensorH1Lagrange(ceed, dim, ncompq, numP, numQ, CEED_GAUSS,
-                                  &basisq);
-  CeedBasisCreateTensorH1Lagrange(ceed, dim, ncompx, 2, numQ, CEED_GAUSS,
-                                  &basisx);
-  CeedBasisCreateTensorH1Lagrange(ceed, dim, ncompx, 2, numP,
-                                  CEED_GAUSS_LOBATTO, &basisxc);
-  ierr = DMGetCoordinateDM(dm, &dmcoord); CHKERRQ(ierr);
-  ierr = DMPlexSetClosurePermutationTensor(dmcoord, PETSC_DETERMINE, NULL);
-  CHKERRQ(ierr);
-
-  // CEED Restrictions
[…]
