+++ examples/fluids/src/setuplibceed.c
@@ -248,3 +248,205 @@ PetscErrorCode CreateOperatorForDomain(Ceed ceed, DM dm, SimpleBC bc,
   }
   PetscFunctionReturn(0);
 }
+
+PetscErrorCode SetupLibceed(Ceed ceed, CeedData ceed_data, DM dm, User user,
+                            AppCtx app_ctx, problemData *problem, SimpleBC bc) {
+
+  PetscErrorCode ierr;
+
+  DM dmcoord;
+  const PetscInt ncompq = 5;
+  PetscInt localNelemVol;
+
+  const CeedInt dim = problem->dim,
+                ncompx = problem->dim,
+                qdatasizeVol = problem->qdatasizeVol,
+                numP = app_ctx->degree + 1,
+                numQ = numP + app_ctx->q_extra;
+
+  // CEED Bases
+  CeedBasisCreateTensorH1Lagrange(ceed, dim, ncompq, numP, numQ, CEED_GAUSS,
+                                  &ceed_data->basisq);
+
+  CeedBasisCreateTensorH1Lagrange(ceed, dim, ncompx, 2, numQ, CEED_GAUSS,
+                                  &ceed_data->basisx);
+
+  CeedBasisCreateTensorH1Lagrange(ceed, dim, ncompx, 2, numP,
+                                  CEED_GAUSS_LOBATTO, &ceed_data->basisxc);
+
+  ierr = DMGetCoordinateDM(dm, &dmcoord); CHKERRQ(ierr);
+  ierr = DMPlexSetClosurePermutationTensor(dmcoord, PETSC_DETERMINE, NULL);
+  CHKERRQ(ierr);
+
+  // CEED Restrictions
+  ierr = GetRestrictionForDomain(ceed, dm, 0, 0, 0, numP, numQ,
+                                 qdatasizeVol, &ceed_data->restrictq, &ceed_data->restrictx,
+                                 &ceed_data->restrictqdi); CHKERRQ(ierr);
+
+  // Create the CEED vectors that will be needed in setup
+  CeedInt NqptsVol;
+  CeedBasisGetNumQuadraturePoints(ceed_data->basisq, &NqptsVol);
+  CeedElemRestrictionGetNumElements(ceed_data->restrictq, &localNelemVol);
+  CeedVectorCreate(ceed, qdatasizeVol*localNelemVol*NqptsVol, &ceed_data->qdata);
+  CeedElemRestrictionCreateVector(ceed_data->restrictq, &ceed_data->q0ceed, NULL);
+
+  // Create the Q-function that builds the quadrature data for the NS operator
+  CeedQFunctionCreateInterior(ceed, 1, problem->setupVol, problem->setupVol_loc,
+                              &ceed_data->qf_setupVol);
+  CeedQFunctionAddInput(ceed_data->qf_setupVol, "dx", ncompx*dim, CEED_EVAL_GRAD);
+  CeedQFunctionAddInput(ceed_data->qf_setupVol, "weight", 1, CEED_EVAL_WEIGHT);
+  CeedQFunctionAddOutput(ceed_data->qf_setupVol, "qdata", qdatasizeVol,
+                         CEED_EVAL_NONE);
+
+  // Create the Q-function that sets the ICs of the operator
+  CeedQFunctionCreateInterior(ceed, 1, problem->ics, problem->ics_loc,
+                              &ceed_data->qf_ics);
+  CeedQFunctionAddInput(ceed_data->qf_ics, "x", ncompx, CEED_EVAL_INTERP);
+  CeedQFunctionAddOutput(ceed_data->qf_ics, "q0", ncompq, CEED_EVAL_NONE);
+
+  if (problem->applyVol_rhs) { // Create the Q-function that defines the action of the RHS operator
[…]
