+++ examples/fluids/navierstokes.c
@@ -67,7 +67,6 @@ int main(int argc, char **argv) {
   PetscScalar ftime;
   Vec Q, Qloc, Xloc;
   Ceed ceed;
-  CeedQFunctionContext ctxSetup, ctxNS, ctxAdvection, ctxEuler;
   PetscInt Xloc_size;
   SimpleBC bc;
   double start, cpu_time_used;
@@ -235,54 +234,8 @@ int main(int argc, char **argv) {
   CHKERRQ(ierr);
 
   // Set up contex for QFunctions
-  CeedQFunctionContextCreate(ceed, &ctxSetup);
-  CeedQFunctionContextSetData(ctxSetup, CEED_MEM_HOST, CEED_USE_POINTER,
-                              sizeof *ctxSetupData, ctxSetupData);
-
-  CeedQFunctionContextCreate(ceed, &ctxNS);
-  CeedQFunctionContextSetData(ctxNS, CEED_MEM_HOST, CEED_USE_POINTER,
-                              sizeof ctxPhysData->dc_ctx_data, ctxPhysData->dc_ctx_data);
-
-  CeedQFunctionContextCreate(ceed, &ctxEuler);
-  CeedQFunctionContextSetData(ctxEuler, CEED_MEM_HOST, CEED_USE_POINTER,
-                              sizeof ctxPhysData->euler_ctx_data, ctxPhysData->euler_ctx_data);
-
-  CeedQFunctionContextCreate(ceed, &ctxAdvection);
-  CeedQFunctionContextSetData(ctxAdvection, CEED_MEM_HOST, CEED_USE_POINTER,
-                              sizeof ctxPhysData->advection_ctx_data, ctxPhysData->advection_ctx_data);
-
-  if (ceed_data->qf_ics && strcmp(app_ctx->problem_name, "euler_vortex") != 0)
-    CeedQFunctionSetContext(ceed_data->qf_ics, ctxSetup);
-
-  if (strcmp(app_ctx->problem_name, "density_current") == 0) {
-    if (ceed_data->qf_rhsVol)
-      CeedQFunctionSetContext(ceed_data->qf_rhsVol, ctxNS);
-
-    if (ceed_data->qf_ifunctionVol)
-      CeedQFunctionSetContext(ceed_data->qf_ifunctionVol, ctxNS);
-
-  } else if (strcmp(app_ctx->problem_name, "euler_vortex") == 0) {
-    if (ceed_data->qf_ics)
-      CeedQFunctionSetContext(ceed_data->qf_ics, ctxEuler);
-
-    if (ceed_data->qf_rhsVol)
-      CeedQFunctionSetContext(ceed_data->qf_rhsVol, ctxEuler);
-
-    if (ceed_data->qf_ifunctionVol)
-      CeedQFunctionSetContext(ceed_data->qf_ifunctionVol, ctxEuler);
-
-    if (ceed_data->qf_applySur)
-      CeedQFunctionSetContext(ceed_data->qf_applySur, ctxEuler);
-  } else {
-    if (ceed_data->qf_rhsVol)
-      CeedQFunctionSetContext(ceed_data->qf_rhsVol, ctxAdvection);
-
-    if (ceed_data->qf_ifunctionVol)
-      CeedQFunctionSetContext(ceed_data->qf_ifunctionVol, ctxAdvection);
-
-    if (ceed_data->qf_applySur)
-      CeedQFunctionSetContext(ceed_data->qf_applySur, ctxAdvection);
-  }
+  ierr = SetupContextForProblems(ceed, ceed_data, app_ctx, ctxSetupData,
+                                 ctxPhysData); CHKERRQ(ierr);
 
   // Set up user structure
   user->comm = comm;
@@ -307,7 +260,7 @@ int main(int argc, char **argv) {
 
[…]
