### 2021-04-13 9141be3
Fluids - Assign structs after they are created
