### 2021-04-13 123fa6e
Fluids - created GetError_NS()
