+++ examples/fluids/src/misc.c
@@ -5,11 +5,13 @@
 // -----------------------------------------------------------------------------
 
 int VectorPlacePetscVec(CeedVector c, Vec p) {
-  PetscErrorCode ierr;
+  
   PetscInt mceed, mpetsc;
   PetscScalar *a;
+  PetscErrorCode ierr;
 
   PetscFunctionBeginUser;
+
   ierr = CeedVectorGetLength(c, &mceed); CHKERRQ(ierr);
   ierr = VecGetLocalSize(p, &mpetsc); CHKERRQ(ierr);
   if (mceed != mpetsc) SETERRQ2(PETSC_COMM_SELF, PETSC_ERR_ARG_INCOMP,
@@ -23,24 +25,28 @@ int VectorPlacePetscVec(CeedVector c, Vec p) {
 PetscErrorCode DMPlexInsertBoundaryValues_NS(DM dm,
     PetscBool insertEssential, Vec Qloc, PetscReal time, Vec faceGeomFVM,
     Vec cellGeomFVM, Vec gradFVM) {
-  PetscErrorCode ierr;
+  
   Vec Qbc;
-
+  PetscErrorCode ierr;
+  
   PetscFunctionBegin;
+
   ierr = DMGetNamedLocalVector(dm, "Qbc", &Qbc); CHKERRQ(ierr);
   ierr = VecAXPY(Qloc, 1., Qbc); CHKERRQ(ierr);
   ierr = DMRestoreNamedLocalVector(dm, "Qbc", &Qbc); CHKERRQ(ierr);
+
   PetscFunctionReturn(0);
 }
 
 // Compare reference solution values with current test run for CI
 PetscErrorCode RegressionTests_NS(AppCtx app_ctx, Vec Q) {
 
-  PetscErrorCode ierr;
-
   Vec Qref;
   PetscViewer viewer;
   PetscReal error, Qrefnorm;
+  PetscErrorCode ierr;
+
+  PetscFunctionBegin;
 
   // Read reference file
   ierr = VecDuplicate(Q, &Qref); CHKERRQ(ierr);
@@ -62,9 +68,47 @@ PetscErrorCode RegressionTests_NS(AppCtx app_ctx, Vec Q) {
                        (double)error); CHKERRQ(ierr);
   }
 
-  // Destroy PETSc objects
+  // Clean up objects
   ierr = PetscViewerDestroy(&viewer); CHKERRQ(ierr);
   ierr = VecDestroy(&Qref); CHKERRQ(ierr);
+
+  PetscFunctionReturn(0);
+}
+
+  // Get error for problems with exact solutions
+PetscErrorCode GetError_NS(CeedData ceed_data, DM dm, AppCtx app_ctx, Vec Q, PetscScalar ftime) {
+
+  PetscInt lnodes;
+  Vec Qexact, Qexactloc;
+  PetscReal rel_error, norm_error, norm_exact;
+  PetscErrorCode ierr;
+
+  PetscFunctionBegin;
+
+  // Get exact solution at final time
+  ierr = DMCreateGlobalVector(dm, &Qexact); CHKERRQ(ierr);
+  ierr = DMGetLocalVector(dm, &Qexactloc); CHKERRQ(ierr);
+  ierr = VecGetSize(Qexactloc, &lnodes); CHKERRQ(ierr);
+  ierr = ICs_FixMultiplicity(ceed_data->op_ics, ceed_data->xcorners,
+                             ceed_data->q0ceed, dm, Qexactloc, Qexact,
+                             ceed_data->restrictq, ceed_data->ctxSetup, ftime); CHKERRQ(ierr);
+  
+  // Get |exact solution - obtained solution|
[…]
