+++ examples/fluids/navierstokes.h
@@ -311,5 +311,8 @@ PetscErrorCode DMPlexInsertBoundaryValues_NS(DM dm,
 // Compare reference solution values with current test run for CI
 PetscErrorCode RegressionTests_NS(AppCtx app_ctx, Vec Q);
 
+// Get error for problems with exact solutions
+PetscErrorCode GetError_NS(CeedData ceed_data, DM dm, AppCtx app_ctx, Vec Q, PetscScalar ftime);
+
 // -----------------------------------------------------------------------------
 #endif
