### 2021-04-13 af19d7c
Fluids - created PrintOutput_NS()
This processes bash output except for the problem summary which will be
included after mem_type is managed properly.
