### 2021-04-13 1ad0cb7
Fluids - remove unused variables
