### 2021-04-14 c6a82cf
tests/python: fix file names
