### 2021-04-14 e6e763e
tests/python: add convenience targets to makefile
