### 2021-04-14 90c958f
Fluids - created TSSolve_NS() which creates, sets up, and solves ts
