+++ examples/fluids/navierstokes.c
@@ -52,8 +52,6 @@ int main(int argc, char **argv) {
   PetscInt ierr;
   MPI_Comm comm;
   DM dm;
-  TS ts;
-  TSAdapt adapt;
   User user;
   Units units;
   CeedData ceed_data;
@@ -64,12 +62,10 @@ int main(int argc, char **argv) {
   PetscInt lnodes, gnodes;
   const PetscInt ncompq = 5;
   PetscMPIInt rank;
-  PetscScalar ftime;
   Vec Q, Qloc, Xloc;
   Ceed ceed;
   PetscInt Xloc_size;
   SimpleBC bc;
-  double start, cpu_time_used;
 
   // todo: define a function
   // Check PETSc CUDA support
@@ -290,67 +286,12 @@ int main(int argc, char **argv) {
   }
   ierr = DMRestoreLocalVector(dm, &Qloc); CHKERRQ(ierr);
 
-  // Create and setup TS
-  ierr = TSCreate(comm, &ts); CHKERRQ(ierr);
-  ierr = TSSetDM(ts, dm); CHKERRQ(ierr);
-  if (ctxPhysData->implicit) {
-    ierr = TSSetType(ts, TSBDF); CHKERRQ(ierr);
-    if (user->op_ifunction) {
-      ierr = TSSetIFunction(ts, NULL, IFunction_NS, &user); CHKERRQ(ierr);
-    } else {                    // Implicit integrators can fall back to using an RHSFunction
-      ierr = TSSetRHSFunction(ts, NULL, RHS_NS, &user); CHKERRQ(ierr);
-    }
-  } else {
-    if (!user->op_rhs) SETERRQ(comm, PETSC_ERR_ARG_NULL,
-                                 "Problem does not provide RHSFunction");
-    ierr = TSSetType(ts, TSRK); CHKERRQ(ierr);
-    ierr = TSRKSetType(ts, TSRK5F); CHKERRQ(ierr);
-    ierr = TSSetRHSFunction(ts, NULL, RHS_NS, &user); CHKERRQ(ierr);
-  }
-  ierr = TSSetMaxTime(ts, 500. * units->second); CHKERRQ(ierr);
-  ierr = TSSetExactFinalTime(ts, TS_EXACTFINALTIME_STEPOVER); CHKERRQ(ierr);
-  ierr = TSSetTimeStep(ts, 1.e-2 * units->second); CHKERRQ(ierr);
-  if (app_ctx->test_mode) {ierr = TSSetMaxSteps(ts, 10); CHKERRQ(ierr);}
-  ierr = TSGetAdapt(ts, &adapt); CHKERRQ(ierr);
-  ierr = TSAdaptSetStepLimits(adapt,
-                              1.e-12 * units->second,
-                              1.e2 * units->second); CHKERRQ(ierr);
-  ierr = TSSetFromOptions(ts); CHKERRQ(ierr);
-  if (!app_ctx->cont_steps) { // print initial condition
-    if (!app_ctx->test_mode) {
-      ierr = TSMonitor_NS(ts, 0, 0., Q, user); CHKERRQ(ierr);
-    }
-  } else { // continue from time of last output
-    PetscReal time;
-    PetscInt count;
-    PetscViewer viewer;
-    char file_path[PETSC_MAX_PATH_LEN];
-    ierr = PetscSNPrintf(file_path, sizeof file_path, "%s/ns-time.bin",
-                         app_ctx->test_mode); CHKERRQ(ierr);
-    ierr = PetscViewerBinaryOpen(comm, file_path, FILE_MODE_READ, &viewer);
-    CHKERRQ(ierr);
[…]
