+++ examples/fluids/src/setupts.c
@@ -236,10 +236,11 @@ PetscErrorCode ICs_FixMultiplicity(CeedOperator op_ics, CeedVector xcorners,
                                    CeedVector q0ceed, DM dm, Vec Qloc, Vec Q,
                                    CeedElemRestriction restrictq,
                                    CeedQFunctionContext ctxSetup, CeedScalar time) {
-  PetscErrorCode ierr;
   CeedVector multlvec;
   Vec Multiplicity, MultiplicityLoc;
+  PetscErrorCode ierr;
 
+  PetscFunctionBeginUser;
   SetupContext ctxSetupData;
   CeedQFunctionContextGetData(ctxSetup, CEED_MEM_HOST, (void **)&ctxSetupData);
   ctxSetupData->time = time;
@@ -266,5 +267,80 @@ PetscErrorCode ICs_FixMultiplicity(CeedOperator op_ics, CeedVector xcorners,
   ierr = DMRestoreLocalVector(dm, &MultiplicityLoc); CHKERRQ(ierr);
   ierr = DMRestoreGlobalVector(dm, &Multiplicity); CHKERRQ(ierr);
 
+  PetscFunctionReturn(0);
+}
+
+// TS: Create, setup, and solve
+PetscErrorCode TSSolve_NS(DM dm, User user, AppCtx app_ctx, Physics phys,
+                          Vec *Q, PetscScalar *f_time, TS *ts) {
+
+  MPI_Comm       comm = user->comm;
+  TSAdapt        adapt;
+  PetscScalar    ftime;
+  PetscErrorCode ierr;
+
+  PetscFunctionBeginUser;
+  ierr = TSCreate(comm, ts); CHKERRQ(ierr);
+  ierr = TSSetDM(*ts, dm); CHKERRQ(ierr);
+  if (phys->implicit) {
+    ierr = TSSetType(*ts, TSBDF); CHKERRQ(ierr);
+    if (user->op_ifunction) {
+      ierr = TSSetIFunction(*ts, NULL, IFunction_NS, &user); CHKERRQ(ierr);
+    } else {                    // Implicit integrators can fall back to using an RHSFunction
+      ierr = TSSetRHSFunction(*ts, NULL, RHS_NS, &user); CHKERRQ(ierr);
+    }
+  } else {
+    if (!user->op_rhs) SETERRQ(comm, PETSC_ERR_ARG_NULL,
+                                 "Problem does not provide RHSFunction");
+    ierr = TSSetType(*ts, TSRK); CHKERRQ(ierr);
+    ierr = TSRKSetType(*ts, TSRK5F); CHKERRQ(ierr);
+    ierr = TSSetRHSFunction(*ts, NULL, RHS_NS, &user); CHKERRQ(ierr);
+  }
+  ierr = TSSetMaxTime(*ts, 500. * user->units->second); CHKERRQ(ierr);
+  ierr = TSSetExactFinalTime(*ts, TS_EXACTFINALTIME_STEPOVER); CHKERRQ(ierr);
+  ierr = TSSetTimeStep(*ts, 1.e-2 * user->units->second); CHKERRQ(ierr);
+  if (app_ctx->test_mode) {ierr = TSSetMaxSteps(*ts, 10); CHKERRQ(ierr);}
+  ierr = TSGetAdapt(*ts, &adapt); CHKERRQ(ierr);
+  ierr = TSAdaptSetStepLimits(adapt,
+                              1.e-12 * user->units->second,
+                              1.e2 * user->units->second); CHKERRQ(ierr);
+  ierr = TSSetFromOptions(*ts); CHKERRQ(ierr);
[…]
