### 2021-04-14 fd936b7
ci - add make tidy to gitlab
