### 2021-04-14 41ed2e8
Fluids - created SetupICsFromBinary() for case of -continue
