+++ examples/fluids/src/misc.c
@@ -146,3 +146,27 @@ PetscErrorCode PostProcess_NS(TS ts, CeedData ceed_data, DM dm,
 
   PetscFunctionReturn(0);
 }
+
+// -- Gather initial Q values in case of continuation of simulation
+PetscErrorCode SetupICsFromBinary(MPI_Comm comm, AppCtx app_ctx, Vec Q) {
+
+  PetscViewer viewer;
+  char file_path[PETSC_MAX_PATH_LEN];
+  PetscErrorCode ierr;
+
+  PetscFunctionBegin;
+
+  // Read input
+  ierr = PetscSNPrintf(file_path, sizeof file_path, "%s/ns-solution.bin",
+                       app_ctx->output_dir); CHKERRQ(ierr);
+  ierr = PetscViewerBinaryOpen(comm, file_path, FILE_MODE_READ, &viewer);
+  CHKERRQ(ierr);
+
+  // Load Q from existent solution
+  ierr = VecLoad(Q, viewer); CHKERRQ(ierr);
+
+  // Clean up PETSc object
+  ierr = PetscViewerDestroy(&viewer); CHKERRQ(ierr);
+
+  PetscFunctionReturn(0);
+}
