### 2021-04-14 619db83
tests/python - add TEST_OPTS and use in GitHub Action
