### 2021-04-14 6cb6732
cleaning up MR dS equations and wording
