### 2021-04-14 0f7fd0f
vec - add PointwiseMult and AXPY convenience functions
