+++ examples/fluids/navierstokes.c
@@ -49,23 +49,6 @@ const char help[] = "Solve Navier-Stokes using PETSc and libCEED\n";
 #include "navierstokes.h"
 
 int main(int argc, char **argv) {
-  PetscInt ierr;
-  MPI_Comm comm;
-  DM dm;
-  User user;
-  Units units;
-  CeedData ceed_data;
-  SetupContext ctxSetupData;
-  Physics ctxPhysData;
-  AppCtx app_ctx;
-  problemData   *problem = NULL;
-  PetscInt lnodes, gnodes;
-  const PetscInt ncompq = 5;
-  PetscMPIInt rank;
-  Vec Q, Qloc, Xloc;
-  Ceed ceed;
-  PetscInt Xloc_size;
-  SimpleBC bc;
 
   // todo: define a function
   // Check PETSc CUDA support
@@ -83,23 +66,37 @@ int main(int argc, char **argv) {
   // ---------------------------------------------------------------------------
   // Initialize PETSc
   // ---------------------------------------------------------------------------
+  PetscInt ierr;
   ierr = PetscInitialize(&argc, &argv, NULL, help);
   if (ierr) return ierr;
 
   // ---------------------------------------------------------------------------
-  // Create contexts
+  // Create structs
   // ---------------------------------------------------------------------------
-  // -- Allocate memory for contexts
-  ierr = PetscCalloc1(1, &user); CHKERRQ(ierr);
-  ierr = PetscMalloc1(1, &units); CHKERRQ(ierr);
-  ierr = PetscCalloc1(1, &problem); CHKERRQ(ierr);
-  ierr = PetscCalloc1(1, &bc); CHKERRQ(ierr);
-  ierr = PetscMalloc1(1, &ctxSetupData); CHKERRQ(ierr);
-  ierr = PetscMalloc1(1, &ctxPhysData); CHKERRQ(ierr);
+  AppCtx app_ctx;
   ierr = PetscCalloc1(1, &app_ctx); CHKERRQ(ierr);
+
+  problemData *problem = NULL; 
+  ierr = PetscCalloc1(1, &problem); CHKERRQ(ierr);
+
+  User user;
+  ierr = PetscCalloc1(1, &user); CHKERRQ(ierr);
+
+  CeedData ceed_data;
   ierr = PetscCalloc1(1, &ceed_data); CHKERRQ(ierr);
 
-  // -- Assign contexts
+  SimpleBC bc;
+  ierr = PetscCalloc1(1, &bc); CHKERRQ(ierr);
+
+  SetupContext ctxSetupData;
+  ierr = PetscMalloc1(1, &ctxSetupData); CHKERRQ(ierr);
+
+  Physics ctxPhysData;
+  ierr = PetscMalloc1(1, &ctxPhysData); CHKERRQ(ierr);
+
+  Units units;
+  ierr = PetscMalloc1(1, &units); CHKERRQ(ierr);
+
   user->app_ctx = app_ctx;
   user->units = units;
   user->phys = ctxPhysData;
@@ -107,11 +104,11 @@ int main(int argc, char **argv) {
   // ---------------------------------------------------------------------------
   // Process command line options
   // ---------------------------------------------------------------------------
-  // Register problems to be available on the command line
+  // -- Register problems to be available on the command line
   ierr = RegisterProblems_NS(app_ctx); CHKERRQ(ierr);
 
-  // Process general command line options
-  comm = PETSC_COMM_WORLD;
+  // -- Process general command line options
+  MPI_Comm comm = PETSC_COMM_WORLD;
   user->comm = comm;
   ierr = ProcessCommandLineOptions(comm, app_ctx); CHKERRQ(ierr);
 
@@ -128,35 +125,10 @@ int main(int argc, char **argv) {
   }
 
[…]
