### 2021-04-26 CEED/libCEED#756 @thelfer
Review of the JOSS Paper
Dear libCEED developpers,

As part of the review of the JOSS paper (https://github.com/openjournals/joss-reviews/issues/2945), please find attached some remarks and questions.

The paper is well written and I was very pleased to review it and play with the library.  Installation is straighforward and the various tests I tried worked out of the box.

However, most of the remarks are related to the fact that the paper shall be made more self-explanatory, notably by picking some information already present in the introduction of the user material.

Do not hesitate to discuss the various questions/remarks and discard irrelevant points.

Regards,
Thomas Helfer

[review.pdf](https://github.com/CEED/libCEED/files/6376430/review.pdf)
[review.md](https://github.com/CEED/libCEED/files/6376431/review.md)
