### 2021-04-26 #2945 @thelfer
@jedbrown, @diehlpk I opened this issue (https://github.com/CEED/libCEED/issues/756) where I uploaded my review of the paper.
