### 2021-04-26 CEED/libCEED#756 @thelfer
Note: for the question concerning boundary conditions, two boundary conditions are of general interest:

- Dirichlet boundary conditions
- Neuman boundary conditions
