### 2021-04-27 CEED/libCEED#756 @jeremylt
Thank you for your review!

I added some additional content to the paper. We'll need to edit flow flow, content, size, etc, but I have initially addressed many of the comments.

---

1.1 - I added a comment on the two use cases as I see them - low level integration and high level use via a library such as MFEM (PETSc integration is still WIP so I de-emphasized it in the intro)
```
Users and library developers can integrate `libCEED` at a low level into existing applications in place of existing matrix-vector products without significant refactoring of their own discretization infrastructure.
Alternatively, users can utilize integrated `libCEED` support in MFEM [@MFEMlibrary; @mfem-paper].
```
We may wish to say more/edit

1.2 - I explained that users can supply their own bases
```
The discrete basis $B$ is the purely algebraic expression of a finite element basis (shape functions) and quadrature; it often possesses structure that is exploited to speed up its action.
Some constructors are provided for arbitrary order $H^1$ Lagrange bases with a tensor-product representation due to the computational efficiency of computing solution values and derivatives at quadrature points via tensor contractions.
However, the user can define a `CeedBasis` for arbitrary geometry elements, to include tetrahedra, by providing quadrature weights and the matrices used to compute solution values and derivatives at quadrature points from the DoFs on the element.
```

1.3 - Clarified libCEED's 'out of the box' MFEM integration
```
If MFEM is built with `libCEED` support, existing MFEM users can pass `-d ceed-cuda:/gpu/cuda/gen` to use a `libCEED` fastest CUDA backend, and similarly for other backends.
The `libCEED` implementations, accessed in this way, currently provide MFEM users with the fastest operator action on CPUs and GPUs (CUDA and HIP/ROCm) without writing any libCEED Q-functions.
```

2.1 - Not yet

2.2 - Added reference

2.3 - Not yet - @valeriabarra, where did we get the 2% number?

2.4 - Mentioned a bit more in 1.2 above

2.5 - Added refs for NVRTC and OCCA there, there is *zero* official HIPRTC documentation

2.6 - Mentioned in 1.2 above

2.7 - Not yet

2.8 - I added some notes
```
`libCEED` provides computation of the true operator diagonal for preconditioning with Jacobi and Chebyshev as well as automatic construction of p-multigrid prolongation and restriction operators from a `CeedOperator`.
Preconditioning matrix-free operators is an active area of research; support for domain decomposition methods and inexact subdomain solvers based on the Fast Diagonalization Method are in active development.
Additionally, `libCEED` supports meshes with mixed order and geometry elements, which means `libCEED` can support solvers and preconditioners utilizing adaptive mesh refinement.
```

2.9 - Added
```
Inhomogeneous Neumann, Robin, and nonlinear boundary conditions can be added in a similar fashion by adding terms integrated over boundary faces while Dirichlet boundary conditions can be added by setting the target values prior to applying the operator representing the weak form.
```

2.10 - Not yet

2.11 - Not yet
