### 2021-04-27 CEED/libCEED#756 @valeriabarra
> 2.3 - Not yet - @valeriabarra, where did we get the 2% number?
>

I believe @jedbrown added that bit. We should add a reference for that to address the reviewer's comment.
