### 2021-04-28 CEED/libCEED#756 @jedbrown
This is a consequence of modern hardware offering about ten f64 flops per byte of DRAM bandwidth. Each matrix entry is 8 bytes plus whatever integer index is required (can be amortized sometime), but you only do two flops (multiply and add), so you're sitting at hardware (2 flop SpMV) / (80 flop hardware capability) = 2.5%. Meanwhile, hardware never achieves its theoretical STREAM bandwidth, while it can (e.g., in large DGEMM) get closer to peak flops, so you're basically at 2% of peak if indices were free (you only store nonzero matrix entries) and only do SpMV.

![image](https://user-images.githubusercontent.com/3303/116347239-13e4b780-a7a9-11eb-8bcd-dcddd4fcecc2.png)
https://github.com/karlrupp/cpu-gpu-mic-comparison
