### 2021-04-28 ae04717
Docs: Tiny rewording in Intro
Since this Introduction was mentioned lately in issue #756 , I was re-reading it and realized this sentence could have been improved. 
The use of the word `they` was a bit confusing in my opinion. 

 @jeremylt I just changed this sentence expanding its line, as recommended in #735 , although the original file had the older style (the max 80 char per line), but I wouldn't change the entire file style now. Do you agree?
