### 2021-05-05 CEED/libCEED#756 @valeriabarra
Thanks @jedbrown  for explaining that. Should we use the [blog post](https://www.karlrupp.net/2013/06/cpu-gpu-and-mic-hardware-characteristics-over-time/) referenced in the GH repo as a citation to address the reviewer comment? Or to you knowledge this was also published somewhere else?
