### 2021-05-05 CEED/libCEED#756 @jeremylt
FWIW, I asked Karl if he'd be willing to add a DOI to the repo to make it easier to cite in an issue in the repo last year, but it seems he's been busy.(https://github.com/karlrupp/cpu-gpu-mic-comparison/issues/7) I just directly cited the repo in my dissertation proposal and plan to do the same for my dissertation.
