### 2021-05-18 CEED/libCEED#756 @diehlpk
Hey, any update on this? When do you estimate to finish that, so we can publish the paper?
