### 2021-05-18 CEED/libCEED#756 @jeremylt
We were just chatting about this offline. We plan to wrap up addressing these comments late this week or early next week.
