### 2021-05-18 CEED/libCEED#756 @valeriabarra
@jeremylt for point 2.2 though, should we upload and cite v0.8 ?
