### 2021-05-19 CEED/libCEED#756 @valeriabarra
Ok.
@jedbrown would you please upload the v0.8 (or v0.8.1 if you prefer) User Manual to Zenodo?
