### 2021-05-21 CEED/libCEED#756 @jeremylt
I believe our ToDos are
- [ ] Update references concurrent with v0.8.1 release
- [ ] Update figure for readability
