### 2021-05-28 CEED/libCEED#756 @diehlpk
@jeremylt When do you think you can finish the last missing ToDos?

Patrick
