### 2021-06-15 #2945 @diehlpk
@danielskatz There are some open issues here https://github.com/CEED/libCEED/issues/756 reported by @thelfer.

The last thing, I heard around 18 days ago was

> We are in the process of wrapping up everything for our v0.8.1 release - it should be very soon.

I contacted @jedbrown today and as far as I understood they're working this week on the open issues.
