### 2021-06-18 CEED/libCEED#756 @diehlpk
Hey @jeremylt @valeriabarra  since I have not heard from you since 21 day, I would like to ask what hinders you to finish the last two items?

If I do not hear from you by June 25th with some finished release, will assume you withdraw the paper.
