### 2021-06-18 CEED/libCEED#756 @jeremylt
Let me double check with @jedbrown on what the release timeline is.
