### 2021-06-18 CEED/libCEED#756 @diehlpk
@jeremylt Note that there is no release necessary if you just change the paper. So update the paper and use the current release if all changes from the review are in that release.
