### 2021-06-18 CEED/libCEED#756 @valeriabarra
> Hey @jeremylt @valeriabarra since I have not heard from you since 21 day, I would like to ask what hinders you to finish the last two items?
>
> If I do not hear from you by June 25th with some finished release, I will assume you withdraw the paper.

Hey, sorry @diehlpk  for our very slow and intermittent communication with you. I apologize.
I tried to have a better communication about the minor outstanding points internally, too. Hopefully @jedbrown can let us know what is the plan, soon
