### 2021-06-18 CEED/libCEED#756 @jedbrown
We need to make a release because we addressed some (relatively minor, but still important) points brought up during review. I'd have been able to finish my part of that long ago if not for recurring sick kids/childcare disruptions and too many deadlines that felt more urgent. I will prioritize this as well as possible.
