### 2021-06-23 CEED/libCEED#756 @valeriabarra
Thanks @diehlpk and again I apologize for the lack of communication.
@jedbrown please let us know what you envisioned timeline is or if more time is needed so that @diehlpk knows. Thanks
