### 2021-06-28 #2945 @jedbrown
Sorry about the big delay. We fixed a release blocking bug yesterday and I made minor copy edits/bib updates so I'm ready to proceed with a v0.9 release (publishing packages for all languages) and this paper if there are no further concerns.
