### 2021-06-28 #2945 @diehlpk
@jedbrown I read the paper and my only concern is Figure 4. For reproducibility, could you add the version of the used library and libCEED itself?
