### 2021-06-28 #2945 @jedbrown
Figure 4 would have been PETSc 3.14 release candidate and libCEED 0.7 release candidate, though it should be equivalent today. Would you prefer that I list detailed provenance (versions and configuration options of transitive dependencies and compilers) in the text or place the log files (which contain comprehensive details) as supplementary material in the software archive or a supplementary archive?
