### 2021-06-28 #2945 @diehlpk
@jedbrown I think adding these two numbers to the paper would and have more details somewhere else would be sufficient.
