### 2021-07-04 CEED/libCEED#756 @thelfer
@jedbrown, @valeriabarra There are still some open questions in my review but those are mostly related to my own curiosity. They  shall no further delay the reviewing process. Feel free to close this ticket as the paper is now ok for me.
