### 2021-07-04 #2945 @thelfer
@diehlpk My concerns about the paper have been adressed in the current version (of the paper). Thus I recommend its publication and asked @jedbrown to close Ticket 756 (https://github.com/CEED/libCEED/issues/756). Please remind me if I have to make other tasks.
Best,
Thomas
