### 2021-07-04 #2945 @diehlpk
> @diehlpk My concerns about the paper have been adressed in the current version (of the paper). Thus I recommend its publication and asked @jedbrown to close Ticket 756 (https://github.com/CEED/libCEED/issues/756). Please remind me if I have to make other tasks.
> Best,
> Thomas

@thelfer Thanks, you are done with the review. I wait for @jedbrown to add address the issue, I mentioned above and I will recommend the paper for acceptance.
