### 2021-07-06 #2945 @diehlpk
> @jedbrown I think adding these two numbers to the paper and have more details somewhere else would be sufficient.

@jedbrown, @valeriabarra Any news on the last task? I really would like to finish this paper and take a new paper. Do you think you can finish it by the end of this week?
