### 2021-07-06 CEED/libCEED#756 @jedbrown
I think I've ensured that the questions are addressed now. With respect to problem sizes, each line ranges from about 30k to 30M dofs/device, which captures an extreme strong scaling (latency-limited) regime all the way to pure throughput-limited (and close to GPU device memory capacity depending on degree).

I did not speak directly to static condensation and hybridization. Indeed, I don't think they make sense in a matrix-free world unless the operator is separable and a tensor-product basis can be used, in which case fast diagonalization can be used to reconstruct the interior without large setup costs. But the applications of HDG are mostly with non-separable operators. There is some ongoing work to apply BDDC in a similar context, though we're mainly interested in approximate element solvers (where you can't eliminate the interior entirely).

The paper discusses some aspects of preconditioning, but the most exciting work in that space is ongoing basic research (in MFEM, PETSc, and libCEED, by multiple groups) and I think it would be presumptive to explain in this paper. libCEED's goal is to provide the common infrastructure, not to take over.
