### 2021-07-06 CEED/libCEED#756 @jedbrown
Closing; feel free to reopen or open a new issue for further discussion.
