### 2021-07-06 #2945 @jedbrown
Yep, the caption has version numbers and range of sizes. I'm working on release QA now.
