### 2021-07-07 #2945 @jedbrown
I think you can do `@whedon recommend-accept from branch joss-paper`
