### 2021-07-09 #2945 @kyleniemeyer
Hi @jedbrown, I was looking over the paper before accepting, and it looks to be missing the required Statement of Need section. Can you please add this? The easiest thing is probably to just add it to the Summary section (i.e., "Summary and Statement of Need") or something.

Also, the Zenodo archive appears to have a number of authors listed that do not appear in the paper. Usually we ask that these match up, unless you are purposely including additional folks.
